% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipd.manipulation.R
\name{alignTimeseriesBin}
\alias{alignTimeseriesBin}
\title{Align and bin two timeseries into comparable bins}
\usage{
alignTimeseriesBin(
  timeX,
  valuesX,
  timeY,
  valuesY,
  binvec = NA,
  binstep = NA,
  binfun = mean,
  max.ens = NA,
  minObs = 10
)
}
\arguments{
\item{timeX}{matrix of age/time ensembles, or single column}

\item{valuesX}{matrix of values ensembles, or single column}

\item{timeY}{matrix of age/time ensembles, or single column}

\item{valuesY}{matrix of values ensembles, or single column}

\item{binvec}{vector of bin edges for binning step}

\item{binstep}{spacing of bins, used to build bin step}

\item{binfun}{function to use during binning (mean, sd, and sum all work)}

\item{max.ens}{maximum number of ensemble members to regress}

\item{minObs}{minimum number of points required to calculate regression}
}
\value{
list of binned data output:
\itemize{
\item binX: binned values from X
\item binY: binned values from Y
\item binstep: interval of the binning
\item yearBins: bins along time
}
}
\description{
Use this to put two timeseries on different timesteps onto equivalent bins
}
\author{
Nick McKay
}
