% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.regression.functions.R
\name{ar1}
\alias{ar1}
\title{Estimate Auto-Regressive coefficient at 1-timesetep}
\usage{
ar1(X)
}
\arguments{
\item{X}{a 1-column matrix or numeric dataset}
}
\value{
ar coefficient estimate of ar1
}
\description{
estimates ar1 using the arima() function
}
\author{
Julien Emile-Geay
}
