% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.spectra.ens.R
\name{ar1Surrogates}
\alias{ar1Surrogates}
\title{Create a synthetic timeseries that emulates the characteristics of a variable}
\usage{
ar1Surrogates(time, vals, detrend_bool = TRUE, method = "redfit",
  nens = 1)
}
\arguments{
\item{time}{LiPD "variable list" or vector of year/age values}

\item{vals}{LiPD "variable list" or vector of values}

\item{detrend_bool}{boolean variable, indicating whether one should remove a linear trend (default) or not. Note: does not affect the AR(1) fit, but does affect the generated timeseries}

\item{method}{Method used for estimation. Possible values are "even" or "redfit"}

\item{nens}{Number of ensemble members to simulate}
}
\value{
a vector or matrix of AR(1) surrogates for series X
}
\description{
create synthetic timeseries based on a timeseries. Useful for null hypothesis testing
}
\seealso{
Other spectra: \code{\link{computeSpectraEns}},
  \code{\link{plotSpectraAnnotate}},
  \code{\link{plotSpectraEns}}

Other pca: \code{\link{pcaEns}}, \code{\link{plotPcaEns}}
}
