% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronBuilding.R
\name{askUser}
\alias{askUser}
\title{askUser}
\usage{
askUser(query)
}
\arguments{
\item{query}{}
}
\value{
user input
}
\description{
Helper function to consistently get user input
}
\seealso{
Other utility: 
\code{\link{concatEnsembleTimeseries}()},
\code{\link{convertAD2BP}()},
\code{\link{convertBP2AD}()},
\code{\link{createChronMeasInputDf}()},
\code{\link{gaussianize}()},
\code{\link{getLastVarString}()},
\code{\link{getOs}()},
\code{\link{heuristicUnits}()},
\code{\link{loadRemote}()},
\code{\link{pullInstance}()},
\code{\link{stringifyVariables}()}
}
\concept{utility}
