% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.functions.R
\name{assignColors}
\alias{assignColors}
\title{Assign colors for map color scale}
\usage{
assignColors(colors = "temp")
}
\arguments{
\item{colors}{string to define color scale. Options are "temp", "precip" and "drought"}
}
\value{
a pair of high/low colors
}
\description{
Quick look up for color scale pairs for mapping
}
\seealso{
Other mapping: 
\code{\link{baseMap}()},
\code{\link{mapLipd}()},
\code{\link{mapTs}()}
}
\author{
Nick McKay
}
\concept{mapping}
