% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.functions.R
\name{baseMap}
\alias{baseMap}
\title{Make a base map}
\usage{
baseMap(
  lon,
  lat,
  map.type = "line",
  f = 0.1,
  restrict.map.range = TRUE,
  projection = "mollweide",
  boundcirc = FALSE,
  global = FALSE,
  extend.range = 5
)
}
\arguments{
\item{lon}{longitude(s) range to map}

\item{lat}{latitude(s) range to map}

\item{map.type}{"stamen" or "line"}

\item{f}{buffer for the map range}

\item{restrict.map.range}{TRUE or FALSE. Trim the size of the map to the points, for "line" map type}

\item{projection}{Map project. All options on: ?mapproject}

\item{boundcirc}{Draw a boundary circle around a polar projection. TRUE or FALSE(default).}

\item{global}{Should the scope of the map be global? TRUE or FALSE(default).}

\item{extend.range}{increase the span of the map by this much (lat/long degrees)}
}
\value{
ggmap base map
}
\description{
Create a stamen or line map of the location of a list of LiPD objects
}
\seealso{
Other mapping: 
\code{\link{assignColors}()},
\code{\link{mapLipd}()},
\code{\link{mapTs}()}
}
\author{
Nick McKay
}
\concept{mapping}
