% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.regression.functions.R
\name{bin}
\alias{bin}
\title{Bin Data}
\usage{
bin(time, values, binvec, binfun = mean)
}
\arguments{
\item{time}{vector of time}

\item{values}{vector of values to bin}

\item{binvec}{vector of bin edges for describing where to bin}

\item{binfun}{function to use during binning (mean, sd, and sum all work)}
}
\value{
A data.frame of (x) binned time, and (y) binned values
}
\description{
function that puts data into appropriate bins, based on the time and the binning vector the bin vector describes the edges of the bins
}
\author{
Nick McKay
}
