% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-helpers.R
\name{bin2d}
\alias{bin2d}
\title{Two dimensional binning}
\usage{
bin2d(
  x,
  y,
  n.bins = 100,
  x.bin = NA,
  y.bin = NA,
  filter.frac = NA,
  interpolate = TRUE
)
}
\arguments{
\item{x}{n by m matrix where n is the number of observations and m is >= 1}

\item{y}{n by j matrix where n is the number of observations and j is >= 1}

\item{n.bins}{number bins over which to calculate intervals. Used to calculate x.bin if not provided.}

\item{x.bin}{vector of bin edges over which to bin.}

\item{y.bin}{vector of bin edges over which to bin.}

\item{filter.frac}{Used to beef up sampling for poorly sampled intervals. Interpolates intervals with less than filter.frac coverage.}

\item{interpolate}{use interpolation? T/F}
}
\value{
A list with a matrix of density, x.bin and y.bin
}
\description{
Calculate the density of samples along a 2-dimensional grid
}
\seealso{
Other gridding: 
\code{\link{kde2d}()},
\code{\link{quantile2d}()}

Other bin: 
\code{\link{alignTimeseriesBin}()},
\code{\link{bin}()},
\code{\link{binEns}()},
\code{\link{binTs}()}

Other plot help: 
\code{\link{AD2BP_trans}()},
\code{\link{BP2AD_trans}()},
\code{\link{axisLabel}()},
\code{\link{geoChronRPlotTheme}()},
\code{\link{getLegend}()},
\code{\link{getPlotRanges}()},
\code{\link{kde2d}()},
\code{\link{meltDistributionTable}()},
\code{\link{modeSelektor}()},
\code{\link{periodAnnotate}()},
\code{\link{quantile2d}()},
\code{\link{reverselog10_trans}()}
}
\concept{bin}
\concept{gridding}
\concept{plot help}
