% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.regression.functions.R
\name{binEns}
\alias{binEns}
\title{Bin ensemble data}
\usage{
binEns(time, values, binvec, binfun = mean, max.ens = NA)
}
\arguments{
\item{time}{single column vector of time}

\item{values}{single column vector of values to bin}

\item{binvec}{vector of bin edges for binning step}

\item{binfun}{function to use during binning (mean, sd, and sum all work)}

\item{max.ens}{maximum number of ensemble members to regress}

\item{binstep}{spacing of bins, used to build bin step}
}
\value{
list that includes matrix of binned data and binned time
}
\description{
takes ensembles in time and/or values and creates a matrix of data for future analysis
}
