% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.regression.functions.R
\name{binTs}
\alias{binTs}
\title{Bin every entry in a Timeseries object}
\usage{
binTs(
  TS,
  time.var = "ageEnsemble",
  bin.vec,
  bin.fun = mean,
  max.ens = 1000,
  na.col.rm = TRUE
)
}
\arguments{
\item{TS}{LiPD timeseries object See \url{http://nickmckay.github.io/LiPD-utilities/r/index.html#what-is-a-time-series}}

\item{time.var}{specify the time variable to bin (default = "ageEnsemble")}

\item{bin.vec}{vector of bin edges for describing where to bin}

\item{bin.fun}{function to use during binning (mean, sd, and sum all work)}

\item{max.ens}{Maximum number of ensemble members.}

\item{na.col.rm}{Remove columns that are all NAs? (TRUE or FALSE)}
}
\value{
A list of binned years and values.
}
\description{
Aggregate data from a timeseries object into the same timeline through binning.
}
\seealso{
Other bin: 
\code{\link{alignTimeseriesBin}()},
\code{\link{bin}()},
\code{\link{bin2d}()},
\code{\link{binEns}()}
}
\author{
Nick McKay
}
\concept{bin}
