% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux.R
\name{calculateFlux}
\alias{calculateFlux}
\title{Calculate a flux ensemble}
\usage{
calculateFlux(
  L,
  sed.rate.ens = "sedimentationRate",
  sed.rate.age = "ageEnsemble",
  sed.rate.table.num = 1,
  sed.rate.table.type = "ensemble",
  density.var = "WBD",
  density.var.age = "ageEnsemble",
  density.table.num = 1,
  density.table.type = "measurement",
  concentration.var = "RABD",
  concentration.var.age = "ageEnsemble",
  concentration.table.num = 1,
  concentration.table.type = "measurement",
  out.age.var = "age",
  ens.units = "mass/cm2/yr",
  ens.var.name = "flux",
  bin.step = 100,
  bin.vec = NA
)
}
\arguments{
\item{L}{A LiPD object}

\item{sed.rate.ens}{sedimentation rate ensemble (potentially from createSedRateEnsemble())}

\item{sed.rate.age}{age ensemble that corresponds to sed.rate.ens}

\item{sed.rate.table.num}{table number for sed rate ensemble}

\item{sed.rate.table.type}{table type for sed rate ensemble}

\item{density.var}{name of the density variable}

\item{density.var.age}{age variable that corresponds to density}

\item{density.table.num}{table number for density}

\item{density.table.type}{table type for density}

\item{concentration.var}{name of the concentration variable}

\item{concentration.var.age}{age variable that corresponds to concentration}

\item{concentration.table.num}{table number for concentration}

\item{concentration.table.type}{name of the concentration variable}

\item{out.age.var}{What is the output age variable (e.g., age or ageEnsemble)}

\item{ens.units}{What are the output units (default is "mass/cm2/yr")}

\item{ens.var.name}{What is the output variable name (default = "flux")}

\item{bin.step}{What timestep are you calculating flux over?}

\item{bin.vec}{Optionall specify a vector to calculate flux over}
}
\value{
A LiPD object
}
\description{
Calculate a flux ensemble
}
\seealso{
Other LiPD manipulation: 
\code{\link{createConcatenatedEnsembleMeasurementTable}()},
\code{\link{createModel}()},
\code{\link{createMultiModelEnsemble}()},
\code{\link{createSedRateEnsemble}()},
\code{\link{createSummaryTableFromEnsembleTable}()},
\code{\link{createTSid}()},
\code{\link{estimateUncertaintyFromRange}()},
\code{\link{getVariableIndex}()},
\code{\link{mapAgeEnsembleToPaleoData}()},
\code{\link{pullTsVariable}()},
\code{\link{selectData}()}
}
\concept{LiPD manipulation}
