% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{clearAll}
\alias{clearAll}
\title{Clear all variables and functions from global environment}
\usage{
clearAll()
}
\description{
Removes all variables and functions from global environment. Use at your own risk.
}
