% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.spectra.ens.R
\name{computeSpectraEns}
\alias{computeSpectraEns}
\title{Calculate ensemble power spectra}
\usage{
computeSpectraEns(time, values, max.ens = NA, method = "lomb-scargle",
  gauss = FALSE, ofac = 1, padfac = 2, tbw = 3, wgtrad = 1,
  sigma = 0.05, mtm_null = "power_law")
}
\arguments{
\item{time}{LiPD "variable list" or vector of year/age values}

\item{values}{LiPD "variable list" or vector of values}

\item{max.ens}{Maximum number of ensemble members to analyze}

\item{ofac}{oversampling factor for lomb::lsp}

\item{padfac}{padding factor for astrochron::mtm___ functions}

\item{tbw}{time-bandwidth product for astrochron::mtm___ functions}

\item{wgtrad}{radius of the nuspectral::nuwaveletcoeff weight function (non-dimensional units)}

\item{sigma}{decay parameter of the nuspectral::nuwaveletcoeff wavelets}
}
\value{
a list of ensemble spectra results
\itemize{
\item freqs: vector of frequencies
\item power: vector of spectral powers
\item powerSyn: matrix of synthetic spectral power results
}
}
\description{
Calculate ensemble power spectra using lomb-scargle
}
\seealso{
Other spectra: \code{\link{ar1Surrogates}},
  \code{\link{plotSpectraAnnotate}},
  \code{\link{plotSpectraEns}}
}
