% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembleComposite.R
\name{concatEnsembleTimeseries}
\alias{concatEnsembleTimeseries}
\title{Concatenate two independent ensemble timeseries}
\usage{
concatEnsembleTimeseries(ae1, pd1, ae2, pd2)
}
\arguments{
\item{ae1}{Age ensemble list for the first dataset}

\item{pd1}{Paleo data list for the first dataset}

\item{ae2}{Age ensemble list for the second dataset}

\item{pd2}{Paleo data  list for the second dataset}
}
\value{
a matrix with one instance of concatenate ensemble data
}
\description{
This "shuffles" together two *independent* ensemble datasets, making no assumptions about stratigraphic order, but assuming that each dataset is representing the same phenomenon.
}
