% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{convertAD2BP}
\alias{convertAD2BP}
\title{Convert years Calendar year (AD/BC or CE) to BP (1950)}
\usage{
convertAD2BP(X)
}
\arguments{
\item{X}{A LiPD variable list or a vector of years AD/BC/CE/BCE}
}
\value{
X A LiPD variable list or a vector of BP
}
\description{
Converts a LiPD variable list, or vector from AD/BC/CE/BCE to BP (1950). Also deals with the lack of a year 0.
}
\seealso{
Other utility: 
\code{\link{askUser}()},
\code{\link{concatEnsembleTimeseries}()},
\code{\link{convertBP2AD}()},
\code{\link{createChronMeasInputDf}()},
\code{\link{gaussianize}()},
\code{\link{generateEnsembleFromUncertainty}()},
\code{\link{getLastVarString}()},
\code{\link{getOs}()},
\code{\link{heuristicUnits}()},
\code{\link{loadRemote}()},
\code{\link{pullInstance}()},
\code{\link{simulateAutoCorrelatedUncertainty}()},
\code{\link{stringifyVariables}()},
\code{\link{surrogateDataFun}()}
}
\concept{utility}
