% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{convertBP2AD}
\alias{convertBP2AD}
\title{Convert years BP to Calendar year (AD/BC or CE)}
\usage{
convertBP2AD(X)
}
\arguments{
\item{X}{A LiPD variable list or a vector of years BP}
}
\value{
X A LiPD variable list or a vector of Calendar years AD
}
\description{
Converts a LiPD variable list, or vector from BP to AD/BC/CE/BCE. Also deals with the lack of a year 0.
}
