% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.regression.functions.R
\name{corEns}
\alias{corEns}
\title{Ensemble correlation}
\usage{
corEns(
  time.1,
  values.1,
  time.2,
  values.2,
  bin.vec = NA,
  bin.step = NA,
  bin.fun = mean,
  max.ens = NA,
  percentiles = c(0.025, 0.25, 0.5, 0.75, 0.975),
  min.obs = 10,
  fdr.qlevel = 0.05,
  gaussianize = TRUE,
  ...
)
}
\arguments{
\item{time.1}{matrix of age/time ensembles, or single column}

\item{values.1}{matrix of values ensembles, or single column}

\item{time.2}{matrix of age/time ensembles, or single column}

\item{values.2}{matrix of values ensembles, or single column}

\item{bin.vec}{vector of bin edges for binning step}

\item{bin.step}{spacing of bins, used to build bin step}

\item{bin.fun}{function to use during binning (mean, sd, and sum all work)}

\item{percentiles}{quantiles to calculate for regression parameters}

\item{min.obs}{minimum number of points required to calculate regression}

\item{fdr.qlevel}{target false discovery rate (most users won't want to change this)}

\item{...}{
  Arguments passed on to \code{\link[=corMatrix]{corMatrix}}
  \describe{
    \item{\code{ens.1}}{matrix of age-uncertain columns to correlate and calculate p-values}
    \item{\code{ens.2}}{matrix of age-uncertain columns to correlate and calculate p-values}
    \item{\code{max.ens}}{optionally limit the number of ensembles calculated (default = NA)}
    \item{\code{isospectral}}{estimate significance using the Ebisuzaki method (default = TRUE)}
    \item{\code{isopersistent}}{estimate significance using the isopersistence method (default = FALSE)}
    \item{\code{p.ens}}{number of ensemble members to use for isospectral and/or isopersistent methods (default = 100)}
    \item{\code{gaussianize}}{Boolean flag indicating whether the values should be mapped to a standard Gaussian prior to analysis.}
  }}

\item{gauss}{Boolean flag indicating whether the values should be mapped to a standard Gaussian prior to analysis}
}
\value{
list of ensemble output and percentile information
}
\description{
Primary function for calculating correlation ensembles
}
\author{
Nick McKay
}
