% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronBuilding.R
\name{createChronMeasInputDf}
\alias{createChronMeasInputDf}
\title{Create a data frame for chron measurement data}
\usage{
createChronMeasInputDf(
  L,
  chron.num = NA,
  meas.table.num = NA,
  lab.id.var = "labID",
  age.14c.var = "age14C",
  age.14c.uncertainty.var = "age14CUnc",
  age.var = "age",
  age.uncertainty.var = "ageUnc",
  depth.var = "depth",
  reservoir.age.14c.var = "reservoirAge",
  reservoir.age.14c.uncertainty.var = "reservoirAge14C",
  rejected.ages.var = "rejected",
  split.ages = TRUE
)
}
\arguments{
\item{L}{A LiPD object - an R serialization of a single LiPD file. It's a list, and is typically created by `readLipd()`}

\item{chron.num}{}

\item{meas.table.num}{an integer that corresponds to paleo.num measurementTable has the variable you want?}

\item{lab.id.var}{}

\item{age.14c.var}{}

\item{age.14c.uncertainty.var}{}

\item{age.var}{}

\item{age.uncertainty.var}{}

\item{depth.var}{}

\item{reservoir.age.14c.var}{}

\item{reservoir.age.14c.uncertainty.var}{}

\item{rejected.ages.var}{}

\item{split.ages}{if there's an age_type column, and only one age column, intelligently split between age and age14C}

\item{overwrite}{}

\item{site.name}{}

\item{use.marine}{}
}
\value{
a standardized dataframe of chron measurements
}
\description{
Create a data frame for chron measurement data
}
