% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembleComposite.R
\name{createConcatenatedEnsembleMeasurementTable}
\alias{createConcatenatedEnsembleMeasurementTable}
\title{Create a measurementTable from concatenated ensembles}
\usage{
createConcatenatedEnsembleMeasurementTable(ae1, pd1, ae2, pd2, nens = 1000)
}
\arguments{
\item{ae1}{Age ensemble list for the first dataset}

\item{pd1}{Paleo data list for the first dataset}

\item{ae2}{Age ensemble list for the second dataset}

\item{pd2}{Paleo data  list for the second dataset}

\item{nens}{How many ensemble members?}
}
\value{

}
\description{
This "shuffles" together two *independent* ensemble datasets, making no assumptions about stratigraphic order, but assuming that each dataset is representing the same phenomenon.
}
\examples{
\dontrun{
library(lipdR)
library(geoChronR)
library(magrittr)

L1 <- readLipd("~/Downloads/ANS1.Kjellman.2019.lpd")
L2 <- readLipd("~/Downloads/ANP3.Kjellman.2019.lpd")


#composite two independent records
L1 <- mapAgeEnsembleToPaleoData(L1,which.pmt = 1,age.var = "ageEnsemble")

ae1 <- selectData(L1,"ageEnsemble",which.mt = 1)
pd1 <- selectData(L1,"C28d2H",,which.mt = 1)

L2 <- mapAgeEnsembleToPaleoData(L2,which.pmt = 1,age.var = "ageEnsemble")
ae2 <- selectData(L2,"ageEnsemble",which.mt = 1)
pd2 <- selectData(L2,"C28d2H",,which.mt = 1)
cmt <- createConcatenatedEnsembleMeasurementTable(ae1,pd1,ae2,pd2)
plotTimeseriesEnsRibbons(X = cmt$ageEnsemble,Y = cmt$C28d2HComposite) \%>\%
plotTimeseriesEnsLines(X = cmt$ageEnsemble,Y = cmt$C28d2HComposite,maxPlotN = 5,color = "red")
}


}
