% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createModel.R
\name{createModel}
\alias{createModel}
\title{Create a chron or paleo model including adding ensemble data}
\usage{
createModel(
  L,
  depth.or.age.vector,
  ensemble.data,
  paleo.or.chron = "chronData",
  paleo.or.chron.num = NA,
  model.num = NA,
  ens.table.num = 1,
  depth.or.age.var = "depth",
  depth.or.age.units = "cm",
  ens.var = "ageEnsemble",
  ens.units = "yr BP",
  make.new = TRUE,
  methods = NA,
  create.summary.table = TRUE
)
}
\arguments{
\item{L}{A LiPD object - an R serialization of a single LiPD file. It's a list, and is typically created by `readLipd()`}

\item{depth.or.age.vector}{A vector of data for the ensemble data that records the position}

\item{ensemble.data}{A matrix or data.frame of ensemble data, with ensemble members in separate columns}

\item{paleo.or.chron}{"paleoData" or "chronData"}

\item{paleo.or.chron.num}{an integer that corresponds to paleo.num or chron Data object (L$<paleo.or.chron>[[?]]) has the variable you want?}

\item{model.num}{an integer that corresponds to model.num that has the variable you want}

\item{ens.table.num}{an integer that corresponds to ensembleTable you want to get the variable from?}

\item{depth.or.age.var}{A string describing the name of the depth or age vector (default = "depth")}

\item{depth.or.age.units}{Units for depth.or.age.var (default = "cm")}

\item{ens.var}{A string describing the name of the ensemble data (default = "age")}

\item{ens.units}{Units for the ensemble data (default = "yr BP")}

\item{make.new}{Create a new model? (default = TRUE)}

\item{methods}{Optionally add in a list describing the methods used to make the model (default = NA)}

\item{create.summary.table}{Optionally create a summary table derived from the ensemble data (default = TRUE)}
}
\value{

}
\description{
This function will create a model in a LiPD object, adding ensemble data
}
