% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxcal.R
\name{createOxcalModel}
\alias{createOxcalModel}
\title{Create the oxcal model script}
\usage{
createOxcalModel(
  cdf,
  depthsToModel = NA,
  depthInterval = 10,
  thin = 50,
  nIt = 10000,
  eventsPerUnitLength = 1,
  eventsPerUnitLengthUncertainty = 2,
  seqType = "P_Sequence",
  topBoundary = "Boundary",
  bottomBoundary = "Boundary",
  outlierProb = 0.05,
  calCurve = "intcal13"
)
}
\arguments{
\item{cdf}{}

\item{depthsToModel}{}

\item{depthInterval}{}

\item{thin}{}

\item{nIt}{}

\item{eventsPerUnitLength}{}

\item{eventsPerUnitLengthUncertainty}{In orders of magnitude}

\item{seqType}{}

\item{topBoundary}{}

\item{bottomBoundary}{}
}
\value{

}
\description{
Create the oxcal model script
}
