% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.spectra.ens.R
\name{createSyntheticTimeseries}
\alias{createSyntheticTimeseries}
\title{Create a synthetic timeseries that emulates the characteristics of a timeseries}
\usage{
createSyntheticTimeseries(
  time,
  values,
  n.ens = 1,
  sameTrend = TRUE,
  index.to.model = NA
)
}
\arguments{
\item{time}{LiPD "variable list" or vector of year/age values}

\item{values}{LiPD "variable list" or vector of values}

\item{n.ens}{Number of ensemble members to simulate}

\item{sameTrend}{should the synthetic data have the same first-order trend as the original? (default  = TRUE)}

\item{index.to.model}{an optional index of values to model (default = NA, the whole series)}
}
\value{
a vector or matrix of synthetic values
}
\description{
create synthetic timeseries based on a timeseries. Useful for null hypothesis testing
}
\seealso{
Other spectra: 
\code{\link{ar1Surrogates}()},
\code{\link{computeSpectraEns}()},
\code{\link{periodAnnotate}()},
\code{\link{plotSpectrum}()},
\code{\link{reverselog10_trans}()}

Other pca: 
\code{\link{ar1Surrogates}()},
\code{\link{pcaEns}()},
\code{\link{plotPcaEns}()},
\code{\link{plotScreeEns}()}
}
\concept{pca}
\concept{spectra}
