% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.regression.functions.R
\name{effectiveN}
\alias{effectiveN}
\title{Estimate effective sample size accounting for autocorrelation}
\usage{
effectiveN(X, Y)
}
\arguments{
\item{X}{a 1-column matrix or numeric dataset}

\item{Y}{a 1-column matrix or numeric dataset of the same length as X}
}
\value{
estimate of the effective sample size
}
\description{
Bretherton et al., 1999 estimate of effective sample size.
}
\seealso{
Other correlation: 
\code{\link{ar1}()},
\code{\link{ar1Surrogates}()},
\code{\link{corMatrix}},
\code{\link{plotCorEns}()},
\code{\link{pvalMonteCarlo}()},
\code{\link{pvalPearsonSerialCorrected}()}
}
\author{
Nick McKay
}
\concept{correlation}
