% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Neotoma2LiPD.R
\name{estimateUncertaintyFromRange}
\alias{estimateUncertaintyFromRange}
\title{Estimate uncertainty estimates from high/low range}
\usage{
estimateUncertaintyFromRange(MT, range1 = "age.young", range2 = "age.old")
}
\arguments{
\item{MT}{LiPD "measurementTable"}

\item{range1}{name of one of the range variables}

\item{range2}{name of the other range variable}
}
\value{
MT: a LiPD measurementTable with a new unc.estimate variable
}
\description{
Estimate uncertainty (plus/minus values) from a range of values
}
\seealso{
Other LiPD manipulation: \code{\link{flipCoords}},
  \code{\link{getVariableIndex}},
  \code{\link{mapAgeEnsembleToPaleoData}},
  \code{\link{selectData}}
}
