% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdr.R
\name{fdr.master}
\alias{fdr.master}
\title{FDR Master}
\usage{
fdr.master(pvals, qlevel = 0.05, method = "original")
}
\arguments{
\item{pvals}{(required):  a vector of pvals on which to conduct the multiple testing}

\item{qlevel}{the proportion of false positives desired}

\item{method}{one of 'original', the original method of Benjamini & Hochberg (1995), or 'general', the method of Benjamini & Yekutieli (2001), which requires no assumptions about the p-values, but which is much more conservative.  We recommend 'original' for climatological data, and suspect it works well generally for spatial data.}
}
\value{
NULL if no significant tests, or a vector of the indices of the significant tests
}
\description{
This is an internal function that performs various versions of the FDR procedure, but without the modification described in section 4 of our J of Climate paper.
}
\seealso{
Other FDR: 
\code{\link{fdrBasic}()},
\code{\link{fdr}()},
\code{\link{propAlt}()},
\code{\link{storey}()}
}
\author{
Chris Paciorek
}
\concept{FDR}
