% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdr.R
\name{fdrBasic}
\alias{fdrBasic}
\title{FDR Basic}
\usage{
fdrBasic(pvals, qlevel = 0.05)
}
\arguments{
\item{pvals}{(required):  a vector of pvals on which to conduct the multiple testing}

\item{qlevel}{the proportion of false positives desired}
}
\value{
NULL if no significant tests, or a vector of the indices of the significant tests
}
\description{
This is an internal function that performs the basic FDR of Benjamini & Hochberg (1995).
}
\seealso{
Other FDR: \code{\link{fdr.master}}, \code{\link{fdr}},
  \code{\link{propAlt}}, \code{\link{storey}}
}
\author{
Chris Paciorek
}
