% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipd.manipulation.R
\name{flipCoords}
\alias{flipCoords}
\title{Flip Coordinates}
\usage{
flipCoords(L)
}
\arguments{
\item{L}{a LiPD object}
}
\value{
a LiPD object
}
\description{
Swap latitude and longitude in a LiPD object
}
\seealso{
Other LiPD manipulation: \code{\link{estimateUncertaintyFromRange}},
  \code{\link{getVariableIndex}},
  \code{\link{mapAgeEnsembleToPaleoData}},
  \code{\link{pullTsVariable}},
  \code{\link{pushTsVariable}}, \code{\link{selectData}},
  \code{\link{tidyTs}}
}
