% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.geoChronR.R
\name{FrequencyAnnotate}
\alias{FrequencyAnnotate}
\title{Annotate plot of spectra with given periodicities}
\usage{
FrequencyAnnotate(specPlot, periods = c(19, 23, 41, 100), colour = "red")
}
\arguments{
\item{specPlot}{Output from plotSpectraEns (or other ggplot)}
}
\value{
ggplot object of spectrum plot
}
\description{
Annotate plot of spectra (ensemble or otherwise) with vertical lines at specific periodicities (assumes log10 scaling)
}
\seealso{
Other plot: \code{\link{AD2BP_trans}},
  \code{\link{BP2AD_trans}}, \code{\link{PeriodAnnotate}},
  \code{\link{axisLabel}},
  \code{\link{geoChronRPlotTheme}},
  \code{\link{getLegend}}, \code{\link{getPlotRanges}},
  \code{\link{plotChronEnsDiff}},
  \code{\link{plotChronEns}}, \code{\link{plotChron}},
  \code{\link{plotCorEns}}, \code{\link{plotHistEns}},
  \code{\link{plotLine}},
  \code{\link{plotModelDistributions}},
  \code{\link{plotPcaEns}}, \code{\link{plotPvalsEnsFdr}},
  \code{\link{plotRegressEns}},
  \code{\link{plotScatterEns}},
  \code{\link{plotSpectraEns}}, \code{\link{plotSpectrum}},
  \code{\link{plotSummary}},
  \code{\link{plotTimeseriesEnsLines}},
  \code{\link{plotTimeseriesEnsRibbons}},
  \code{\link{plotTimeseriesStack}},
  \code{\link{plotTrendLinesEns}},
  \code{\link{reverselog10_trans}}

Other spectra: \code{\link{PeriodAnnotate}},
  \code{\link{ar1Surrogates}},
  \code{\link{computeSpectraEns}},
  \code{\link{createSyntheticTimeseries}},
  \code{\link{plotSpectraEns}}, \code{\link{plotSpectrum}},
  \code{\link{reverselog10_trans}}
}
\author{
Julien Emile-Geay
}
