% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoChronR-package.R
\docType{package}
\name{geoChronR-package}
\alias{geoChronR}
\alias{geoChronR-package}
\title{geoChronR: Tools to Analyze and Visualize Time-Uncertain Data}
\description{
Time uncertain chronology developement and analysis, including correlation, regression, spectral, and PCA for paleogeoscientific data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nickmckay/GeoChronR}
  \item Report bugs at \url{https://github.com/nickmckay/GeoChronR/issues}
}

}
\author{
\strong{Maintainer}: Nicholas McKay \email{nick@nau.edu} (\href{https://orcid.org/0000-0003-3598-5113}{ORCID})

Authors:
\itemize{
  \item Julien Emile-Geay (\href{https://orcid.org/0000-0001-5920-4751}{ORCID})
  \item Deborah Khider (\href{https://orcid.org/0000-0001-7501-8430}{ORCID})
}

}
\keyword{internal}
