% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacon.lipd.R
\name{getBaconDir}
\alias{getBaconDir}
\title{Get the name of bacon core directory}
\usage{
getBaconDir(bacon.dir = NA, overwrite.tmp.dir = TRUE)
}
\arguments{
\item{bacon.dir}{if you already have it, it just returns this, (default=NA)}

\item{overwrite.tmp.dir}{Erase if using a tempdir before running?}
}
\value{
bacon.dir the bacon core directory
}
\description{
This is a mostly internal function that returns the core directory. Interactive if given a NA
}
\seealso{
Other Bacon: 
\code{\link{loadBaconOutput}()},
\code{\link{runBacon}()},
\code{\link{sampleBaconAges}()},
\code{\link{setBaconDir}()},
\code{\link{writeBacon}()}
}
\author{
Nick McKay
}
\concept{Bacon}
