% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronBuilding.R
\name{getLastVarString}
\alias{getLastVarString}
\title{Get the last set of parameters you used in createChronMeasInputDf()}
\usage{
getLastVarString()
}
\value{
the last set of parameters you used in createChronMeasInputDf()
}
\description{
Get the last set of parameters you used in createChronMeasInputDf()
}
\seealso{
Other utility: 
\code{\link{askUser}()},
\code{\link{concatEnsembleTimeseries}()},
\code{\link{convertAD2BP}()},
\code{\link{convertBP2AD}()},
\code{\link{createChronMeasInputDf}()},
\code{\link{gaussianize}()},
\code{\link{generateEnsembleFromUncertainty}()},
\code{\link{getOs}()},
\code{\link{heuristicUnits}()},
\code{\link{loadRemote}()},
\code{\link{pullInstance}()},
\code{\link{simulateAutoCorrelatedUncertainty}()},
\code{\link{stringifyVariables}()},
\code{\link{surrogateDataFun}()}
}
\concept{utility}
