% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacon.lipd.R
\name{getOxcalPath}
\alias{getOxcalPath}
\title{Get the name of  directory}
\usage{
getOxcalPath(path = NA)
}
\arguments{
\item{baconDir}{if you already have it, it just returns this, (default=NA)}
}
\value{
baconDir the bacon core directory
}
\description{
This is a mostly internal function that returns the core directory. Interactive if given a NA
}
\author{
Nick McKay
}
