% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-helpers.R
\name{getPlotRanges}
\alias{getPlotRanges}
\title{Get ggplot x and y ranges}
\usage{
getPlotRanges(h)
}
\arguments{
\item{h}{a ggplot object}
}
\value{
a list of x and y ranges
}
\description{
Use this to extract x and y ranges from ggplot, dealing with changes in version
}
\seealso{
Other plot help: 
\code{\link{AD2BP_trans}()},
\code{\link{BP2AD_trans}()},
\code{\link{axisLabel}()},
\code{\link{bin2d}()},
\code{\link{geoChronRPlotTheme}()},
\code{\link{getLegend}()},
\code{\link{kde2d}()},
\code{\link{meltDistributionTable}()},
\code{\link{modeSelektor}()},
\code{\link{periodAnnotate}()},
\code{\link{quantile2d}()},
\code{\link{reverselog10_trans}()}
}
\concept{plot help}
