% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{heuristicUnits}
\alias{heuristicUnits}
\title{Makes a guess about the units of a time vector based on the values}
\usage{
heuristicUnits(X, rangeMin = 25)
}
\arguments{
\item{X}{A LiPD variable list or a vector of years BP}
}
\value{
A *guessed* string unit label
}
\description{
Guesses year units based on some simple heuristics
}
