% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxcal.R
\name{loadOxcalOutput}
\alias{loadOxcalOutput}
\title{Load oxcal output}
\usage{
loadOxcalOutput(
  L,
  oxcal.result.file.path,
  model.parameters,
  depth.units = "cm",
  chron.num = NA,
  model.num = NA,
  make.new = NA,
  max.ens = 1000
)
}
\arguments{
\item{L}{a lipd object}

\item{oxcal.result.file.path}{Path to oxcal output data (from oxcAAR::executeOxcalScript())}

\item{model.parameters}{model parameters to write into the lipd file}

\item{depth.units}{Depth units to assign in (default = "cm")}

\item{model.num}{an integer that corresponds to model.num that has the variable you want}

\item{make.new}{Create a new model?}

\item{max.ens}{Maximum number of ensemble members to import (default = 1000)}
}
\value{

}
\description{
Load oxcal output
}
