% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{loadRemote}
\alias{loadRemote}
\title{Load a remote RData file}
\usage{
loadRemote(url, useSavedNames = FALSE)
}
\arguments{
\item{url}{the url of a remote data file}

\item{useSavedNames}{use the names saved in the file?}
}
\value{
the loadeds object
}
\description{
Load a remote RData file
}
\seealso{
Other utility: 
\code{\link{askUser}()},
\code{\link{concatEnsembleTimeseries}()},
\code{\link{convertAD2BP}()},
\code{\link{convertBP2AD}()},
\code{\link{createChronMeasInputDf}()},
\code{\link{gaussianize}()},
\code{\link{getLastVarString}()},
\code{\link{getOs}()},
\code{\link{heuristicUnits}()},
\code{\link{pullInstance}()},
\code{\link{stringifyVariables}()},
\code{\link{surrogateDataFun}()}
}
\concept{utility}
