% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipd.manipulation.R
\name{mapAgeEnsembleToPaleoData}
\alias{mapAgeEnsembleToPaleoData}
\title{Map an ageEnsemble variable from a chron model to a paleoMeasurement Table}
\usage{
mapAgeEnsembleToPaleoData(L, which.paleo = NA, which.pmt = NA,
  which.chron = NA, which.model = NA, max.ensemble.members = NA,
  strictSearch = FALSE, which.ens = NA)
}
\arguments{
\item{L}{a lipd object}

\item{which.paleo}{an integer that corresponds to which paleoData object (L$paleoData[[?]]) has the measurementTable you want to modify}
}
\value{
L a lipd object
}
\description{
Copies an ageEnsemble from chronData (model) to paleoData (measurementTable), by matching depth and interpolating (extrapolating) as necessary.
}
