% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.functions.R
\name{mapLipds}
\alias{mapLipds}
\title{Map a list of LiPD objects}
\usage{
mapLipds(D, shape = 21, size = 8, color = sapply(D, "[[", "archiveType"),
  map.type = "google", f = 0.3, restrict.map.range = TRUE,
  boundcirc = FALSE, global = FALSE, projection = "mercator")
}
\arguments{
\item{D}{A list of LiPD objects}

\item{shape}{Shape of the location marker}

\item{size}{Size of the location marker}

\item{color}{Color of the location marker}

\item{map.type}{"google" or "line"}

\item{f}{buffer for the map range}

\item{restrict.map.range}{TRUE or FALSE. Trim the size of the map to the points, for "line" map type}

\item{boundcirc}{Draw a boundary circle around a polar projection. TRUE or FALSE(default).}

\item{global}{Should the scope of the map be global? TRUE or FALSE(default).}

\item{projection}{Map project. All options on: ?mapproject}

\item{shape.by.archive}{TRUE or FALSE. Use archiveType to assign shapes.}
}
\value{
ggmap object
}
\description{
Create a google or line map of the location of a list of LiPD objects
}
\seealso{
Other mapping: \code{\link{assignColors}},
  \code{\link{baseMap}}, \code{\link{mapLipd}}
}
\author{
Nick McKay
}
