% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipd.manipulation.R
\name{mapPaleoEnsembleToPaleoData}
\alias{mapPaleoEnsembleToPaleoData}
\title{Map a paleoEnsemble variable from a paleo model to a paleoMeasurement Table}
\usage{
mapPaleoEnsembleToPaleoData(
  L,
  paleo.model.var = NA,
  paleo.model.depth.var = "depth",
  paleo.depth.var = "depth",
  paleo.age.var = "age",
  map.median = NA,
  paleo.num = NA,
  paleo.meas.table.num = NA,
  paleo.mod.num = NA,
  model.num = NA,
  ens.table.num = 1,
  max.ens = NA,
  strict.search = FALSE
)
}
\arguments{
\item{L}{A LiPD object - an R serialization of a single LiPD file. It's a list, and is typically created by `readLipd()`}

\item{paleo.depth.var}{name of the depth variable to search for in the paleo measurement table}

\item{paleo.age.var}{name of the age variable to search for in the paleo measurement table, critical if no depth data are available}

\item{map.median}{Do you want to also map the median of the ageEnsemble to paleoData? (TRUE or FALSE, or NA, the default, which will do so unless that variable already exists)}

\item{paleo.num}{an integer that corresponds to paleo.numData object (L$paleoData[[?]]) has the measurementTable you want to modify}

\item{paleo.meas.table.num}{an integer that corresponds to paleo.num measurementTable you want to add the ensemble to?}

\item{model.num}{an integer that corresponds to chron.num model you want to get the ensemble from?}

\item{ens.table.num}{an integer that corresponds to ensembleTable you want to get the variable from?}

\item{max.ens}{Maximum number of ensemble members to map}

\item{strict.search}{Use a strict.search to look for the ageEnsemble and depth variables. TRUE(default) or FALSE.}

\item{age.var}{name of the age ensemble variable to search for}

\item{chron.depth.var}{name of the depth variable to search for in the ensemble table}

\item{chron.num}{an integer that corresponds to chron.numData object (L$chronData[[?]]) has the model you want to get the ensemble from}
}
\value{
L a lipd object
}
\description{
Copies a paleoEnsemble from (model) to paleoData (measurementTable), by matching depth and interpolating (extrapolating) as necessary.
}
\seealso{
Other LiPD manipulation: 
\code{\link{createConcatenatedEnsembleMeasurementTable}()},
\code{\link{createModel}()},
\code{\link{createMultiModelEnsemble}()},
\code{\link{createSummaryTableFromEnsembleTable}()},
\code{\link{createTSid}()},
\code{\link{estimateUncertaintyFromRange}()},
\code{\link{getVariableIndex}()},
\code{\link{mapAgeEnsembleToPaleoData}()},
\code{\link{mapAgeSummaryToPaleoData}()},
\code{\link{pullTsVariable}()},
\code{\link{selectData}()}
}
\concept{LiPD manipulation}
