% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-chron.R
\name{meltDistributionTable}
\alias{meltDistributionTable}
\title{Melt distribution}
\usage{
meltDistributionTable(this.dist, dist.plot = 1:length(this.dist))
}
\arguments{
\item{this.dist}{LiPD "distributionTable" object}

\item{dist.plot}{vector of distribution tables to plot}
}
\value{
data.frame of melted distribution objects.
}
\description{
Takes a LiPD model distribution and melt it into a single data.frame
}
\seealso{
Other plot help: 
\code{\link{AD2BP_trans}()},
\code{\link{BP2AD_trans}()},
\code{\link{axisLabel}()},
\code{\link{bin2d}()},
\code{\link{geoChronRPlotTheme}()},
\code{\link{getLegend}()},
\code{\link{getPlotRanges}()},
\code{\link{kde2d}()},
\code{\link{modeSelektor}()},
\code{\link{periodAnnotate}()},
\code{\link{quantile2d}()},
\code{\link{reverselog10_trans}()}
}
\author{
Nick McKay
}
\concept{plot help}
