% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Neotoma2LiPD.R
\name{neotoma2Lipd}
\alias{neotoma2Lipd}
\title{Create a LiPD object from Neotoma}
\usage{
neotoma2Lipd(site)
}
\arguments{
\item{site}{the site object from the R Neotoma package, output of  neotoma::get_site()}
}
\value{
A LiPD object
}
\description{
Uses the Neotoma API to create a LiPD file?
}
\details{
Super alpha version 0.00001. Expect updates!
get site fist
site = get_site("Potato Lake")
}
