% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxcal.R
\name{oxCalDateExpression}
\alias{oxCalDateExpression}
\title{write oxcal expression for each date.}
\usage{
oxCalDateExpression(
  ...,
  default.outlier.prob = 0.05,
  cal.curve = "intcal20",
  unknown.delta.r = FALSE
)
}
\arguments{
\item{...}{data to pass into a tibble}

\item{default.outlier.prob}{Outlier Probability}

\item{cal.curve}{Calibration curve}

\item{unknown.delta.r}{Use an unknown delta R (T/f)}
}
\value{
an oxcal date expression
}
\description{
write oxcal expression for each date.
}
\seealso{
Other oxcal: 
\code{\link{createOxcalModel}()},
\code{\link{getModelParametersFromOxcalText}()},
\code{\link{loadOxcalOutput}()},
\code{\link{runOxcal}()}
}
\concept{oxcal}
