% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.geoChronR.R
\name{pcaEns}
\alias{pcaEns}
\title{Perform principle components analysis (PCA) across an ensemble}
\usage{
pcaEns(
  bin.list,
  method = "ppca",
  weights = NA,
  pca.type = "corr",
  n.pcs = 4,
  nEns = 1000,
  simulateTrendInNull = FALSE
)
}
\arguments{
\item{bin.list}{A list of binned data, the output of binTs()}

\item{method}{What method to use for PCA? pcaMethods::listPcaMethods() for options. "ppca" is default. Other options may not work in GeoChronR.}

\item{weights}{Vector of weights to apply to timeseries in the bin.list}

\item{pca.type}{Correlation ("corr" - default) or Covariance ("cov"), matrix}

\item{n.pcs}{number of PCs/EOFs to calculate}

\item{n.ens}{how many ensemble members?}
}
\description{
Ensemble PCA, or Monte Carlo Empirical Orthogonal Functions as described in Anchukaitis and Tierney 2012.
}
\seealso{
Other pca: 
\code{\link{ar1Surrogates}()},
\code{\link{createSyntheticTimeseries}()},
\code{\link{getLegend}()},
\code{\link{plotPcaEns}()}
}
\concept{pca}
