% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.geoChronR.R
\name{plotChron}
\alias{plotChron}
\title{High-level chron plotting}
\usage{
plotChron(L, chron.number = NA, meas.num = NA, depth.var = "depth",
  age.var = "age", age14C.var = "age14C", dotSize = 5,
  legend.position = c(0.7, 0.3), ...)
}
\arguments{
\item{L}{A LiPD object}

\item{chron.number}{which chronData object to use (NA by default, will ask if needed)}

\item{meas.num}{which chronData model to use (NA by default, will ask if needed)}

\item{depth.var}{variableName to use for depth ("depth" by default)}

\item{age.var}{variableName to use for age ensemble ("age" by default)}

\item{age14C.var}{variableName to use for age ensemble ("age14C" by default)}

\item{legend.position}{where to put the legend on the chron plot?}

\item{...}{arguments to pass on to plotChronEns()}

\item{dotsize}{what size dot for the chron plot? Only used if not plotting by plotChronEns() (default = 5)}
}
\value{
a ggplot object, or NA if there's chronData to plot
}
\description{
Plot a chronology, either from a chron model (preferred) or from a chron measurement table if there is no model
}
\seealso{
Other plot: \code{\link{axisLabel}},
  \code{\link{geoChronRPlotTheme}},
  \code{\link{getLegend}}, \code{\link{getPlotRanges}},
  \code{\link{plotChronEnsDiff}},
  \code{\link{plotChronEns}}, \code{\link{plotCorEns}},
  \code{\link{plotHistEns}}, \code{\link{plotLine}},
  \code{\link{plotModelDistributions}},
  \code{\link{plotPcaEns}}, \code{\link{plotPvalsEnsFdr}},
  \code{\link{plotRegressEns}},
  \code{\link{plotScatterEns}},
  \code{\link{plotSpectraEns}}, \code{\link{plotSummary}},
  \code{\link{plotTimeseriesEnsLines}},
  \code{\link{plotTimeseriesEnsRibbons}},
  \code{\link{plotTimeseriesStack}},
  \code{\link{plotTrendLinesEns}}

Other chron: \code{\link{plotChronEnsDiff}},
  \code{\link{plotChronEns}},
  \code{\link{plotModelDistributions}}
}
\author{
Nick McKay
}
