% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.geoChronR.R
\name{plotChronEns}
\alias{plotChronEns}
\title{Plot chron ensemble}
\usage{
plotChronEns(
  L,
  age.var = "ageEnsemble",
  depth.var = "depth",
  chron.number = NA,
  model.num = NA,
  probs = c(0.025, 0.25, 0.5, 0.75, 0.975),
  x.bin = NA,
  y.bin = NA,
  n.bins = 100,
  color.low = "white",
  color.high = "grey70",
  alp = 1,
  color.line = "Black",
  line.width = 1,
  add.to.plot = ggplot2::ggplot(),
  n.ens.plot = 5,
  color.ens.line = "red",
  alp.ens.line = 0.7,
  distAlp = 0.3,
  dist.type = "violin",
  dist.color = "purple",
  dist.thick = 0.1,
  dist.scale = 0.02,
  truncate.dist = NA,
  add.paleo.age.depth = FALSE,
  paleo.number = NA,
  meas.num = NA,
  color.line.paleo = "cyan"
)
}
\arguments{
\item{L}{A LiPD object}

\item{age.var}{ariableName to use for age ensemble ("ageEnsemble" by default)}

\item{depth.var}{variableName to use for depth ("depth" by default)}

\item{chron.number}{chron.numData object to use (NA by default, will ask if needed)}

\item{model.num}{chron.numData model to use (NA by default, will ask if needed)}

\item{probs}{quantiles to calculate and plot}

\item{x.bin}{vector of bin edges over which to bin.}

\item{y.bin}{vector of bin edges over which to bin.}

\item{n.bins}{number bins over which to calculate intervals. Used to calculate x.bin if not provided.}

\item{color.low}{Band color of the outer most band.}

\item{color.high}{Band color of the inner most band.}

\item{alp}{Transparency of the band plot}

\item{color.line}{Line color (following ggplot rules)}

\item{line.width}{Width of the line}

\item{add.to.plot}{A ggplot object to add this plot to. Default is ggplot() .}

\item{n.ens.plot}{Number of ensemble members to plot}

\item{color.ens.line}{color of the ensemble lines}

\item{alp.ens.line}{transparency of the lines}

\item{dist.type}{"violin" (default), "up" for one-sided distributions pointed up, "down" for one-sided distributions pointed down}

\item{dist.color}{distribution color (following ggplot rules)}

\item{dist.thick}{thickness of the line around the distribution}

\item{dist.scale}{controls the vertical span of the probability distribution. Approximately the vertical fraction of the plot that the distribution will cover.}

\item{truncate.dist}{truncate probability density values below this number. NA (default) means no truncation}

\item{add.paleo.age.depth}{add a line that shows the paleoData age depth.}

\item{paleo.number}{paleo.num number for the paleoData age-depth}

\item{meas.num}{which measurement Table for the paleoData age-depth}

\item{color.line.paleo}{line color of the paleoData age-depth (following ggplot rules)}
}
\value{
A ggplot object
}
\description{
Plot creates an age model plot with all the bells and whistles, including a spread of ensemble members, probability distributions, and a few example ensemble members.
}
\seealso{
Other plot: 
\code{\link{AD2BP_trans}()},
\code{\link{BP2AD_trans}()},
\code{\link{axisLabel}()},
\code{\link{geoChronRPlotTheme}()},
\code{\link{getLegend}()},
\code{\link{getPlotRanges}()},
\code{\link{periodAnnotate}()},
\code{\link{plotChronEnsDiff}()},
\code{\link{plotChron}()},
\code{\link{plotCorEns}()},
\code{\link{plotHistEns}()},
\code{\link{plotLine}()},
\code{\link{plotModelDistributions}()},
\code{\link{plotPcaEns}()},
\code{\link{plotPvalsEnsFdr}()},
\code{\link{plotRegressEns}()},
\code{\link{plotScatterEns}()},
\code{\link{plotSpectraEns}()},
\code{\link{plotSpectrum}()},
\code{\link{plotSummary}()},
\code{\link{plotTimeseriesEnsLines}()},
\code{\link{plotTimeseriesEnsRibbons}()},
\code{\link{plotTimeseriesStack}()},
\code{\link{plotTrendLinesEns}()},
\code{\link{reverselog10_trans}()}

Other chron: 
\code{\link{plotChronEnsDiff}()},
\code{\link{plotChron}()},
\code{\link{plotModelDistributions}()}
}
\author{
Nick McKay
}
\concept{chron}
\concept{plot}
