% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.geoChronR.R
\name{plotChronEnsDiff}
\alias{plotChronEnsDiff}
\title{Compare chron ensemble with paleoData age-model}
\usage{
plotChronEnsDiff(
  L,
  ageEnsVar = "ageEnsemble",
  age.var = "age",
  depth.var = "depth",
  paleo.num = NA,
  paleo.meas.table.num = NA,
  chron.num = NA,
  model.num = NA,
  ens.table.num = NA,
  max.ensemble.members = NA,
  strict.search = FALSE,
  probs = c(0.025, 0.25, 0.5, 0.75, 0.975),
  x.bin = NA,
  y.bin = NA,
  n.bins = 100,
  color.low = "white",
  color.high = "grey70",
  alp = 1,
  color.line = "Black",
  line.width = 1,
  add.to.plot = ggplot2::ggplot(),
  n.ens.plot = 5,
  color.ens.line = "red",
  alp.ens.line = 0.7
)
}
\arguments{
\item{L}{A LiPD object - an R serialization of a single LiPD file. It's a list, and is typically created by `readLipd()`}

\item{ageEnsVar}{name of the age ensemble variable in the chronData to search for}

\item{age.var}{name of the age variable in the paleoData to search for}

\item{depth.var}{name of the depth variable to search for}

\item{paleo.num}{an integer that corresponds to paleo.numData object (L$paleoData[[?]]) has the measurementTable you want to modify}

\item{paleo.meas.table.num}{an integer that corresponds to paleo.num measurementTable you want to add the ensemble to?}

\item{chron.num}{an integer that corresponds to chron.numData object (L$crhonData[[?]]) has the model you want to get the ensemble from}

\item{model.num}{an integer that corresponds to chron.num model you want to get the ensemble from?}

\item{ens.table.num}{an integer that corresponds to chron.num model ensembleTable you want to get the ensemble from?}

\item{max.ensemble.members}{Maximum number of ensemble members to map}

\item{strict.search}{Use a strict.search to look for the ageEnsemble and depth variables. TRUE(default) or FALSE   #' @param probs quantiles to calculate and plot}

\item{probs}{quantiles to plot with ribbons}

\item{x.bin}{vector of bin edges over which to bin.}

\item{y.bin}{vector of bin edges over which to bin.}

\item{n.bins}{number bins over which to calculate intervals. Used to calculate x.bin if not provided.}

\item{color.low}{Band color of the outer most band.}

\item{color.high}{Band color of the inner most band.}

\item{alp}{Transparency of the band plot}

\item{color.line}{Line color (following ggplot rules)}

\item{line.width}{Width of the line}

\item{add.to.plot}{A ggplot object to add this plot to. Default is ggplot() .}

\item{n.ens.plot}{Number of ensemble members to plot}

\item{color.ens.line}{color of the ensemble lines}

\item{alp.ens.line}{transparency of the lines}
}
\value{
A ggplot object
}
\description{
Plots the difference of an chron ensembleTable with the paleoData age
}
\seealso{
Other plot: 
\code{\link{plotChron}()},
\code{\link{plotChronEns}()},
\code{\link{plotCorEns}()},
\code{\link{plotHistEns}()},
\code{\link{plotLine}()},
\code{\link{plotModelDistributions}()},
\code{\link{plotPcaEns}()},
\code{\link{plotPvalsEnsFdr}()},
\code{\link{plotRegressEns}()},
\code{\link{plotScatterEns}()},
\code{\link{plotScreeEns}()},
\code{\link{plotSpectraEns}()},
\code{\link{plotSpectrum}()},
\code{\link{plotSummary}()},
\code{\link{plotSummaryTs}()},
\code{\link{plotTimeseriesEnsLines}()},
\code{\link{plotTimeseriesEnsRibbons}()},
\code{\link{plotTimeseriesStack}()},
\code{\link{plotTrendLinesEns}()}

Other chron: 
\code{\link{plotChron}()},
\code{\link{plotChronEns}()},
\code{\link{plotModelDistributions}()}
}
\author{
Nick McKay
}
\concept{chron}
\concept{plot}
