% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-analysis.R
\name{plotCorEns}
\alias{plotCorEns}
\title{Plot the results of an ensemble correlation}
\usage{
plotCorEns(
  corout,
  bins = 40,
  line.labels = corout$cor.stats$percentiles,
  add.to.plot = ggplot2::ggplot(),
  legend.position = c(0.2, 0.8),
  f.sig.lab.position = c(0.15, 0.4),
  sig.level = 0.05,
  significance.option = "isospectral",
  use.fdr = TRUE,
  bar.colors = c("grey50", "Chartreuse4", "DarkOrange")
)
}
\arguments{
\item{corout}{output from corEns()}

\item{bins}{Number of bins in the histogram}

\item{line.labels}{Labels for the quantiles lines}

\item{add.to.plot}{A ggplot object to add these lines to. Default is ggplot()}

\item{legend.position}{Where to put the map legend?}

\item{f.sig.lab.position}{x,y (0-1) position of the fraction of significant correlation labels}

\item{sig.level}{What significance level to plot?}

\item{significance.option}{Choose how handle significance. Options are:
\itemize{
\item "raw" for uncorrected p-values
\item "eff-n" to adjust the test's sample size to reflect the reduction in degrees of freedom due to autocorrelation
\item "isopersistent" to estimate significance by generating surrogates, or random synthetic timeseries, that emulate the persistence characteristics of the series.
\item "isospectral" A non-parametric alternative which estimates significance by generating surrogates by scrambling the spectral phases of the two datasets, thus preserving their power spectrum while destroying the correlated signal. This is the recommended (and default) option.
}}

\item{use.fdr}{Use results from False Discovery Rate testing in plot?}

\item{bar.colors}{What colors to use for the bars, formatted as (insignificant, significant, significant after FDR)}
}
\value{
A ggplot object
}
\description{
Plots the output of an ensemble correlation analysis.
}
\seealso{
Other plot: 
\code{\link{plotChron}()},
\code{\link{plotChronEns}()},
\code{\link{plotChronEnsDiff}()},
\code{\link{plotHistEns}()},
\code{\link{plotLine}()},
\code{\link{plotModelDistributions}()},
\code{\link{plotPcaEns}()},
\code{\link{plotPvalsEnsFdr}()},
\code{\link{plotRegressEns}()},
\code{\link{plotScatterEns}()},
\code{\link{plotScreeEns}()},
\code{\link{plotSpectraEns}()},
\code{\link{plotSpectrum}()},
\code{\link{plotSummary}()},
\code{\link{plotSummaryTs}()},
\code{\link{plotTimeseriesEnsLines}()},
\code{\link{plotTimeseriesEnsRibbons}()},
\code{\link{plotTimeseriesStack}()},
\code{\link{plotTrendLinesEns}()}

Other correlation: 
\code{\link{ar1}()},
\code{\link{ar1Surrogates}()},
\code{\link{corMatrix}},
\code{\link{effectiveN}()},
\code{\link{pvalMonteCarlo}()},
\code{\link{pvalPearsonSerialCorrected}()}
}
\author{
Julien Emile-Geay

Nick McKay
}
\concept{correlation}
\concept{plot}
