% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.geoChronR.R
\name{plotCorrEns}
\alias{plotCorrEns}
\title{Plot the results of an ensemble correlation}
\usage{
plotCorrEns(cor.df, corStats, bins = 40, lineLabels = rownames(corStats),
  add.to.plot = ggplot())
}
\arguments{
\item{cor.df}{A data.frame correlation r and p-values. Output from corEns()}

\item{corStats}{A data.frame of correlation quantiles. Output from corEns()}

\item{bins}{Number of bins in the histogram}

\item{lineLabels}{Labels for the quantiles lines}

\item{add.to.plot}{A ggplot object to add these lines to. Default is ggplot() .}
}
\value{
A ggplot object
}
\description{
Plots the output of an ensemble correlation analysis.
}
\seealso{
Other plot: \code{\link{axisLabel}},
  \code{\link{geoChronRPlotTheme}},
  \code{\link{plotChron}}, \code{\link{plotHistEns}},
  \code{\link{plotLine}},
  \code{\link{plotModelDistributions}},
  \code{\link{plotPcaEns}}, \code{\link{plotPvalsEnsFdr}},
  \code{\link{plotRegressEns}},
  \code{\link{plotScatterEns}},
  \code{\link{plotSpectraAnnotate}},
  \code{\link{plotSpectraEns}}, \code{\link{plotSummary}},
  \code{\link{plotTimeseriesEnsLines}},
  \code{\link{plotTimeseriesEnsRibbons}},
  \code{\link{plotTrendLinesEns}}
}
\author{
Julien Emile-Geay

Nick McKay
}
