% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.geoChronR.R
\name{plotModelDistributions}
\alias{plotModelDistributions}
\title{Plot probability distributions}
\usage{
plotModelDistributions(L, dist.var = "age", y.var = "depth",
  mode = "chron", which.data = 1, model.num = 1, add.to.plot = ggplot(),
  alp = 0.5, color = "purple", scaleFrac = 0.02, equalArea = TRUE,
  dist.plot = NA, distType = "violin", thick = 0.1, truncateDist = NA)
}
\arguments{
\item{L}{A LiPD object}

\item{dist.var}{Name of the distribution variable, will be plotted along the x-axis. Use coord_flip() after running the function if you want vertical distributions. "age" by default.}

\item{y.var}{Name of the y-axis variable. "depth" by default.}

\item{mode}{chron or paleo}

\item{which.data}{number of the chron or paleo Data object}

\item{model.num}{number of the model object}

\item{add.to.plot}{A ggplot object to add this plot to. Default is ggplot() .}

\item{color}{distribution color (following ggplot rules)}

\item{scaleFrac}{controls the vertical span of the probability distribution. Approximately the vertical fraction of the plot that the distribution will cover.}

\item{dist.plot}{vector of distribution tables to plot}

\item{distType}{"violin" (default), "up" for one-sided distributions pointed up, "down" for one-sided distributions pointed down}

\item{thick}{thickness of the line around the distribution}

\item{truncateDist}{truncate probability density values below this number. NA (default) means no truncation}
}
\value{
A ggplot object
}
\description{
Plot or add probability distributions from a paleo or chron model to a plot.
}
\seealso{
Other plot: \code{\link{AD2BP_trans}},
  \code{\link{BP2AD_trans}},
  \code{\link{FrequencyAnnotate}},
  \code{\link{PeriodAnnotate}}, \code{\link{axisLabel}},
  \code{\link{geoChronRPlotTheme}},
  \code{\link{getLegend}}, \code{\link{getPlotRanges}},
  \code{\link{plotChronEnsDiff}},
  \code{\link{plotChronEns}}, \code{\link{plotChron}},
  \code{\link{plotCorEns}}, \code{\link{plotHistEns}},
  \code{\link{plotLine}}, \code{\link{plotPcaEns}},
  \code{\link{plotPvalsEnsFdr}},
  \code{\link{plotRegressEns}},
  \code{\link{plotScatterEns}},
  \code{\link{plotSpectraEns}}, \code{\link{plotSpectrum}},
  \code{\link{plotSummary}},
  \code{\link{plotTimeseriesEnsLines}},
  \code{\link{plotTimeseriesEnsRibbons}},
  \code{\link{plotTimeseriesStack}},
  \code{\link{plotTrendLinesEns}},
  \code{\link{reverselog10_trans}}

Other chron: \code{\link{plotChronEnsDiff}},
  \code{\link{plotChronEns}}, \code{\link{plotChron}}
}
\author{
Nick McKay
}
