% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-analysis.R
\name{plotScreeEns}
\alias{plotScreeEns}
\title{create a scree plot for PCA analysis}
\usage{
plotScreeEns(pcaout, null.color = "red", null.significance = 0.05, ...)
}
\arguments{
\item{pcaout}{a list of results output from pcaEns()}

\item{null.color}{color of the line of the null hypothesis results}

\item{null.significance}{Significance level to of the null to plot (default = 0.05).}

\item{...}{
  Arguments passed on to \code{\link[=plotTimeseriesEnsRibbons]{plotTimeseriesEnsRibbons}}
  \describe{
    \item{\code{X}}{A LiPD variable list to plot, including values, units, names, and more}
    \item{\code{Y}}{A LiPD variable list to plot, including values, units, names, and more}
    \item{\code{probs}}{a vector of probabilities to plot as ribbons. It will create bands as ribbons of quantiles moving inward. If there's an odd number, it plots the middle quantile as a line.}
    \item{\code{color.low}}{Color of the outermost band; the extreme quantiles of the distribution}
    \item{\code{color.high}}{Color of the innermost band; the central quantiles of the distribution}
    \item{\code{color.line}}{Line color (following ggplot rules)}
    \item{\code{color.vector}}{A vector (of length equal to the number of bands) that specifies the colors for the ribbons from the outermost band in (default = NA). Colors specified as string according to ggplot2 conventions. If present, this overrules color.high and color.low}
    \item{\code{line.width}}{Width of the line}
    \item{\code{n.bins}}{number bins over which to calculate intervals. Used to calculate x.bin if not provided.}
    \item{\code{x.bin}}{vector of bin edges over which to bin.}
    \item{\code{y.bin}}{vector of bin edges over which to bin.}
    \item{\code{alp}}{alpha (transparency) parameter for the ribbons}
    \item{\code{add.to.plot}}{A ggplot object to add this plot to. Default is ggplot() .}
    \item{\code{export.quantiles}}{If TRUE, return the plotted quantiles rather than the plot}
  }}
}
\value{
a ggplot plot
}
\description{
create a scree plot for PCA analysis
}
\seealso{
Other pca: 
\code{\link{ar1Surrogates}()},
\code{\link{createSyntheticTimeseries}()},
\code{\link{pcaEns}()},
\code{\link{plotPcaEns}()}

Other plot: 
\code{\link{plotChron}()},
\code{\link{plotChronEns}()},
\code{\link{plotChronEnsDiff}()},
\code{\link{plotCorEns}()},
\code{\link{plotHistEns}()},
\code{\link{plotLine}()},
\code{\link{plotModelDistributions}()},
\code{\link{plotPcaEns}()},
\code{\link{plotPvalsEnsFdr}()},
\code{\link{plotRegressEns}()},
\code{\link{plotScatterEns}()},
\code{\link{plotSpectraEns}()},
\code{\link{plotSpectrum}()},
\code{\link{plotSummary}()},
\code{\link{plotSummaryTs}()},
\code{\link{plotTimeseriesEnsLines}()},
\code{\link{plotTimeseriesEnsRibbons}()},
\code{\link{plotTimeseriesStack}()},
\code{\link{plotTrendLinesEns}()}
}
\concept{pca}
\concept{plot}
