% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.geoChronR.R
\name{plotSpectraAnnotate}
\alias{plotSpectraAnnotate}
\title{Annotate plot of spectra with periodicities}
\usage{
plotSpectraAnnotate(specPlot, periods = c(19, 23, 41, 100))
}
\arguments{
\item{specPlot}{Output from plotSpectraEns (or other ggplot)}
}
\value{
ggplot object of spectrum plot
}
\description{
Annotate plot of spectra (ensemble or otherwise) with vertical lines at specific periodicities
}
\seealso{
Other plot: \code{\link{axisLabel}},
  \code{\link{geoChronRPlotTheme}},
  \code{\link{plotChron}}, \code{\link{plotCorrEns}},
  \code{\link{plotHistEns}}, \code{\link{plotLine}},
  \code{\link{plotModelDistributions}},
  \code{\link{plotPcaEns}}, \code{\link{plotPvalsEnsFdr}},
  \code{\link{plotRegressEns}},
  \code{\link{plotScatterEns}},
  \code{\link{plotSpectraEns}}, \code{\link{plotSummary}},
  \code{\link{plotTimeseriesEnsLines}},
  \code{\link{plotTimeseriesEnsRibbons}},
  \code{\link{plotTrendLinesEns}}

Other spectra: \code{\link{ar1Surrogates}},
  \code{\link{computeSpectraEns}},
  \code{\link{plotSpectraEns}}
}
