% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.geoChronR.R
\name{plotSpectraEns}
\alias{plotSpectraEns}
\title{Plot ensemble spectra output}
\usage{
plotSpectraEns(spec.ens)
}
\arguments{
\item{spec.ens}{Output from powerSpectrumEns()}
}
\value{
ggplot object of spectrum plot
}
\description{
Plot the output of powerSpectrumEns() as a ribbon plot of distributions, plus confidence levels
}
\seealso{
Other plot: \code{\link{geoChronRPlotTheme}},
  \code{\link{plotCorrEns}}, \code{\link{plotHistEns}},
  \code{\link{plotLine}}, \code{\link{plotPvalsEnsFdr}},
  \code{\link{plotScatterEns}}, \code{\link{plotSummary}},
  \code{\link{plotTimeseriesEnsLines}},
  \code{\link{plotTimeseriesEnsRibbons}},
  \code{\link{plotTrendLinesEns}}
}
