% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.geoChronR.R
\name{plotSummary}
\alias{plotSummary}
\title{Plot a summary figure}
\usage{
plotSummary(
  L,
  paleo.age.var = "age",
  paleo.data.var = NA,
  chron.number = NA,
  paleo.meas.num = NA,
  chron.meas.num = NA,
  chron.depth.var = "depth",
  chron.age.var = "age",
  dotSize = 5,
  summary.font.size = 10,
  text.width = 400/summary.font.size,
  legend.position = c(0.7, 0.3),
  ...
)
}
\arguments{
\item{L}{A LiPD object - an R serialization of a single LiPD file. It's a list, and is typically created by `readLipd()`}

\item{paleo.age.var}{variableName to use for x axis of the paleo plot ("age" by default)}

\item{paleo.data.var}{variableName to use for the y axis of the paleo plot (NA by default, which lets you choose)}

\item{chron.number}{chron.numData object to use (NA by default, will ask if needed)}

\item{paleo.meas.num}{paleo.num measurement table to use (NA by default, will ask if needed)}

\item{chron.meas.num}{chron.num measurement table to use (NA by default, will ask if needed)}

\item{chron.depth.var}{variableName to use for chron depth ("depth" by default)}

\item{chron.age.var}{variableName to use for chron age ("age" by default)}

\item{...}{
  Arguments passed on to \code{\link[=plotChronEns]{plotChronEns}}
  \describe{
    \item{\code{depth.var}}{variableName to use for depth ("depth" by default)}
    \item{\code{age.var}}{ariableName to use for age ensemble ("ageEnsemble" by default)}
    \item{\code{model.num}}{chron.numData model to use (NA by default, will ask if needed)}
    \item{\code{probs}}{quantiles to calculate and plot}
    \item{\code{n.bins}}{number bins over which to calculate intervals. Used to calculate x.bin if not provided.}
    \item{\code{x.bin}}{vector of bin edges over which to bin.}
    \item{\code{y.bin}}{vector of bin edges over which to bin.}
    \item{\code{color.low}}{Band color of the outer most band.}
    \item{\code{color.high}}{Band color of the inner most band.}
    \item{\code{alp}}{Transparency of the band plot}
    \item{\code{color.line}}{Line color (following ggplot rules)}
    \item{\code{line.width}}{Width of the line}
    \item{\code{add.to.plot}}{A ggplot object to add this plot to. Default is ggplot() .}
    \item{\code{n.ens.plot}}{Number of ensemble members to plot}
    \item{\code{color.ens.line}}{color of the ensemble lines}
    \item{\code{alp.ens.line}}{transparency of the lines}
    \item{\code{dist.color}}{distribution color (following ggplot rules)}
    \item{\code{dist.type}}{"violin" (default), "up" for one-sided distributions pointed up, "down" for one-sided distributions pointed down}
    \item{\code{dist.thick}}{thickness of the line around the distribution}
    \item{\code{truncate.dist}}{truncate probability density values below this number. NA (default) means no truncation}
    \item{\code{dist.scale}}{controls the vertical span of the probability distribution. Approximately the vertical fraction of the plot that the distribution will cover.}
    \item{\code{add.paleo.age.depth}}{add a line that shows the paleoData age depth.}
    \item{\code{paleo.number}}{paleo.num number for the paleoData age-depth}
    \item{\code{meas.num}}{which measurement Table for the paleoData age-depth}
    \item{\code{color.line.paleo}}{line color of the paleoData age-depth (following ggplot rules)}
  }}

\item{dot.size}{what size dot for the chron plot? Only used if not plotting by plotChronEns() (default = 5)}
}
\value{
A gridArrange of ggplot grobs
}
\description{
shows a map, timeseries, and age model diagram, and basic simple metadata
}
\examples{
\dontrun{
myPlot = summaryPlot(L)
}
}
\seealso{
Other plot: 
\code{\link{AD2BP_trans}()},
\code{\link{BP2AD_trans}()},
\code{\link{axisLabel}()},
\code{\link{geoChronRPlotTheme}()},
\code{\link{getLegend}()},
\code{\link{getPlotRanges}()},
\code{\link{periodAnnotate}()},
\code{\link{plotChronEnsDiff}()},
\code{\link{plotChronEns}()},
\code{\link{plotChron}()},
\code{\link{plotCorEns}()},
\code{\link{plotHistEns}()},
\code{\link{plotLine}()},
\code{\link{plotModelDistributions}()},
\code{\link{plotPcaEns}()},
\code{\link{plotPvalsEnsFdr}()},
\code{\link{plotRegressEns}()},
\code{\link{plotScatterEns}()},
\code{\link{plotSpectraEns}()},
\code{\link{plotSpectrum}()},
\code{\link{plotTimeseriesEnsLines}()},
\code{\link{plotTimeseriesEnsRibbons}()},
\code{\link{plotTimeseriesStack}()},
\code{\link{plotTrendLinesEns}()},
\code{\link{reverselog10_trans}()}
}
\author{
Nick McKay
}
\concept{plot}
