% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.functions.R
\name{plotSummaryTs}
\alias{plotSummaryTs}
\title{TS overview plot}
\usage{
plotSummaryTs(
  TS,
  sort.var = "archiveType",
  age.range = NA,
  age.var = "age",
  step = NA,
  ...
)
}
\arguments{
\item{TS}{LiPD timeseries object See \url{http://nickmckay.github.io/LiPD-utilities/r/index.html#what-is-a-time-series}}

\item{sort.var}{What variable in the TS should the map and histogram be organized by? (default = "archiveType)}

\item{age.range}{a two element vector that includes the range over which you want to count (default = NA, which calculates the full range in the TS)}

\item{age.var}{What age variable should be used (default = "age)}

\item{step}{What step size should be used along the age axis? (default = NA, which will calculate 0.5\% of the age.range)}

\item{...}{
  Arguments passed on to \code{\link[=mapTs]{mapTs}}
  \describe{
    \item{\code{color}}{variable (in TS) by which to color, or a string specifying a static ggplot color}
    \item{\code{size}}{Size of the location marker}
    \item{\code{lat.range}}{Latitudes to use to create baseMap range (default = NA, which determines from TS)}
    \item{\code{lon.range}}{Longitudes to use to create baseMap range (default = NA, which determines from TS)}
    \item{\code{shape}}{variable (in TS) by which to adjust shape, or an integer specifying a static ggplot shape}
  }}
}
\value{
a gridExtra ggplot output
}
\description{
TS overview plot
}
\seealso{
Other plot: 
\code{\link{plotChronEnsDiff}()},
\code{\link{plotChronEns}()},
\code{\link{plotChron}()},
\code{\link{plotCorEns}()},
\code{\link{plotHistEns}()},
\code{\link{plotLine}()},
\code{\link{plotModelDistributions}()},
\code{\link{plotPcaEns}()},
\code{\link{plotPvalsEnsFdr}()},
\code{\link{plotRegressEns}()},
\code{\link{plotScatterEns}()},
\code{\link{plotScreeEns}()},
\code{\link{plotSpectraEns}()},
\code{\link{plotSpectrum}()},
\code{\link{plotSummary}()},
\code{\link{plotTimeseriesEnsLines}()},
\code{\link{plotTimeseriesEnsRibbons}()},
\code{\link{plotTimeseriesStack}()},
\code{\link{plotTrendLinesEns}()}
}
\concept{plot}
