% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.geoChronR.R
\name{plotTimeseriesEnsRibbons}
\alias{plotTimeseriesEnsRibbons}
\title{Plot an ensemble timeseries as ribbons of probabilities}
\usage{
plotTimeseriesEnsRibbons(
  add.to.plot = ggplot(),
  X,
  Y,
  alp = 1,
  probs = c(0.025, 0.25, 0.5, 0.75, 0.975),
  x.bin = NA,
  y.bin = NA,
  n.bins = 200,
  color.low = "white",
  color.high = "grey70",
  color.line = "Black",
  color.vector = NA,
  line.width = 1,
  export.quantiles = FALSE,
  ...
)
}
\arguments{
\item{add.to.plot}{A ggplot object to add this plot to. Default is ggplot() .}

\item{X}{A LiPD variable list to plot, including values, units, names, and more}

\item{Y}{A LiPD variable list to plot, including values, units, names, and more}

\item{alp}{alpha (transparency) parameter for the ribbons}

\item{probs}{a vector of probabilities to plot as ribbons. It will create bands as ribbons of quantiles moving inward. If there's an odd number, it plots the middle quantile as a line.}

\item{x.bin}{vector of bin edges over which to bin.}

\item{y.bin}{vector of bin edges over which to bin.}

\item{n.bins}{number bins over which to calculate intervals. Used to calculate x.bin if not provided.}

\item{color.low}{Color of the outermost band; the extreme quantiles of the distribution}

\item{color.high}{Color of the innermost band; the central quantiles of the distribution}

\item{color.line}{Line color (following ggplot rules)}

\item{color.vector}{A vector (of length equal to the number of bands) that specifies the colors for the ribbons from the outermost band in (default = NA). Colors specified as string according to ggplot2 conventions. If present, this overrules color.high and color.low}

\item{line.width}{Width of the line}

\item{export.quantiles}{If TRUE, return the plotted quantiles rather than the plot}

\item{...}{
  Arguments passed on to \code{\link[=quantile2d]{quantile2d}}
  \describe{
    \item{\code{x}}{n by m matrix where n is the number of observations and m is >= 1}
    \item{\code{y}}{n by j matrix where n is the number of observations and j is >= 1}
    \item{\code{seed}}{set a seed for reproducibility}
    \item{\code{n.ens}}{how many ensemble members?}
    \item{\code{limit.outliers.x}}{limit the plotting of outliers on the x axis to exclude values below this probability limit (default = 0.025)}
  }}
}
\value{
A ggplot object OR list of plotted quantiles, depending on export.quantiles
}
\description{
Plot an ensemble timeseries as a set of bands of probability. Useful for displaying the full range of probability across ensemble members.
}
\seealso{
Other plot: 
\code{\link{plotChronEnsDiff}()},
\code{\link{plotChronEns}()},
\code{\link{plotChron}()},
\code{\link{plotCorEns}()},
\code{\link{plotHistEns}()},
\code{\link{plotLine}()},
\code{\link{plotModelDistributions}()},
\code{\link{plotPcaEns}()},
\code{\link{plotPvalsEnsFdr}()},
\code{\link{plotRegressEns}()},
\code{\link{plotScatterEns}()},
\code{\link{plotScreeEns}()},
\code{\link{plotSpectraEns}()},
\code{\link{plotSpectrum}()},
\code{\link{plotSummaryTs}()},
\code{\link{plotSummary}()},
\code{\link{plotTimeseriesEnsLines}()},
\code{\link{plotTimeseriesStack}()},
\code{\link{plotTrendLinesEns}()}
}
\author{
Nick McKay
}
\concept{plot}
