% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.geoChronR.R
\name{plotTimeseriesStack}
\alias{plotTimeseriesStack}
\title{Plot a bunch of timeseries in a vertical stack}
\usage{
plotTimeseriesStack(
  plot.df,
  time.var = "year",
  color.var = "paleoData_TSid",
  invert.var = NA,
  fill.alpha = 0.2,
  line.size = 0.5,
  scale.factor = 1/3,
  scale.height = 0.75,
  lab.buff = 0.02,
  lab.size = 3,
  lab.space = 2,
  color.ramp = function(nColors) {     RColorBrewer::brewer.pal(nColors, "Dark2") }
)
}
\arguments{
\item{plot.df}{A tidy data.frame, typically the output of tidyTs()}

\item{time.var}{Which variable to put on the x-axis. Must be in plot.df. Typically "year", "age", or "depth"}

\item{color.var}{Which variable to color the timeseries by. The default ("paleoData_TSid") will give each timeseries it's own color. Common other options include "paleoData_variable", "archiveType", or "paleoData_units", but any variable in plot.df should work.}

\item{invert.var}{Which variable to use to invert the timeseries. This should point to a variable of "positive" and "negative" (searches on "neg"), or a vector of 1s and -1s. (default = NA, which flips nothing)}

\item{fill.alpha}{Transparency of the shading}

\item{line.size}{thickness of the line (default = 0.5)}

\item{scale.factor}{Controls how much the timeseries should overlap, with larger numbers overlapping more. (default = 1/3)}

\item{scale.height}{Controls how large the y-axes will be. 1 is equivalent to end-to-end coverage with no space. (default = 0.75)}

\item{lab.buff}{Fraction of the x axis to space the tick marks away from the axes bars (default = 0.02)}

\item{lab.size}{Font size for the ylabels}

\item{lab.space}{Multiplier on lab.buff for the axis label separation from the y-scale}

\item{color.ramp}{Specify the colors to use in the plot arranged along color.var. (default = function(nColors){RColorBrewer::brewer.pal(nColors,"Dark2")})}
}
\value{
A ggplot object of the plot
}
\description{
Creates a suite of plots to characterize the results of an ensemble regression.
}
\section{Long-form example}{

\href{http://nickmckay.github.io/GeoChronR/articles/PlotTimeseriesStack.html}{View a full-fledged example of how to use this function.}
}

\seealso{
Other plot: 
\code{\link{plotChronEnsDiff}()},
\code{\link{plotChronEns}()},
\code{\link{plotChron}()},
\code{\link{plotCorEns}()},
\code{\link{plotHistEns}()},
\code{\link{plotLine}()},
\code{\link{plotModelDistributions}()},
\code{\link{plotPcaEns}()},
\code{\link{plotPvalsEnsFdr}()},
\code{\link{plotRegressEns}()},
\code{\link{plotScatterEns}()},
\code{\link{plotScreeEns}()},
\code{\link{plotSpectraEns}()},
\code{\link{plotSpectrum}()},
\code{\link{plotSummaryTs}()},
\code{\link{plotSummary}()},
\code{\link{plotTimeseriesEnsLines}()},
\code{\link{plotTimeseriesEnsRibbons}()},
\code{\link{plotTrendLinesEns}()}
}
\author{
Nick McKay
}
\concept{plot}
