% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.spectra.ens.R
\name{powerSpectrumEns}
\alias{powerSpectrumEns}
\title{Calculate ensemble power spectra}
\usage{
powerSpectrumEns(time, values, max.ens = NA, ofac = 1,
  gaussianize = FALSE)
}
\arguments{
\item{time}{LiPD "variable list" or vector of year/age values}

\item{values}{LiPD "variable list" or vector of values}

\item{max.ens}{Maximum number of ensemble members to analyze}

\item{ofac}{oversampling factor for lomb::lsp}
}
\value{
a list of ensemble spectra results
\itemize{
\item freqs: vector of frequencies
\item power: vector of spectral powers
\item powerSyn: matrix of synthetic spectral power results
}
}
\description{
Calculate ensemble power spectra using lomb-scargle
}
\seealso{
Other spectra: \code{\link{ar1Surrogates}},
  \code{\link{createSyntheticTimeseries}},
  \code{\link{plotSpectraEns}}
}
