% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdr.R
\name{propAlt}
\alias{propAlt}
\title{Proportion of alternate hypotheses}
\usage{
propAlt(pvals, adjustment.method = "mean", adjustment.args = list(edf.lower
  = 0.8, num.steps = 20))
}
\arguments{
\item{pvals}{a vector of pvals on which to conduct the multiple testing}

\item{adjustment.method}{method for increasing the power of the procedure by estimating the proportion of alternative p-values, one of "mean", the modified Storey estimator that we suggest in Ventura et al. (2004), "storey", the method of Storey (2002), or "two-stage", the iterative approach of Benjamini et al. (2001)}

\item{adjustment.args}{arguments to adjustment.method; see propAlt() for description, but note that for "two-stage", qlevel and fdr.method are taken from the qlevel and method arguments to fdr()}

\item{qlevel}{the proportion of false positives desired}
}
\value{
estimate of a, the number of alternative hypotheses
}
\description{
This is an internal function that calculates an estimate of a, the proportion of alternative hypotheses, using one of several methods.
}
\examples{
a <- propAlt(pvals,adjustment.method="mean")
}
\seealso{
Other FDR: \code{\link{fdr.master}},
  \code{\link{fdrBasic}}, \code{\link{fdr}},
  \code{\link{storey}}
}
\author{
Chris Paciorek
}
