% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembleComposite.R
\name{pullInstance}
\alias{pullInstance}
\title{Pull one instance from an ensemble}
\usage{
pullInstance(X)
}
\arguments{
\item{X}{A LiPD column object}
}
\value{
a vector randomly drawn from the matrix
}
\description{
Pull one instance from an ensemble
}
