% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipd.manipulation.R
\name{pullTsVariable}
\alias{pullTsVariable}
\title{pull variable out of TS object}
\usage{
pullTsVariable(TS, variable)
}
\arguments{
\item{TS}{a LiPD Timeseries object}

\item{variable}{the name of variable in a TS object}
}
\description{
pulls all instances of a single variable out of a TS
}
\seealso{
Other LiPD manipulation: \code{\link{estimateUncertaintyFromRange}},
  \code{\link{flipCoords}}, \code{\link{getVariableIndex}},
  \code{\link{mapAgeEnsembleToPaleoData}},
  \code{\link{pushTsVariable}}, \code{\link{selectData}}
}
