% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipd.manipulation.R
\name{pushTsVariable}
\alias{pushTsVariable}
\title{push variable into of TS object}
\usage{
pushTsVariable(TS, variable, vec, createNew = FALSE)
}
\arguments{
\item{TS}{a LiPD Timeseries object}

\item{variable}{the name of variable in a TS object}

\item{vec}{a vector of data to be added to the TS object}

\item{createNew}{allow the function to create a new variable in the TS?}
}
\description{
pulls all instances of a single variable out of a TS
}
\seealso{
Other LiPD manipulation: \code{\link{estimateUncertaintyFromRange}},
  \code{\link{flipCoords}}, \code{\link{getVariableIndex}},
  \code{\link{mapAgeEnsembleToPaleoData}},
  \code{\link{pullTsVariable}}, \code{\link{selectData}},
  \code{\link{tidyTs}}
}
