% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.regression.functions.R
\name{pvalMonteCarlo}
\alias{pvalMonteCarlo}
\title{Correlations and their significance according to AR(1) benchmarks}
\usage{
pvalMonteCarlo(X, Y, n.sim = 100, method = "ebisuzaki")
}
\arguments{
\item{X}{a 1-column vector}

\item{Y}{a 1-column vector of the same}

\item{nsim}{number of simulations}
}
\value{
output
}
\description{
Fits AR(1) model to two series X & Y
}
\author{
Julien Emile-Geay

Nick McKay
}
