% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.regression.functions.R
\name{regress}
\alias{regress}
\title{Simple ordinary least squeares regression}
\usage{
regress(X, Y)
}
\arguments{
\item{X}{a matrix of predictor data}

\item{Y}{a vector of predictand data}
}
\value{
model coefficients
}
\description{
Simple regression function. Faster than lm()
}
\author{
Nick McKay
}
