% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.regression.functions.R
\name{regressEns}
\alias{regressEns}
\title{Ensemble regression}
\usage{
regressEns(
  time.1,
  values.1,
  time.2,
  values.2,
  bin.vec = NA,
  bin.step = NA,
  bin.fun = mean,
  max.ens = NA,
  percentiles = c(0.025, 0.25, 0.5, 0.75, 0.975),
  recon.bin.vec = NA,
  min.obs = 10
)
}
\arguments{
\item{time.1}{matrix of age/time ensembles, or single column}

\item{values.1}{matrix of values ensembles, or single column}

\item{time.2}{matrix of age/time ensembles, or single column}

\item{values.2}{matrix of values ensembles, or single column}

\item{bin.vec}{vector of bin edges for binning step}

\item{bin.step}{spacing of bins, used to build bin step}

\item{bin.fun}{function to use during binning (mean, sd, and sum all work)}

\item{max.ens}{maximum number of ensemble members to regress}

\item{percentiles}{quantiles to calculate for regression parameters}

\item{recon.bin.vec}{bin vector to use for the modeled regression.}

\item{min.obs}{minimum number of points required to calculate regression}
}
\value{
list of ensemble output
}
\description{
This is the primary function for ensemble regression. It will take ensemble values in time and/or values in the predictor (X), and regress them on ensemble values in time and/or values in Y (the predictand). The function will then apply the ensemble linear model to the full length of X to create a modeled Y. Will also optionally create plots.
}
\seealso{
Other regress: 
\code{\link{plotRegressEns}()},
\code{\link{plotScatterEns}()},
\code{\link{plotTrendLinesEns}()}
}
\author{
Nick McKay
}
\concept{regress}
