% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacon.lipd.R
\name{runBacon}
\alias{runBacon}
\title{Generate a Bayesian Reconstruction Age Model  (Bacon) and add it into a LiPD object}
\usage{
runBacon(L, which.chron = NA, which.mt = NA, baconDir = NA,
  site.name = L$dataSetName, modelNum = NA, remove.rejected = TRUE,
  overwrite = TRUE, cc = NA, maxEns = 1000, useMarine = NULL,
  labIDVar = "labID", age14CVar = "age14C",
  age14CuncertaintyVar = "age14CUnc", ageVar = "age",
  ageUncertaintyVar = "ageUnc", depthVar = "depth",
  reservoirAge14CVar = "reservoirAge",
  reservoirAge14CUncertaintyVar = "reservoirAge14C",
  rejectedAgesVar = "rejected", BaconAsk = TRUE, BaconSuggest = TRUE,
  baconThick = 5, baconAccMean = 20)
}
\arguments{
\item{L}{a single LiPD object}

\item{which.chron}{the number of the chronData object that you'll be working in}

\item{which.mt}{the number of the measurementTable you'll be working in}

\item{baconDir}{the directory where Bacon is installed on this computer.}

\item{site.name}{the name used for the bacon model (and directories)}

\item{modelNum}{which chronModel do you want to use?}

\item{remove.rejected}{don't write out dates that are marked as rejected}

\item{overwrite}{overwrite files and directories}

\item{cc}{An integer, or vector of integers corresponding to age that describes the calibration curve. You can specify here (see below) or if it's NA the code will guess based on archiveType
\itemize{
\item cc=1 IntCal13
\item cc=2 MarineCal 
\item cc=3 SHCal13
}}

\item{maxEns}{the maximum number of ensembles to load in (default = 1000)}
}
\value{
L The single LiPD object that was entered, with methods, ensembleTable, summaryTable and distributionTable added to the chronData model.
}
\description{
This is a high-level function that uses Bacon to simulate an age model, and stores this as an age-ensemble in a model in chronData. If needed input variables are not entered, and cannot be deduced, it will run in interactive mode. See Blaauw and Christen (2011) doi:10.1214/11-BA618 for details.
}
\examples{
Run in interactive mode:
L = runBacon(L)

Run in noninteractive mode, describing everything:
L = runBacon(L,which.chron = 1, which.mt = 1, modelNum = 3, baconDir = "~/Bacon/",site.name = "MSB2K", cc = 1)
}
\seealso{
Other Bacon: \code{\link{loadBaconOutput}},
  \code{\link{sampleBaconAges}}, \code{\link{writeBacon}}
}
\author{
Nick McKay

Maarten Blaauw (Bacon)
}
