% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.clam.LiPD.R
\name{runClam}
\alias{runClam}
\title{Run a clam model}
\usage{
runClam(L, which.chron = 1, clamDir = NA, remove.reverse = TRUE,
  overwrite = TRUE, cc = NA, site.name = L$dataSetName, modelNum = NA)
}
\description{
Not functional yet
}
