% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxcal.R
\name{runOxcal}
\alias{runOxcal}
\title{Generate an oxcal Age Model and add it into a LiPD object}
\usage{
runOxcal(L, which.chron = NA, which.mt = NA, oxcalPath = NA,
  modelNum = NA, remove.rejected = TRUE, overwrite = TRUE,
  maxEns = 1000, staticReservoirAge = NA, staticReservoirAgeUnc = NA,
  surfaceAge = NA, surfaceAgeUnc = NA, surfaceAgeDepth = 0,
  depthInterval = NA, eventsPerUnitLength = NA,
  eventsPerUnitLengthUncertainty = 0, outlierProb = 0.05,
  calCurve = "IntCal13", ...)
}
\arguments{
\item{L}{a single LiPD object}

\item{which.chron}{the number of the chronData object that you'll be working in}

\item{which.mt}{the number of the measurementTable you'll be working in}

\item{modelNum}{which chronModel do you want to use?}

\item{remove.rejected}{don't write out dates that are marked as rejected}

\item{overwrite}{overwrite files and directories}

\item{maxEns}{the maximum number of ensembles to load in (default = 1000)}

\item{staticReservoirAge}{optionally assign a deltaR to all radiocarbon dates}

\item{staticReservoirAgeUnc}{optionally assign a deltaR uncertainty to all radiocarbon dates}

\item{...}{arguments to pass to createChronMeasInputDf}
}
\value{
L The single LiPD object that was entered, with methods, ensembleTable, summaryTable and distributionTable added to the chronData model.
}
\description{
This is a high-level function that uses oxcal to simulate an age model, and stores this as an age-ensemble in a model in chronData. If needed input variables are not entered, and cannot be deduced, it will run in interactive mode. See Bronk Ramsey et al. 2008 doi:10.1016/j.quascirev.2007.01.019
}
\examples{
Run in interactive mode:
L = runOxcal(L)

Run in noninteractive mode, describing everything:
L = runOxcal(L,which.chron = 1, which.mt = 1, modelNum = 3, baconDir = "~/Bacon/",site.name = "MSB2K", cc = 1)
}
\author{
Nick McKay
}
