% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxcal.R
\name{runOxcalModel}
\alias{runOxcalModel}
\title{Title}
\usage{
runOxcalModel(
  L,
  model.text,
  model.num = NA,
  chron.num = NA,
  oxcal.path = "~/OxCal/",
  overwrite = TRUE,
  max.ens = 1000
)
}
\arguments{
\item{L}{a LiPD object}

\item{model.text}{a character vector of the model input}

\item{model.num}{chron.numModel do you want to use?}

\item{chron.num}{the number of the chronData object that you'll be working in}

\item{oxcal.path}{path to oxcal binaries}

\item{overwrite}{overwrite the previous run}

\item{max.ens}{the maximum number of ensembles to load in (default = 1000)}
}
\value{
a LiPD object
}
\description{
Title
}
