% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipd.manipulation.R
\name{selectData}
\alias{selectData}
\title{Select a LiPD "variable list"}
\usage{
selectData(
  L,
  var.name = NA,
  paleo.or.chron = "paleoData",
  paleo.or.chron.num = NA,
  table.type = "measurement",
  meas.table.num = NA,
  always.choose = FALSE,
  alt.names = NA,
  model.num = 1,
  ens.table.num = 1,
  sum.table.num = 1,
  strict.search = FALSE
)
}
\arguments{
\item{L}{A LiPD object - an R serialization of a single LiPD file. It's a list, and is typically created by `readLipd()`}

\item{var.name}{string name of the variable to extract}

\item{paleo.or.chron}{"paleoData" or "chronData"}

\item{paleo.or.chron.num}{an integer that corresponds to paleo.num or chron Data object (L$<paleo.or.chron>[[?]]) has the variable you want?}

\item{table.type}{What type of table do you want to select data from? ("measurement", "summary" or "ensemble")}

\item{meas.table.num}{an integer that corresponds to paleo.num measurementTable has the variable you want?}

\item{always.choose}{Force selection of the variable from a list}

\item{alt.names}{A vector of strings for alternative names to search for}

\item{model.num}{an integer that corresponds to model.num that has the variable you want}

\item{ens.table.num}{an integer that corresponds to ensembleTable you want to get the variable from?}

\item{sum.table.num}{an integer that corresponds to which summaryTable you want to get the variable from?}

\item{strict.search}{Use a strict.search to look for the ageEnsemble and depth variables. TRUE(default) or FALSE.}
}
\value{
A LiPD "variable list" object
}
\description{
Selects and extracts a LiPD "variable list"
}
\seealso{
Other LiPD manipulation: 
\code{\link{createConcatenatedEnsembleMeasurementTable}()},
\code{\link{createModel}()},
\code{\link{createMultiModelEnsemble}()},
\code{\link{createSummaryTableFromEnsembleTable}()},
\code{\link{createTSid}()},
\code{\link{estimateUncertaintyFromRange}()},
\code{\link{getVariableIndex}()},
\code{\link{mapAgeEnsembleToPaleoData}()},
\code{\link{mapPaleoEnsembleToPaleoData}()},
\code{\link{pullTsVariable}()}
}
\concept{LiPD manipulation}
