% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacon.lipd.R
\name{setBaconDir}
\alias{setBaconDir}
\title{Set the name of bacon core directory}
\usage{
setBaconDir(bacon.dir)
}
\arguments{
\item{bacon.dir}{if you already have it}
}
\value{
bacon.dir the bacon core directory
}
\description{
Use this to programmatically set the bacon.dir
}
\seealso{
Other Bacon: 
\code{\link{getBaconDir}()},
\code{\link{loadBaconOutput}()},
\code{\link{runBacon}()},
\code{\link{sampleBaconAges}()},
\code{\link{writeBacon}()}
}
\author{
Nick McKay
}
\concept{Bacon}
