% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacon.lipd.R
\name{setOxcalPath}
\alias{setOxcalPath}
\title{Set the oxcal path into geochronR's memory}
\usage{
setOxcalPath(path)
}
\arguments{
\item{path}{if you already have it}
}
\value{
baconDir the bacon core directory
}
\description{
Use this to programmatically set the oxcal path
}
\author{
Nick McKay
}
