% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{simulateAutoCorrelatedUncertainty}
\alias{simulateAutoCorrelatedUncertainty}
\title{Simulate uncertainty in a dataset}
\usage{
simulateAutoCorrelatedUncertainty(
  n,
  sd,
  mean = 0,
  ar = sqrt(0.5),
  arima.order = c(1, 0, 0),
  seed = NA
)
}
\arguments{
\item{n}{length of output vector}

\item{sd}{standard deviation of the output vector}

\item{mean}{mean of the output vector}

\item{ar}{Autocorrelation coefficient to use for modelling uncertainty, what fraction of the uncertainties are autocorrelated? (default = sqrt(0.5); or 50 percent autocorrelated uncertainty)}

\item{arima.order}{Order to use for ARIMA model used in modelling uncertainty (default = c(1,0,0))}

\item{seed}{Optionally enter an integer to set a seed for the random number generation}
}
\value{
ts simulated from a from an ARIMA model with a defined mean and variance
}
\description{
Simulate uncertainty in a dataset given an estimate of standard deviation and an ARIMA model
}
