% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BAM.lipd.R
\name{simulateBam}
\alias{simulateBam}
\title{Simulate a Banded Age Model (BAM)}
\usage{
simulateBam(X, t, model = NULL, ageEnsOut = FALSE)
}
\arguments{
\item{X}{data (vector or matrix n*p)}

\item{t}{chronology for data X (n*1)}

\item{model}{a list that describes the model to use in BAM
\itemize{
\item model$ns: number of samples 
\item model$name: 'poisson' or 'bernoulli' 
\item model$param: probability of growth band being perturbed (default: prob of missing band = prob of doubly-counted band = 0.05)
\itemize{
     \item if model$param is a single argument, then the perturbations are symmetric (prob of missing band = prob of doubly-counted band)
     \item if model$param = [a1 a2] and a1 neq a2 the model is asymmetric
     \itemize{
                 \item a1 = prob(missing layer) - undercounted
                 \item a2 = prob(layer counted multiple times) - overcounted
                 }
     \item if model$param: 2xp matrix, then different miscounting prob. are defined for each time series.
     }
\item model$resize: do not resize: 0 (default), resize to shortest sample: -1, resize to longest sample: 1
\item model$tm: if a time model is provided, the code returns the corresponding perturbed data
}}

\item{ageEnsOut}{TRUE or FALSE - return the ageEnsemble}
}
\value{
res a list with
\itemize{
\item res$Xc: realizations of age-perturbed data matrix of size tn*p*ns (could be 2 or 3d)
\item res$tc: new chronology tn*1
\item res$tmc: corresponding ensemble of time-correction matrices (tn*p*ns) to map realizations in Xp back to the original data X (2=insert nan, 0=remove double band) (2 or 3d)
where tn is the chronology length = n (default), shortest sample or longest sample depending on the chosen resizing option.
\item res$ageEnsemble (optional): Returnd the full age ensemble if desired. 
}
}
\description{
Generate an ensemble of possible age corrected data:See www.clim-past-discuss.net/9/6077/2013/ for a detailed description of the model.
The time series in X are automatically flipped to range from most recent to oldest measurements when the intput t is given in increasing order.
}
\examples{
res <- simulateBam(X,t) 
#will generate an ensemble of 1000 age models randomly following
#a Poisson process with rate parameter theta=0.05 used to perturb data X

res <- simulateBam(X,t,model) 
#will perturb data X  with the model specified in
#the model structure
}
\seealso{
Other BAM: \code{\link{bamCorrect}}, \code{\link{runBam}}
}
\author{
Maud Comboul
}
