% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdr.R
\name{storey}
\alias{storey}
\title{Storey estimator}
\usage{
storey(edf.quantile, pvals)
}
\arguments{
\item{edf.quantile}{(required):  the quantile of the empirical distribution function at which to estimate a}

\item{pvals}{(required):  a vector of pvals on which to conduct the multiple testing}
}
\value{
estimate of a, the number of alternative hypotheses
}
\description{
This is an internal function that calculates the basic Storey (2002) estimator of a, the proportion of alternative hypotheses.
}
\seealso{
Other FDR: 
\code{\link{fdr.master}()},
\code{\link{fdrBasic}()},
\code{\link{fdr}()},
\code{\link{propAlt}()}
}
\author{
Chris Paciorek
}
\concept{FDR}
