% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.regression.functions.R
\name{surrogateDataFun}
\alias{surrogateDataFun}
\title{Use the write function from rEDM depending on the version}
\usage{
surrogateDataFun(...)
}
\arguments{
\item{...}{Arguments to pass to surrogate}
}
\value{
Surrogate output
}
\description{
Use the write function from rEDM depending on the version
}
\seealso{
Other utility: 
\code{\link{askUser}()},
\code{\link{concatEnsembleTimeseries}()},
\code{\link{convertAD2BP}()},
\code{\link{convertBP2AD}()},
\code{\link{createChronMeasInputDf}()},
\code{\link{gaussianize}()},
\code{\link{generateEnsembleFromUncertainty}()},
\code{\link{getLastVarString}()},
\code{\link{getOs}()},
\code{\link{heuristicUnits}()},
\code{\link{loadRemote}()},
\code{\link{pullInstance}()},
\code{\link{simulateAutoCorrelatedUncertainty}()},
\code{\link{stringifyVariables}()}
}
\concept{utility}
