% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipd.manipulation.R
\name{tidyTs}
\alias{tidyTs}
\title{create tidy data.frame from TS}
\usage{
tidyTs(TS)
}
\arguments{
\item{TS}{a LiPD Timeseries object}
}
\description{
takes a TS object and turns it into a long, tidy, data.frame. Useful for data manipulation and analysis in the tidyverse and plotting
}
\seealso{
Other LiPD manipulation: \code{\link{estimateUncertaintyFromRange}},
  \code{\link{flipCoords}}, \code{\link{getVariableIndex}},
  \code{\link{mapAgeEnsembleToPaleoData}},
  \code{\link{pullTsVariable}},
  \code{\link{pushTsVariable}}, \code{\link{selectData}}
}
