% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.clam.LiPD.R
\name{writeClam}
\alias{writeClam}
\title{write files for running a clam model}
\usage{
writeClam(L, which.chron = 1, clamDir = NA, remove.reverse = TRUE,
  overwrite = TRUE, cc = NA, site.name = L$dataSetName, modelNum = NA)
}
\description{
Not functional yet
}
