% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipd.manipulation.R
\name{alignTimeseriesBin}
\alias{alignTimeseriesBin}
\title{Align and bin two timeseries into comparable bins}
\usage{
alignTimeseriesBin(
  time.1,
  values.1,
  time.2,
  values.2,
  bin.vec = NA,
  bin.step = NA,
  bin.fun = mean,
  max.ens = NA,
  min.obs = 10
)
}
\arguments{
\item{time.1}{matrix of age/time ensembles, or single column}

\item{values.1}{matrix of values ensembles, or single column}

\item{time.2}{matrix of age/time ensembles, or single column}

\item{values.2}{matrix of values ensembles, or single column}

\item{bin.vec}{vector of bin edges for binning step}

\item{bin.step}{spacing of bins, used to build bin step}

\item{bin.fun}{function to use during binning (mean, sd, and sum all work)}

\item{max.ens}{maximum number of ensemble members to regress}

\item{min.obs}{minimum number of points required to calculate regression}
}
\value{
list of binned data output:
\itemize{
\item binX: binned values from X
\item binY: binned values from Y
\item bin.step: interval of the binning
\item yearBins: bins along time
}
}
\description{
Use this to put two timeseries on different timesteps onto equivalent bins
}
\seealso{
Other bin: 
\code{\link{bin2d}()},
\code{\link{binEns}()},
\code{\link{binTs}()},
\code{\link{bin}()}
}
\author{
Nick McKay
}
\concept{bin}
