% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.regression.functions.R
\name{ar1}
\alias{ar1}
\title{Estimate Auto-Regressive coefficient at 1-timesetep}
\usage{
ar1(x)
}
\arguments{
\item{x}{a 1-column matrix or numeric dataset}
}
\value{
ar coefficient estimate of ar1
}
\description{
estimates ar1 using the arima() function
}
\seealso{
Other correlation: 
\code{\link{ar1Surrogates}()},
\code{\link{corMatrix}},
\code{\link{effectiveN}()},
\code{\link{plotCorEns}()},
\code{\link{pvalMonteCarlo}()},
\code{\link{pvalPearsonSerialCorrected}()}
}
\author{
Julien Emile-Geay
}
\concept{correlation}
