% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.spectra.ens.R
\name{ar1Surrogates}
\alias{ar1Surrogates}
\title{Create a synthetic timeseries that emulates the characteristics of a variable}
\usage{
ar1Surrogates(time, vals, detrend = TRUE, method = "redfit", n.ens = 1)
}
\arguments{
\item{time}{LiPD "variable list" or vector of year/age values}

\item{vals}{LiPD "variable list" or vector of values}

\item{detrend}{boolean variable, indicating whether one should remove a linear trend (default) or not. Note: does not affect the AR(1) fit, but does affect the generated timeseries}

\item{method}{Method used for estimation. Possible values are "even" or "redfit"}

\item{n.ens}{Number of ensemble members to simulate}
}
\value{
a vector or matrix of AR(1) surrogates for series X
}
\description{
create synthetic timeseries based on a timeseries. Useful for null hypothesis testing
}
\seealso{
Other spectra: 
\code{\link{createSyntheticTimeseries}()},
\code{\link{periodAnnotate}()},
\code{\link{plotSpectrum}()},
\code{\link{reverselog10_trans}()}

Other pca: 
\code{\link{createSyntheticTimeseries}()},
\code{\link{pcaEns}()},
\code{\link{plotPcaEns}()},
\code{\link{plotScreeEns}()}

Other correlation: 
\code{\link{ar1}()},
\code{\link{corMatrix}},
\code{\link{effectiveN}()},
\code{\link{plotCorEns}()},
\code{\link{pvalMonteCarlo}()},
\code{\link{pvalPearsonSerialCorrected}()}
}
\concept{correlation}
\concept{pca}
\concept{spectra}
