% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.functions.R
\name{assignColors}
\alias{assignColors}
\title{Make a base map}
\usage{
assignColors(colors = "temp")
}
\arguments{
\item{colors}{string to define color scale. Options are "temp", "precip" and "drought"}
}
\value{
a pair of high/low colors
}
\description{
Create a google or line map of the location of a list of LiPD objects
}
\seealso{
Other mapping: \code{\link{baseMap}},
  \code{\link{mapLipds}}, \code{\link{mapLipd}}
}
\author{
Nick McKay
}
