% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.geoChronR.R
\name{axisLabel}
\alias{axisLabel}
\title{Label axes}
\usage{
axisLabel(varList)
}
\arguments{
\item{varList}{LiPD "variable list"}
}
\value{
axis label as a string
}
\description{
Create an axis label string from a LiPD column vector
}
\seealso{
Other plot: \code{\link{geoChronRPlotTheme}},
  \code{\link{getLegend}}, \code{\link{plotChron}},
  \code{\link{plotCorEns}}, \code{\link{plotHistEns}},
  \code{\link{plotLine}},
  \code{\link{plotModelDistributions}},
  \code{\link{plotPcaEns}}, \code{\link{plotPvalsEnsFdr}},
  \code{\link{plotRegressEns}},
  \code{\link{plotScatterEns}},
  \code{\link{plotSpectraEns}}, \code{\link{plotSummary}},
  \code{\link{plotTimeseriesEnsLines}},
  \code{\link{plotTimeseriesEnsRibbons}},
  \code{\link{plotTimeseriesStack}},
  \code{\link{plotTrendLinesEns}}
}
\author{
Nick McKay
}
