% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.geoChronR.R
\name{axisLabel}
\alias{axisLabel}
\title{Label axes}
\usage{
axisLabel(varList)
}
\arguments{
\item{varList}{LiPD "variable list"}
}
\value{
axis label as a string
}
\description{
Create an axis label string from a LiPD column vector
}
\seealso{
Other plot help: 
\code{\link{AD2BP_trans}()},
\code{\link{BP2AD_trans}()},
\code{\link{bin2d}()},
\code{\link{geoChronRPlotTheme}()},
\code{\link{getLegend}()},
\code{\link{getPlotRanges}()},
\code{\link{kde2d}()},
\code{\link{meltDistributionTable}()},
\code{\link{modeSelektor}()},
\code{\link{periodAnnotate}()},
\code{\link{quantile2d}()},
\code{\link{reverselog10_trans}()}
}
\author{
Nick McKay
}
\concept{plot help}
