% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.regression.functions.R
\name{binEns}
\alias{binEns}
\title{Bin ensemble data}
\usage{
binEns(time, values, bin.vec, bin.fun = mean, max.ens = NA)
}
\arguments{
\item{time}{single column vector of time}

\item{values}{single column vector of values to bin}

\item{bin.vec}{vector of bin edges for binning step}

\item{bin.fun}{function to use during binning (mean, sd, and sum all work)}

\item{max.ens}{maximum number of ensemble members to regress}
}
\value{
list that includes matrix of binned data and binned time
}
\description{
takes ensembles in time and/or values and creates a matrix of data for future analysis
}
\seealso{
Other bin: 
\code{\link{alignTimeseriesBin}()},
\code{\link{bin}()},
\code{\link{bin2d}()},
\code{\link{binTs}()}
}
\concept{bin}
