% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.geoChronR.R
\name{BP2AD_trans}
\alias{BP2AD_trans}
\title{convert a BP age scale to AD}
\usage{
BP2AD_trans()
}
\description{
a ggplot scale to convert a BP axis to AD
}
\seealso{
Other plot help: 
\code{\link{AD2BP_trans}()},
\code{\link{axisLabel}()},
\code{\link{bin2d}()},
\code{\link{geoChronRPlotTheme}()},
\code{\link{getLegend}()},
\code{\link{getPlotRanges}()},
\code{\link{kde2d}()},
\code{\link{meltDistributionTable}()},
\code{\link{modeSelektor}()},
\code{\link{periodAnnotate}()},
\code{\link{quantile2d}()},
\code{\link{reverselog10_trans}()}
}
\concept{plot help}
