% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.geoChronR.R
\name{BP2AD_trans}
\alias{BP2AD_trans}
\title{convert a BP age scale to AD}
\usage{
BP2AD_trans()
}
\description{
a ggplot scale to convert a BP axis to AD
}
\seealso{
Other plot: \code{\link{AD2BP_trans}},
  \code{\link{FrequencyAnnotate}},
  \code{\link{PeriodAnnotate}}, \code{\link{axisLabel}},
  \code{\link{geoChronRPlotTheme}},
  \code{\link{getLegend}}, \code{\link{getPlotRanges}},
  \code{\link{plotChronEnsDiff}},
  \code{\link{plotChronEns}}, \code{\link{plotChron}},
  \code{\link{plotCorEns}}, \code{\link{plotHistEns}},
  \code{\link{plotLine}},
  \code{\link{plotModelDistributions}},
  \code{\link{plotPcaEns}}, \code{\link{plotPvalsEnsFdr}},
  \code{\link{plotRegressEns}},
  \code{\link{plotScatterEns}},
  \code{\link{plotSpectraEns}}, \code{\link{plotSpectrum}},
  \code{\link{plotSummary}},
  \code{\link{plotTimeseriesEnsLines}},
  \code{\link{plotTimeseriesEnsRibbons}},
  \code{\link{plotTimeseriesStack}},
  \code{\link{plotTrendLinesEns}},
  \code{\link{reverselog10_trans}}
}
