% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.spectra.ens.R
\name{computeSpectraEns}
\alias{computeSpectraEns}
\title{Calculate ensemble power spectra}
\usage{
computeSpectraEns(
  time,
  values,
  max.ens = NA,
  method = "mtm",
  probs = 0.95,
  gaussianize = TRUE,
  ofac = 4,
  padfac = 2,
  tbw = 3,
  wgtrad = 1,
  sigma = 0.02,
  mtm_null = "power_law"
)
}
\arguments{
\item{time}{LiPD "variable list" or vector of year/age values}

\item{values}{LiPD "variable list" or vector of values}

\item{max.ens}{Maximum number of ensemble members to analyze}

\item{probs}{vector of probabilities for the siginificance levels. To avoid such computations, pass `probs = NA`.}

\item{ofac}{oversampling factor for lomb::lsp and dplR::redfit}

\item{padfac}{padding factor for astrochron::mtm___ functions}

\item{tbw}{time-bandwidth product for astrochron::mtm___ functions}

\item{wgtrad}{radius of the nuspectral::nuwaveletcoeff weight function (non-dimensional units)}

\item{sigma}{decay parameter of the nuspectral::nuwaveletcoeff wavelets}
}
\value{
a list of ensemble spectra results
\itemize{
\item freqs: vector of frequencies
\item power: vector of spectral powers
\item power.CL: matrix of confidence limits for spectral power
}
}
\description{
Calculate ensemble power spectra using 4 spectral methods: 
\enumerate{
\item Lomb-Scargle periodogram
\item REDFIT (Mudelsee et al, 2009) 
\item Multitaper Method (Thomson et al. 1982)
\item nuspectral (Mathias et al 2004)  (no confidence estimation)
}
For REDFIT and MTM, only the 95% confidence limit is computed. The estimation thereof is as described in Mudelsee et al 2009 for a single ensemble member.
For multiple ensemble members, the median of the ensemble of 95% CLs is taken.  
For Lomb-Scargle, the parameter `probs` determines the quantiles of the surrogate spectrum distribution extracted as confidence limits.
}
\details{
Calculate ensemble power spectra
}
\references{
Mudelsee, M., D. Scholz, R. Röthlisberger, D. Fleitmann, A. Mangini, and E. W. Wolff (2009), Climate spectrum estimation in the presence of timescale errors, Nonlinear Processes in Geophysics, 16(1), 43–56, doi:10.5194/npg-16-43-2009.

Mathias, A., F. Grond, R. Guardans, D. Seese, M. Canela, and H. Diebner (2004), Algorithms for spectral analysis of irregularly sampled time series, Journal of Statistical Software, Articles, 11(2), 1–27, doi:10.18637/jss.v011.i02.

Thomson, D. J. (1982), Spectrum estimation and harmonic analysis, Proc. IEEE, 70(9), 1055–1096.
}
\seealso{
Other spectra: 
\code{\link{ar1Surrogates}()},
\code{\link{createSyntheticTimeseries}()},
\code{\link{periodAnnotate}()},
\code{\link{plotSpectraEns}()},
\code{\link{plotSpectrum}()},
\code{\link{reverselog10_trans}()}
}
\concept{spectra}
