% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.regression.functions.R
\name{corEns}
\alias{corEns}
\title{Ensemble correlation}
\usage{
corEns(time1, values1, time2, values2, binvec = NA, binstep = NA,
  binfun = mean, max.ens = NA, percentiles = c(0.025, 0.25, 0.5, 0.75,
  0.975), minObs = 10)
}
\arguments{
\item{time1}{matrix of age/time ensembles, or single column}

\item{values1}{matrix of values ensembles, or single column}

\item{time2}{matrix of age/time ensembles, or single column}

\item{values2}{matrix of values ensembles, or single column}

\item{binvec}{vector of bin edges for binning step}

\item{binstep}{spacing of bins, used to build bin step}

\item{binfun}{function to use during binning (mean, sd, and sum all work)}

\item{max.ens}{maximum number of ensemble members to correlate}

\item{percentiles}{quantiles to calculate for regression parameters}

\item{minObs}{minimum number of points required to calculate regression}
}
\value{
list of ensemble output and percentile information
}
\description{
Primary function for calculating correlation ensembles
}
\author{
Nick McKay
}
