% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.regression.functions.R
\name{corMatrix}
\alias{corMatrix}
\title{Matrix correlation}
\usage{
corMatrix(
  ens.1,
  ens.2,
  max.ens = NA,
  isospectral = TRUE,
  isopersistent = FALSE,
  p.ens = 100,
  gaussianize = TRUE,
  cor.method = "pearson"
)
}
\arguments{
\item{ens.1}{matrix of age-uncertain columns to correlate and calculate p-values}

\item{ens.2}{matrix of age-uncertain columns to correlate and calculate p-values}

\item{max.ens}{optionally limit the number of ensembles calculated (default = NA)}

\item{isospectral}{estimate significance using the Ebisuzaki method (default = TRUE)}

\item{isopersistent}{estimate significance using the isopersistence method (default = FALSE)}

\item{p.ens}{number of ensemble members to use for isospectral and/or isopersistent methods (default = 100)}

\item{gaussianize}{Boolean flag indicating whether the values should be mapped to a standard Gaussian prior to analysis.}

\item{cor.method}{correlation method to pass to cor() "pearson" (default), "kendall", or "spearman". Note that because the standard Student's T-test for significance is inappropriate for Kendall's Tau correlations, the raw and effective-N significance estimates will be NA when using "kendall"}
}
\value{
out list of correlation coefficients (r) p-values (p) and autocorrelation corrected p-values (pAdj)
}
\description{
Calculates correlations and associated p-values for two ensemble matrices (or vectors)
}
\seealso{
Other correlation: 
\code{\link{ar1Surrogates}()},
\code{\link{ar1}()},
\code{\link{effectiveN}()},
\code{\link{plotCorEns}()},
\code{\link{pvalMonteCarlo}()},
\code{\link{pvalPearsonSerialCorrected}()}
}
\author{
Nick McKay

Julien Emile-Geay
}
\concept{correlation}
