% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.regression.functions.R
\name{corMatrix}
\alias{corMatrix}
\title{Matrix correlation}
\usage{
corMatrix(ens.1, ens.2)
}
\arguments{
\item{ens.1}{matrix of age-uncertain columns to correlate and calculate p-values}

\item{ens.2}{matrix of age-uncertain columns to correlate and calculate p-values}
}
\value{
out list of correlation coefficients (r) p-values (p) and autocorrelation corrected p-values (pAdj)
}
\description{
Calculates correlations and associated p-values for two ensemble matrices (or vectors)
}
\author{
Nick McKay

Julien Emile-Geay
}
