% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronBuilding.R
\name{createChronMeasInputDf}
\alias{createChronMeasInputDf}
\title{Title}
\usage{
createChronMeasInputDf(L, which.chron = NA, which.mt = NA,
  overwrite = TRUE, site.name = L$dataSetName, modelNum = NA,
  useMarine = NULL, labIDVar = "labID", age14CVar = "age14C",
  age14CuncertaintyVar = "age14CUnc", ageVar = "age",
  ageUncertaintyVar = "ageUnc", depthVar = "depth",
  reservoirAge14CVar = "reservoirAge",
  reservoirAge14CUncertaintyVar = "reservoirAge14C",
  rejectedAgesVar = "rejected")
}
\arguments{
\item{L}{}

\item{which.chron}{}

\item{which.mt}{}

\item{overwrite}{}

\item{site.name}{}

\item{modelNum}{}

\item{useMarine}{}

\item{labIDVar}{}

\item{age14CVar}{}

\item{age14CuncertaintyVar}{}

\item{ageVar}{}

\item{ageUncertaintyVar}{}

\item{depthVar}{}

\item{reservoirAge14CVar}{}

\item{reservoirAge14CUncertaintyVar}{}

\item{rejectedAgesVar}{}
}
\value{
a standardized dataframe of chron measurements
}
\description{
Title
}
