% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembleComposite.R
\name{createConcatenatedEnsembleMeasurementTable}
\alias{createConcatenatedEnsembleMeasurementTable}
\title{Create a measurementTable from concatenated ensembles}
\usage{
createConcatenatedEnsembleMeasurementTable(
  age.ens.1,
  paleo.data.1,
  age.ens.2,
  paleo.data.2,
  n.ens = 1000
)
}
\arguments{
\item{age.ens.1}{Age ensemble list for the first dataset}

\item{paleo.data.1}{Paleo data list for the first dataset (ensembles optional)}

\item{age.ens.2}{Age ensemble list for the second dataset}

\item{paleo.data.2}{Paleo data  list for the second dataset (ensembles optional)}

\item{n.ens}{How many ensemble members?}
}
\value{
A measurement table object
}
\description{
This "shuffles" together two *independent* ensemble datasets, making no assumptions about stratigraphic order, but assuming that each dataset is representing the same phenomenon.
}
\examples{
\dontrun{
library(lipdR)
library(geoChronR)
library(magrittr)

L1 <- readLipd("~/Downloads/ANS1.Kjellman.2019.lpd")
L2 <- readLipd("~/Downloads/ANP3.Kjellman.2019.lpd")


#composite two independent records
L1 <- mapAgeEnsembleToPaleoData(L1,paleo.meas.table.num = 1,age.var = "ageEnsemble")

age.ens.1 <- selectData(L1,"ageEnsemble",meas.table.num = 1)
paleo.data.1 <- selectData(L1,"C28d2H",,meas.table.num = 1)

L2 <- mapAgeEnsembleToPaleoData(L2,paleo.meas.table.num = 1,age.var = "ageEnsemble")
age.ens.2 <- selectData(L2,"ageEnsemble",meas.table.num = 1)
paleo.data.2 <- selectData(L2,"C28d2H",,meas.table.num = 1)
cmt <- createConcatenatedEnsembleMeasurementTable(age.ens.1,paleo.data.1,age.ens.2,paleo.data.2)
plotTimeseriesEnsRibbons(X = cmt$ageEnsemble,Y = cmt$C28d2HComposite) \%>\%
plotTimeseriesEnsLines(X = cmt$ageEnsemble,Y = cmt$C28d2HComposite,n.ens.plot = 5,color = "red")
}


}
\seealso{
Other LiPD manipulation: 
\code{\link{createModel}()},
\code{\link{createMultiModelEnsemble}()},
\code{\link{createSummaryTableFromEnsembleTable}()},
\code{\link{createTSid}()},
\code{\link{estimateUncertaintyFromRange}()},
\code{\link{getVariableIndex}()},
\code{\link{mapAgeEnsembleToPaleoData}()},
\code{\link{mapPaleoEnsembleToPaleoData}()},
\code{\link{pullTsVariable}()},
\code{\link{selectData}()}
}
\concept{LiPD manipulation}
