% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxcal.R
\name{createOxcalModel}
\alias{createOxcalModel}
\title{Create the oxcal model script}
\usage{
createOxcalModel(
  cdf,
  depths.to.model = NA,
  depth.interval = 10,
  thin = 50,
  n.it = 10000,
  events.per.unit.length = 1,
  events.per.unit.length.uncertainty = NA,
  seq.type = "P_Sequence",
  top.boundary = "Boundary",
  bottom.boundary = "Boundary",
  outlier.prob = 0.05,
  cal.curve = "intcal13"
)
}
\arguments{
\item{cdf}{A data.frame of chron measurement data (output from createChronMeasInputDf())}

\item{depths.to.model}{A vector of depths to model (default = create a sequence using depth.interval)}

\item{depth.interval}{Depth interval to use to create depths.to.model (default = 10)}

\item{thin}{Thinning parameter for OxCal (default = 50)}

\item{n.it}{Number of iteration in model (default = 10000)}

\item{events.per.unit.length}{How many events per unit length for the Poisson model? This controls the flexibility on the model, the lower the number the more flexibile.}

\item{events.per.unit.length.uncertainty}{Uncertainty on events.per.unit.length in orders of magnitude. This makes for better models but *greatly* increases the run time. If NA or < 0 it runs without uncertainty (default = NA)}

\item{seq.type}{What type of Oxcal sequence to use? (default = "P_Sequence")}

\item{top.boundary}{What type of Oxcal boundary to use for the top (default = "Boundary")}

\item{bottom.boundary}{What type of Oxcal boundary to use for the bottom (default = "Boundary")}

\item{outlier.prob}{What's the probability of a date being an outlier if not otherwise specified (default= 0.05)}

\item{cal.curve}{What calibration curve to use? (default = "intcal13")}
}
\value{
a list of an oxcal model
}
\description{
Create the oxcal model script
}
\seealso{
Other oxcal: 
\code{\link{loadOxcalOutput}()},
\code{\link{oxCalDateExpression}()},
\code{\link{runOxcal}()}
}
\concept{oxcal}
