% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipd.manipulation.R
\name{estimateUncertaintyFromRange}
\alias{estimateUncertaintyFromRange}
\title{Estimate uncertainty estimates from high/low range}
\usage{
estimateUncertaintyFromRange(L, range1 = NA, range2 = NA,
  chronPaleo = "chronData", sigmaRange = 2, which.mt = 1,
  which.data = 1)
}
\arguments{
\item{L}{A lipd dataset}

\item{range1}{name of one of the range variables}

\item{range2}{name of the other range variable}

\item{sigmaRange}{what sigma range are the measurement uncertainties (default is 2)}

\item{which.mt}{which measurment table}

\item{which.data}{which chron or paleoData?}

\item{where}{chron or paleo}
}
\value{
MT: a LiPD measurementTable with a new unc.estimate variable
}
\description{
Estimate uncertainty (plus/minus values) from a range of values
}
\seealso{
Other LiPD manipulation: \code{\link{getVariableIndex}},
  \code{\link{mapAgeEnsembleToPaleoData}},
  \code{\link{selectData}}
}
