% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipd.manipulation.R
\name{estimateUncertaintyFromRange}
\alias{estimateUncertaintyFromRange}
\title{Estimate uncertainty estimates from high/low range}
\usage{
estimateUncertaintyFromRange(
  L,
  range.1 = NA,
  range.2 = NA,
  paleo.or.chron = "chronData",
  sigma.range = 2,
  meas.table.num = 1,
  paleo.or.chron.num = 1
)
}
\arguments{
\item{L}{a lipd object}

\item{range.1}{name of one of the range variables}

\item{range.2}{name of the other range variable}

\item{paleo.or.chron}{"paleoData" or "chronData"}

\item{sigma.range}{what sigma range are the measurement uncertainties (default is 2)}

\item{meas.table.num}{an integer that corresponds to paleo.num measurementTable has the variable you want?}

\item{paleo.or.chron.num}{an integer that corresponds to paleo.num or chron Data object (L$<paleo.or.chron>[[?]]) has the variable you want?}
}
\value{
MT: a LiPD measurementTable with a new unc.estimate variable
}
\description{
Estimate uncertainty (plus/minus values) from a range of values
}
\seealso{
Other LiPD manipulation: 
\code{\link{getVariableIndex}()},
\code{\link{mapAgeEnsembleToPaleoData}()},
\code{\link{pullTsVariable}()},
\code{\link{selectData}()}
}
\concept{LiPD manipulation}
