% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{gaussianize}
\alias{gaussianize}
\title{Gaussianize}
\usage{
gaussianize(X, jitter = FALSE)
}
\arguments{
\item{X}{data matrix (2D array)}

\item{jitter}{boolean variable ; if TRUE, add jitter to data to prevent ties}
}
\value{
gaussianized data matrix, Xn
}
\description{
Transforms each column of data matrix X to normality using the inverse Rosenblatt transform. Tolerant to missing values (NA entries).
}
\references{
Emile-Geay, J., and M. Tingley (2016), Inferring climate variability from nonlinear proxies: application to palaeo-enso studies, Climate of the Past, 12 (1), 31–50, doi:10.5194/cp-12-31-2016.

Van Albada, S.J., Robinson P.A. (2006), Transformation of arbitrary distributions to the normal distribution with application to EEG test-retest reliability. J Neurosci Meth, doi:10.1016/j.jneumeth.2006.11.004
}
\author{
Julien Emile-Geay

Nick McKay
}
