% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{generateEnsembleFromUncertainty}
\alias{generateEnsembleFromUncertainty}
\title{Create an ensemble from an uncertainty and AR1 estimate}
\usage{
generateEnsembleFromUncertainty(
  variable,
  n.ens,
  sd,
  ar = sqrt(0.5),
  arima.order = c(1, 0, 0)
)
}
\arguments{
\item{variable}{LiPD variable object or vector of data}

\item{n.ens}{Number of ensemble members to create (default = 1000)}

\item{sd}{Standard deviation of the uncertainty to simulate}

\item{ar}{Autocorrelation coefficient to use for modelling uncertainty, what fraction of the uncertainties are autocorrelated? (default = sqrt(0.5); or 50 percent autocorrelated uncertainty)}

\item{arima.order}{Order to use for ARIMA model used in modelling uncertainty (default = c(1,0,0))}
}
\value{
a matrix of ensemble values
}
\description{
Create an ensemble from an uncertainty and AR1 estimate
}
\seealso{
Other utility: 
\code{\link{askUser}()},
\code{\link{concatEnsembleTimeseries}()},
\code{\link{convertAD2BP}()},
\code{\link{convertBP2AD}()},
\code{\link{createChronMeasInputDf}()},
\code{\link{gaussianize}()},
\code{\link{getLastVarString}()},
\code{\link{getOs}()},
\code{\link{heuristicUnits}()},
\code{\link{loadRemote}()},
\code{\link{pullInstance}()},
\code{\link{simulateAutoCorrelatedUncertainty}()},
\code{\link{stringifyVariables}()},
\code{\link{surrogateDataFun}()}
}
\concept{utility}
