% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.geoChronR.R
\name{geoChronRPlotTheme}
\alias{geoChronRPlotTheme}
\title{Define a plot theme for GeoChronR}
\usage{
geoChronRPlotTheme(
  base_size = 12,
  base_family = "sans",
  style = c("default", "darkunica"),
  bgcolor = NULL
)
}
\description{
Use this to define a theme across geoChronR
}
\seealso{
Other plot help: 
\code{\link{AD2BP_trans}()},
\code{\link{BP2AD_trans}()},
\code{\link{axisLabel}()},
\code{\link{bin2d}()},
\code{\link{getLegend}()},
\code{\link{getPlotRanges}()},
\code{\link{kde2d}()},
\code{\link{meltDistributionTable}()},
\code{\link{modeSelektor}()},
\code{\link{periodAnnotate}()},
\code{\link{quantile2d}()},
\code{\link{reverselog10_trans}()}
}
\concept{plot help}
