% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacon.lipd.R
\name{getBaconDir}
\alias{getBaconDir}
\title{Get the name of bacon core directory}
\usage{
getBaconDir(bacon.dir = NA)
}
\arguments{
\item{bacon.dir}{if you already have it, it just returns this, (default=NA)}
}
\value{
bacon.dir the bacon core directory
}
\description{
This is a mostly internal function that returns the core directory. Interactive if given a NA
}
\author{
Nick McKay
}
