% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronBuilding.R
\name{getLastVarString}
\alias{getLastVarString}
\title{Get the last set of parameters you used in createChronMeasInputDf()}
\usage{
getLastVarString()
}
\value{

}
\description{
Get the last set of parameters you used in createChronMeasInputDf()
}
