% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.geoChronR.R
\name{getLegend}
\alias{getLegend}
\title{get a ggplot legend object}
\usage{
getLegend(a.gplot)
}
\arguments{
\item{a.gplot}{ggplot object}
}
\value{
a legend grob
}
\description{
get a ggplot legend object
}
\seealso{
Other plot help: 
\code{\link{AD2BP_trans}()},
\code{\link{BP2AD_trans}()},
\code{\link{axisLabel}()},
\code{\link{bin2d}()},
\code{\link{geoChronRPlotTheme}()},
\code{\link{getPlotRanges}()},
\code{\link{kde2d}()},
\code{\link{meltDistributionTable}()},
\code{\link{modeSelektor}()},
\code{\link{periodAnnotate}()},
\code{\link{quantile2d}()},
\code{\link{reverselog10_trans}()}
}
\author{
Nick McKay
}
\concept{plot help}
