% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.geoChronR.R
\name{getLegend}
\alias{getLegend}
\title{get a ggplot legend object}
\usage{
getLegend(a.gplot)
}
\arguments{
\item{a.gplot}{ggplot object}
}
\value{
a legend grob
}
\seealso{
Other plot: \code{\link{AD2BP_trans}},
  \code{\link{BP2AD_trans}}, \code{\link{PeriodAnnotate}},
  \code{\link{axisLabel}},
  \code{\link{geoChronRPlotTheme}},
  \code{\link{getPlotRanges}},
  \code{\link{plotChronEnsDiff}},
  \code{\link{plotChronEns}}, \code{\link{plotChron}},
  \code{\link{plotCorEns}}, \code{\link{plotHistEns}},
  \code{\link{plotLine}},
  \code{\link{plotModelDistributions}},
  \code{\link{plotPcaEns}}, \code{\link{plotPvalsEnsFdr}},
  \code{\link{plotRegressEns}},
  \code{\link{plotScatterEns}},
  \code{\link{plotSpectraEns}}, \code{\link{plotSpectrum}},
  \code{\link{plotSummary}},
  \code{\link{plotTimeseriesEnsLines}},
  \code{\link{plotTimeseriesEnsRibbons}},
  \code{\link{plotTimeseriesStack}},
  \code{\link{plotTrendLinesEns}},
  \code{\link{reverselog10_trans}}

Other pca: \code{\link{ar1Surrogates}},
  \code{\link{createSyntheticTimeseries}},
  \code{\link{pcaEns}}, \code{\link{plotPcaEns}}
}
\author{
Nick McKay
}
