% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.geoChronR.R
\name{getLegend}
\alias{getLegend}
\title{get a ggplot legend object}
\usage{
getLegend(a.gplot)
}
\arguments{
\item{a.gplot}{ggplot object}
}
\value{
a legend grob
}
\seealso{
Other plot: \code{\link{axisLabel}},
  \code{\link{geoChronRPlotTheme}},
  \code{\link{plotChronEnsDiff}},
  \code{\link{plotChronEns}}, \code{\link{plotCorEns}},
  \code{\link{plotHistEns}}, \code{\link{plotLine}},
  \code{\link{plotModelDistributions}},
  \code{\link{plotPcaEns}}, \code{\link{plotPvalsEnsFdr}},
  \code{\link{plotRegressEns}},
  \code{\link{plotScatterEns}},
  \code{\link{plotSpectraEns}}, \code{\link{plotSummary}},
  \code{\link{plotTimeseriesEnsLines}},
  \code{\link{plotTimeseriesEnsRibbons}},
  \code{\link{plotTimeseriesStack}},
  \code{\link{plotTrendLinesEns}}

Other pca: \code{\link{ar1Surrogates}},
  \code{\link{createSyntheticTimeseries}},
  \code{\link{pcaEns}}, \code{\link{plotPcaEns}}
}
\author{
Nick McKay
}
