% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxcal.R
\name{getModelParametersFromOxcalText}
\alias{getModelParametersFromOxcalText}
\title{Pull model parameters from OxCal Code}
\usage{
getModelParametersFromOxcalText(modText)
}
\arguments{
\item{modText}{The model text}
}
\value{
The list of model parameters
}
\description{
Pull model parameters from OxCal Code
}
\seealso{
Other oxcal: 
\code{\link{createOxcalModel}()},
\code{\link{loadOxcalOutput}()},
\code{\link{oxCalDateExpression}()},
\code{\link{runOxcal}()}
}
\concept{oxcal}
