% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipd.manipulation.R
\name{getVariableIndex}
\alias{getVariableIndex}
\title{Get the index of variable list}
\usage{
getVariableIndex(
  table,
  var.name = NA,
  alt.names = var.name,
  ignore = NA,
  always.choose = FALSE,
  strict.search = FALSE
)
}
\arguments{
\item{table}{a LiPD measurement, ensemble or summary Table}

\item{var.name}{string name of the variable to extract}

\item{alt.names}{A vector of strings for alternative names to search for}

\item{ignore}{A vector of strings of variableNames to ignore}

\item{always.choose}{Force selection of the variable from a list}

\item{strict.search}{Use a strict.search to look for the ageEnsemble and depth variables. TRUE(default) or FALSE.}
}
\value{
An integer index
}
\description{
Gets the index for a LiPD "variable list"
}
\seealso{
Other LiPD manipulation: 
\code{\link{estimateUncertaintyFromRange}()},
\code{\link{mapAgeEnsembleToPaleoData}()},
\code{\link{pullTsVariable}()},
\code{\link{selectData}()}
}
\concept{LiPD manipulation}
