% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipd.manipulation.R
\name{getVariableIndex}
\alias{getVariableIndex}
\title{Get the index of variable list}
\usage{
getVariableIndex(
  table,
  varName = NA,
  altNames = varName,
  ignore = NA,
  always.choose = FALSE,
  strictSearch = FALSE
)
}
\arguments{
\item{table}{a LiPD measurement, ensemble or summary Table}

\item{varName}{string name of the variable to extract}

\item{altNames}{A vector of strings for alternative names to search for}

\item{ignore}{A vector of strings of variableNames to ignore}

\item{always.choose}{Force selection of the variable from a list}

\item{strictSearch}{Use a strictSearch to look for the ageEnsemble and depth variables. TRUE(default) or FALSE.}
}
\value{
An integer index
}
\description{
Gets the index for a LiPD "variable list"
}
\seealso{
Other LiPD manipulation: 
\code{\link{estimateUncertaintyFromRange}()},
\code{\link{mapAgeEnsembleToPaleoData}()},
\code{\link{selectData}()}
}
\concept{LiPD manipulation}
