% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{heuristicUnits}
\alias{heuristicUnits}
\title{Makes a guess about the units of a time vector based on the values}
\usage{
heuristicUnits(X, range.min = 25)
}
\arguments{
\item{X}{A LiPD variable list or a vector of years BP}

\item{range.min}{What's the smallest range that you'd believe is yr not ka? (default = 25)}
}
\value{
A *guessed* string unit label
}
\description{
Guesses year units based on some simple heuristics
}
\seealso{
Other utility: 
\code{\link{askUser}()},
\code{\link{concatEnsembleTimeseries}()},
\code{\link{convertAD2BP}()},
\code{\link{convertBP2AD}()},
\code{\link{createChronMeasInputDf}()},
\code{\link{gaussianize}()},
\code{\link{getLastVarString}()},
\code{\link{getOs}()},
\code{\link{loadRemote}()},
\code{\link{pullInstance}()},
\code{\link{stringifyVariables}()},
\code{\link{surrogateDataFun}()}
}
\concept{utility}
