% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.geoChronR.R
\name{kde2d}
\alias{kde2d}
\title{Two dimensional kernel density estimation}
\usage{
kde2d(x, y, n.bins = 100, x.bin = NA, y.bin = NA)
}
\arguments{
\item{x}{n by m matrix where n is the number of observations and m is >= 1}

\item{y}{n by j matrix where n is the number of observations and j is >= 1}

\item{n.bins}{number bins over which to calculate intervals. Used to calculate x.bin if not provided.}

\item{x.bin}{vector of bin edges over which to bin.}

\item{y.bin}{vector of bin edges over which to bin.}
}
\value{
A list with a matrix of density, x.bin and y.bin
}
\description{
Use a kernel density estimator to model the density of samples along a 2-dimensional grid
}
\seealso{
Other gridding: 
\code{\link{bin2d}()},
\code{\link{quantile2d}()}

Other plot help: 
\code{\link{AD2BP_trans}()},
\code{\link{BP2AD_trans}()},
\code{\link{axisLabel}()},
\code{\link{bin2d}()},
\code{\link{geoChronRPlotTheme}()},
\code{\link{getLegend}()},
\code{\link{getPlotRanges}()},
\code{\link{meltDistributionTable}()},
\code{\link{modeSelektor}()},
\code{\link{periodAnnotate}()},
\code{\link{quantile2d}()},
\code{\link{reverselog10_trans}()}
}
\concept{gridding}
\concept{plot help}
