% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacon.lipd.R
\name{loadBaconOutput}
\alias{loadBaconOutput}
\title{Load the ensemble, summary and distribution data from a Bacon simulation}
\usage{
loadBaconOutput(
  L,
  site.name = L$dataSetName,
  K = NA,
  chron.num = NA,
  bacon.dir = NA,
  model.num = NA,
  make.new = NA,
  max.ens = 1000
)
}
\arguments{
\item{L}{a single LiPD object}

\item{site.name}{the name used for the bacon model (and directories)}

\item{chron.num}{the number of the chronData object that you'll be working in}

\item{bacon.dir}{the directory where Bacon is installed on this computer.}

\item{model.num}{chron.numModel do you want to use?}

\item{make.new}{do you want to create a new model in chronData? (TRUE, FALSE, NA). NA will try be smart, or ask you for advice.}
}
\value{
L the input LiPD file with methods and data added to the chronModel.
}
\description{
Loads the ensemble, summary and distribution data from a Bacon simulation and stores them in the LiPD structure. Will be run in interactive mode if necessary parameters aren't specified. Most users will want to use runBacon for their bacon needs.
}
\examples{
loadBaconOutput(L)
#Run in interactive mode
}
\seealso{
Other Bacon: 
\code{\link{runBacon}()},
\code{\link{sampleBaconAges}()},
\code{\link{writeBacon}()}
}
\author{
Nick McKay
}
\concept{Bacon}
