% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxcal.R
\name{loadOxcalOutput}
\alias{loadOxcalOutput}
\title{Load oxcal output}
\usage{
loadOxcalOutput(L, oxcalResultFilePath, modelParameters, depthUnits = "cm",
  which.chron = NA, modelNum = NA, makeNew = NA, maxEns = 1000)
}
\arguments{
\item{L}{}

\item{oxcalResultFilePath}{}

\item{modelParameters}{}

\item{depthUnits}{}

\item{which.chron}{}

\item{modelNum}{}

\item{makeNew}{}

\item{maxEns}{}
}
\description{
Load oxcal output
}
