% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{loadRemote}
\alias{loadRemote}
\title{Load a remote RData file}
\usage{
loadRemote(url, useSavedNames = FALSE)
}
\arguments{
\item{url}{the url of a remote data file}
}
\value{
the loadeds object
}
\description{
Load a remote RData file
}
