% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipd.manipulation.R
\name{mapAgeEnsembleToPaleoData}
\alias{mapAgeEnsembleToPaleoData}
\title{Map an ageEnsemble variable from a chron model to a paleoMeasurement Table}
\usage{
mapAgeEnsembleToPaleoData(
  L,
  age.var = "age",
  depth.var = "depth",
  paleo.num = NA,
  paleo.meas.table.num = NA,
  chron.num = NA,
  model.num = NA,
  ens.table.num = NA,
  max.ens = NA,
  strict.search = FALSE
)
}
\arguments{
\item{L}{A LiPD object - an R serialization of a single LiPD file. It's a list, and is typically created by `readLipd()`}

\item{age.var}{name of the age ensemble variable to search for}

\item{depth.var}{name of the depth variable to search for}

\item{paleo.num}{an integer that corresponds to paleo.numData object (L$paleoData[[?]]) has the measurementTable you want to modify}

\item{paleo.meas.table.num}{an integer that corresponds to paleo.num measurementTable you want to add the ensemble to?}

\item{chron.num}{an integer that corresponds to chron.numData object (L$crhonData[[?]]) has the model you want to get the ensemble from}

\item{model.num}{an integer that corresponds to chron.num model you want to get the ensemble from?}

\item{ens.table.num}{an integer that corresponds to ensembleTable you want to get the variable from?}

\item{max.ens}{Maximum number of ensemble members to map}

\item{strict.search}{Use a strict.search to look for the ageEnsemble and depth variables. TRUE(default) or FALSE.}
}
\value{
L a lipd object
}
\description{
Copies an ageEnsemble from chronData (model) to paleoData (measurementTable), by matching depth and interpolating (extrapolating) as necessary.
}
\seealso{
Other LiPD manipulation: 
\code{\link{createConcatenatedEnsembleMeasurementTable}()},
\code{\link{createModel}()},
\code{\link{createSummaryTableFromEnsembleTable}()},
\code{\link{createTSid}()},
\code{\link{estimateUncertaintyFromRange}()},
\code{\link{getVariableIndex}()},
\code{\link{pullTsVariable}()},
\code{\link{selectData}()}
}
\concept{LiPD manipulation}
