% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipd.manipulation.R
\name{mapAgeEnsembleToPaleoData}
\alias{mapAgeEnsembleToPaleoData}
\title{Create an ageEnsemble variable in a paleoMeasurement Table}
\usage{
mapAgeEnsembleToPaleoData(L, which.paleo = NA, which.pmt = NA,
  which.chron = NA, which.model = NA, max.ensemble.members = NA,
  strictSearch = FALSE)
}
\arguments{
\item{L}{a lipd object}

\item{which.paleo}{an integer that corresponds to which paleoData object (L$paleoData[[?]]) has the measurementTable you want to modify}
}
\value{
L a lipd object
}
\description{
Create an ageEnsemble variable in a paleoMeasurement Table
}
