% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.functions.R
\name{mapLipd}
\alias{mapLipd}
\title{Map a LiPD object}
\usage{
mapLipd(L, color = "red", size = 8, shape = 16, map.type = "google",
  extend.range = 10)
}
\arguments{
\item{L}{Single LiPD object}

\item{color}{Color of the location marker}

\item{size}{Size of the location marker}

\item{shape}{Shape of the location marker}

\item{map.type}{"google" or "line"}

\item{extend.range}{increase the span of the map by this much (lat/long degrees)}
}
\value{
ggmap object
}
\description{
Create a google or line map of the location of the LiPD object
}
\seealso{
Other mapping: \code{\link{assignColors}},
  \code{\link{baseMap}}, \code{\link{mapLipds}}
}
\author{
Nick McKay
}
