% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.functions.R
\name{mapLipd}
\alias{mapLipd}
\title{Map a list of LiPD objects}
\usage{
mapLipd(D, color = NA, size = 8, shape = 21, label.site = TRUE, ...)
}
\arguments{
\item{D}{A list of LiPD objects}

\item{color}{Color of the location marker}

\item{size}{Size of the location marker}

\item{shape}{Shape of the location marker}

\item{label.site}{Label the sites on the map (T/f)}

\item{...}{
  Arguments passed on to \code{\link[=baseMap]{baseMap}}
  \describe{
    \item{\code{lat}}{latitude(s) range to map}
    \item{\code{lon}}{longitude(s) range to map}
    \item{\code{map.type}}{"stamen" or "line"}
    \item{\code{f}}{buffer for the map range}
    \item{\code{restrict.map.range}}{TRUE or FALSE. Trim the size of the map to the points, for "line" map type}
    \item{\code{bound.circ}}{Draw a boundary circle around a polar projection. TRUE or FALSE(default).}
    \item{\code{global}}{Should the scope of the map be global? TRUE or FALSE(default).}
    \item{\code{projection}}{Map project. All options on: ?mapproject}
    \item{\code{extend.range}}{increase the span of the map by this much (lat/long degrees)}
  }}
}
\value{
ggmap object
}
\description{
Create a google or line map of the location of a list of LiPD objects
}
\seealso{
Other mapping: 
\code{\link{baseMap}()},
\code{\link{mapTs}()}
}
\author{
Nick McKay
}
\concept{mapping}
