% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.functions.R
\name{mapLipd}
\alias{mapLipd}
\title{Map a list of LiPD objects}
\usage{
mapLipd(D, color = NA, size = 8, shape = 21, labelSite = TRUE, ...)
}
\arguments{
\item{D}{A list of LiPD objects}

\item{color}{Color of the location marker}

\item{size}{Size of the location marker}

\item{shape}{Shape of the location marker}
}
\value{
ggmap object
}
\description{
Create a google or line map of the location of a list of LiPD objects
}
\seealso{
Other mapping: 
\code{\link{assignColors}()},
\code{\link{baseMap}()},
\code{\link{mapTs}()}
}
\author{
Nick McKay
}
\concept{mapping}
