% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.functions.R
\name{mapTs}
\alias{mapTs}
\title{Map a TS object}
\usage{
mapTs(TS, color = "archiveType", size = 6, shape = 21, ...)
}
\arguments{
\item{TS}{A list of LiPD objects}

\item{color}{variable by which to color}

\item{size}{Size of the location marker}

\item{shape}{Shape of the location marker}
}
\value{
ggmap object
}
\description{
Create a stamen or line map of the location of a list of LiPD objects
}
\seealso{
Other mapping: \code{\link{assignColors}},
  \code{\link{baseMap}}, \code{\link{mapLipd}}
}
\author{
Nick McKay
}
