% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.geoChronR.R
\name{meltDistributionTable}
\alias{meltDistributionTable}
\title{Melt distribution}
\usage{
meltDistributionTable(this.dist, dist.plot = 1:length(this.dist))
}
\arguments{
\item{this.dist}{LiPD "distributionTable" object}

\item{dist.plot}{vector of distribution tables to plot}
}
\value{
data.frame of melted distribution objects.
}
\description{
Takes a LiPD model distribution and melt it into a single data.frame
}
\author{
Nick McKay
}
