% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxcal.R
\name{oxCalDateExpression}
\alias{oxCalDateExpression}
\title{write oxcal expression for each date.}
\usage{
oxCalDateExpression(
  ...,
  defaultOutlierProb = 0.05,
  calCurve = "intcal13",
  unknownDeltaR = FALSE
)
}
\arguments{
\item{...}{}

\item{defaultOutlierProb}{}
}
\value{

}
\description{
write oxcal expression for each date.
}
