% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxcal.R
\name{oxCalDateExpression}
\alias{oxCalDateExpression}
\title{write oxcal expression for each date.}
\usage{
oxCalDateExpression(
  ...,
  default.outlier.prob = 0.05,
  cal.curve = "intcal13",
  unknown.delta.r = FALSE
)
}
\arguments{
\item{...}{}

\item{default.outlier.prob}{Outlier Probability}

\item{cal.curve}{Calibration curve}

\item{unknown.delta.r}{Use an unknown delta R (T/f)}
}
\value{

}
\description{
write oxcal expression for each date.
}
