% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.geoChronR.R
\name{pcaEns}
\alias{pcaEns}
\title{Perform principle components analysis (PCA) across an ensemble}
\usage{
pcaEns(bin.list, method = "ppca", weights = NA, PCAtype = "corr",
  nPCs = 4, nEns = 1000)
}
\arguments{
\item{bin.list}{A list of binned data, the output of binTs()}

\item{method}{methods}
}
\description{
Ensemble PCA, or Monte Carlo Empirical Orthogonal Functions as described in Anchukaitis and Tierney 2012.
}
