% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.geoChronR.R
\name{pcaEns}
\alias{pcaEns}
\title{Perform principle components analysis (PCA) across an ensemble}
\usage{
pcaEns(
  bin.list,
  method = "ppca",
  weights = NA,
  pca.type = "corr",
  gaussianize = TRUE,
  n.pcs = 8,
  n.ens = 1000,
  simulateTrendInNull = FALSE
)
}
\arguments{
\item{bin.list}{A list of binned data, the output of binTs()}

\item{method}{What method to use for PCA? pcaMethods::listPcaMethods() for options. "ppca" is default. Other options may not work in GeoChronR.}

\item{weights}{Vector of weights to apply to timeseries in the bin.list}

\item{pca.type}{Correlation ("corr" - default) or Covariance ("cov"), matrix}

\item{gaussianize}{Map input data to a standard Gaussian distribution? This is only relevant for correlation matrices, covariance matrices will not be gaussianized. (default = TRUE)}

\item{n.pcs}{number of PCs/EOFs to calculate}

\item{n.ens}{how many ensemble members to calculate}

\item{simulateTrendInNull}{Should the null include the trend?}
}
\description{
Ensemble PCA, or Monte Carlo Empirical Orthogonal Functions
}
\section{Long-form example}{

\href{http://nickmckay.github.io/GeoChronR/articles/PCA.html}{View a full-fledged example of how to use this function.}
}

\seealso{
Other pca: 
\code{\link{ar1Surrogates}()},
\code{\link{createSyntheticTimeseries}()},
\code{\link{plotPcaEns}()},
\code{\link{plotScreeEns}()}
}
\concept{pca}
