% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.geoChronR.R
\name{PeriodAnnotate}
\alias{PeriodAnnotate}
\title{Annotate plot of spectra with given periodicities}
\usage{
PeriodAnnotate(specPlot, periods, colour = "orange", log10scale = T,
  ylims = NULL, size = 4)
}
\arguments{
\item{specPlot}{ggplot handle to figure containing spectrum}

\item{periods}{the periods to highlight in the spectrum}

\item{colour}{the color of the text and lines}
}
\value{
ggplot object of spectrum plot
}
\description{
Annotate plot of spectra (ensemble or otherwise) with vertical lines at specific periodicities (assumes log10 scaling)
}
\seealso{
Other plot: \code{\link{AD2BP_trans}},
  \code{\link{BP2AD_trans}},
  \code{\link{FrequencyAnnotate}}, \code{\link{axisLabel}},
  \code{\link{geoChronRPlotTheme}},
  \code{\link{getLegend}}, \code{\link{getPlotRanges}},
  \code{\link{plotChronEnsDiff}},
  \code{\link{plotChronEns}}, \code{\link{plotChron}},
  \code{\link{plotCorEns}}, \code{\link{plotHistEns}},
  \code{\link{plotLine}},
  \code{\link{plotModelDistributions}},
  \code{\link{plotPcaEns}}, \code{\link{plotPvalsEnsFdr}},
  \code{\link{plotRegressEns}},
  \code{\link{plotScatterEns}},
  \code{\link{plotSpectraEns}}, \code{\link{plotSpectrum}},
  \code{\link{plotSummary}},
  \code{\link{plotTimeseriesEnsLines}},
  \code{\link{plotTimeseriesEnsRibbons}},
  \code{\link{plotTimeseriesStack}},
  \code{\link{plotTrendLinesEns}},
  \code{\link{reverselog10_trans}}

Other spectra: \code{\link{FrequencyAnnotate}},
  \code{\link{ar1Surrogates}},
  \code{\link{computeSpectraEns}},
  \code{\link{createSyntheticTimeseries}},
  \code{\link{plotSpectraEns}}, \code{\link{plotSpectrum}},
  \code{\link{reverselog10_trans}}
}
\author{
Julien Emile-Geay
}
