% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.geoChronR.R
\name{plotChronEns}
\alias{plotChronEns}
\title{Plot chron ensemble}
\usage{
plotChronEns(L, ageVar = "ageEnsemble", depthVar = "depth",
  chron.number = NA, model.num = NA, probs = c(0.025, 0.25, 0.5, 0.75,
  0.975), x.bin = NA, y.bin = NA, nbins = 100, bandColorLow = "white",
  bandColorHigh = "grey70", bandAlp = 1, lineColor = "Black",
  lineWidth = 1, add.to.plot = ggplot2::ggplot(), nEnsLines = 5,
  ensLineColor = "red", ensLineAlp = 0.7, distAlp = 0.3,
  distType = "violin", distColor = "purple", distThick = 0.1,
  distScale = 0.02, truncateDist = NA, addPaleoAgeDepth = FALSE,
  paleo.number = NA, meas.num = NA, paleoColor = "cyan")
}
\arguments{
\item{L}{A LiPD object}

\item{ageVar}{ariableName to use for age ensemble ("ageEnsemble" by default)}

\item{depthVar}{variableName to use for depth ("depth" by default)}

\item{chron.number}{which chronData object to use (NA by default, will ask if needed)}

\item{model.num}{which chronData model to use (NA by default, will ask if needed)}

\item{probs}{quantiles to calculate and plot}

\item{x.bin}{vector of bin edges over which to bin.}

\item{y.bin}{vector of bin edges over which to bin.}

\item{nbins}{number bins over which to calculate intervals. Used to calculate x.bin if not provided.}

\item{bandColorLow}{Band color of the outer most band.}

\item{bandColorHigh}{Band color of the inner most band.}

\item{lineColor}{Line color (following ggplot rules)}

\item{lineWidth}{Width of the line}

\item{add.to.plot}{A ggplot object to add this plot to. Default is ggplot() .}

\item{nEnsLines}{Number of ensemble members to plot}

\item{ensLineColor}{color of the ensemble lines}

\item{ensLineAlp}{transparency of the lines}

\item{distType}{"violin" (default), "up" for one-sided distributions pointed up, "down" for one-sided distributions pointed down}

\item{distColor}{distribution color (following ggplot rules)}

\item{distThick}{thickness of the line around the distribution}

\item{distScale}{controls the vertical span of the probability distribution. Approximately the vertical fraction of the plot that the distribution will cover.}

\item{truncateDist}{truncate probability density values below this number. NA (default) means no truncation}

\item{addPaleoAgeDepth}{add a line that shows the paleoData age depth.}

\item{paleo.number}{which paleo number for the paleoData age-depth}

\item{meas.num}{which measurement Table for the paleoData age-depth}

\item{paleoColor}{line color of the paleoData age-depth (following ggplot rules)}

\item{bandAlpha}{Transparency of the band plot}
}
\value{
A ggplot object
}
\description{
Plot creates an age model plot with all the bells and whistles, including a spread of ensemble members, probability distributions, and a few example ensemble members.
}
\seealso{
Other plot: \code{\link{axisLabel}},
  \code{\link{geoChronRPlotTheme}},
  \code{\link{getLegend}}, \code{\link{getPlotRanges}},
  \code{\link{plotChronEnsDiff}}, \code{\link{plotChron}},
  \code{\link{plotCorEns}}, \code{\link{plotHistEns}},
  \code{\link{plotLine}},
  \code{\link{plotModelDistributions}},
  \code{\link{plotPcaEns}}, \code{\link{plotPvalsEnsFdr}},
  \code{\link{plotRegressEns}},
  \code{\link{plotScatterEns}},
  \code{\link{plotSpectraEns}}, \code{\link{plotSummary}},
  \code{\link{plotTimeseriesEnsLines}},
  \code{\link{plotTimeseriesEnsRibbons}},
  \code{\link{plotTimeseriesStack}},
  \code{\link{plotTrendLinesEns}}

Other chron: \code{\link{plotChronEnsDiff}},
  \code{\link{plotChron}},
  \code{\link{plotModelDistributions}}
}
\author{
Nick McKay
}
