% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.geoChronR.R
\name{plotChronEnsDiff}
\alias{plotChronEnsDiff}
\title{Compare chron ensemble with paleoData age-model}
\usage{
plotChronEnsDiff(
  L,
  ageEnsVar = "ageEnsemble",
  ageVar = "age",
  depthVar = "depth",
  which.paleo = NA,
  which.pmt = NA,
  which.chron = NA,
  which.model = NA,
  which.ens = NA,
  max.ensemble.members = NA,
  strictSearch = FALSE,
  probs = c(0.025, 0.25, 0.5, 0.75, 0.975),
  x.bin = NA,
  y.bin = NA,
  nbins = 100,
  bandcolor.low = "white",
  bandcolor.high = "grey70",
  bandAlp = 1,
  color.line = "Black",
  lineWidth = 1,
  add.to.plot = ggplot2::ggplot(),
  nEnsLines = 5,
  enscolor.line = "red",
  ensLineAlp = 0.7
)
}
\arguments{
\item{L}{A LiPD object}

\item{ageEnsVar}{name of the age ensemble variable in the chronData to search for}

\item{ageVar}{name of the age variable in the paleoData to search for}

\item{which.paleo}{an integer that corresponds to which paleoData object (L$paleoData[[?]]) has the measurementTable you want to modify}

\item{which.pmt}{an integer that corresponds to which paleo measurementTable you want to add the ensemble to?}

\item{which.chron}{an integer that corresponds to which chronData object (L$crhonData[[?]]) has the model you want to get the ensemble from}

\item{which.model}{an integer that corresponds to which chron model you want to get the ensemble from?}

\item{which.ens}{an integer that corresponds to which chron model ensembleTable you want to get the ensemble from?}

\item{max.ensemble.members}{Maximum number of ensemble members to map}

\item{strictSearch}{Use a strictSearch to look for the ageEnsemble and depth variables. TRUE(default) or FALSE.}

\item{probs}{quantiles to calculate and plot}

\item{x.bin}{vector of bin edges over which to bin.}

\item{y.bin}{vector of bin edges over which to bin.}

\item{nbins}{number bins over which to calculate intervals. Used to calculate x.bin if not provided.}

\item{bandcolor.low}{Band color of the outer most band.}

\item{bandcolor.high}{Band color of the inner most band.}

\item{color.line}{Line color (following ggplot rules)}

\item{lineWidth}{Width of the line}

\item{add.to.plot}{A ggplot object to add this plot to. Default is ggplot() .}

\item{nEnsLines}{Number of ensemble members to plot}

\item{enscolor.line}{color of the ensemble lines}

\item{ensLineAlp}{transparency of the lines}

\item{depth.var}{name of the depth variable to search for}

\item{bandAlpha}{Transparency of the band plot}
}
\value{
A ggplot object
}
\description{
Plots the difference of an chron ensembleTable with the paleoData age
}
\seealso{
Other plot: 
\code{\link{AD2BP_trans}()},
\code{\link{BP2AD_trans}()},
\code{\link{PeriodAnnotate}()},
\code{\link{axisLabel}()},
\code{\link{geoChronRPlotTheme}()},
\code{\link{getLegend}()},
\code{\link{getPlotRanges}()},
\code{\link{plotChronEns}()},
\code{\link{plotChron}()},
\code{\link{plotCorEns}()},
\code{\link{plotHistEns}()},
\code{\link{plotLine}()},
\code{\link{plotModelDistributions}()},
\code{\link{plotPcaEns}()},
\code{\link{plotPvalsEnsFdr}()},
\code{\link{plotRegressEns}()},
\code{\link{plotScatterEns}()},
\code{\link{plotSpectraEns}()},
\code{\link{plotSpectrum}()},
\code{\link{plotSummary}()},
\code{\link{plotTimeseriesEnsLines}()},
\code{\link{plotTimeseriesEnsRibbons}()},
\code{\link{plotTimeseriesStack}()},
\code{\link{plotTrendLinesEns}()},
\code{\link{reverselog10_trans}()}

Other chron: 
\code{\link{plotChronEns}()},
\code{\link{plotChron}()},
\code{\link{plotModelDistributions}()}
}
\author{
Nick McKay
}
\concept{chron}
\concept{plot}
