% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.geoChronR.R
\name{plotCorEns}
\alias{plotCorEns}
\title{Plot the results of an ensemble correlation}
\usage{
plotCorEns(
  corEns,
  bins = 40,
  line.labels = rownames(cor.stats),
  add.to.plot = ggplot(),
  legend.position = c(0.2, 0.8),
  significance.option = "autocor"
)
}
\arguments{
\item{bins}{Number of bins in the histogram}

\item{line.labels}{Labels for the quantiles lines}

\item{add.to.plot}{A ggplot object to add these lines to. Default is ggplot()}

\item{legend.position}{Where to put the map legend?}

\item{significance.option}{Choose how handle significance. Options are:
\itemize{
\item "autocor" (default) for serial-autocorrelation corrected p-values
\item "raw" for uncorrected p-values
\item "FDR" for autocorrelation and False-discovery-rate corrected p-values
}}

\item{cor.df}{A data.frame correlation r and p-values. Output from corEns()}

\item{cor.stats}{A data.frame of correlation quantiles. Output from corEns()}
}
\value{
A ggplot object
}
\description{
Plots the output of an ensemble correlation analysis.
}
\seealso{
Other plot: 
\code{\link{AD2BP_trans}()},
\code{\link{BP2AD_trans}()},
\code{\link{axisLabel}()},
\code{\link{geoChronRPlotTheme}()},
\code{\link{getLegend}()},
\code{\link{getPlotRanges}()},
\code{\link{periodAnnotate}()},
\code{\link{plotChronEnsDiff}()},
\code{\link{plotChronEns}()},
\code{\link{plotChron}()},
\code{\link{plotHistEns}()},
\code{\link{plotLine}()},
\code{\link{plotModelDistributions}()},
\code{\link{plotPcaEns}()},
\code{\link{plotPvalsEnsFdr}()},
\code{\link{plotRegressEns}()},
\code{\link{plotScatterEns}()},
\code{\link{plotSpectraEns}()},
\code{\link{plotSpectrum}()},
\code{\link{plotSummary}()},
\code{\link{plotTimeseriesEnsLines}()},
\code{\link{plotTimeseriesEnsRibbons}()},
\code{\link{plotTimeseriesStack}()},
\code{\link{plotTrendLinesEns}()},
\code{\link{reverselog10_trans}()}
}
\author{
Julien Emile-Geay

Nick McKay
}
\concept{plot}
