% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.geoChronR.R
\name{plotHistEns}
\alias{plotHistEns}
\title{Plot an ensemble dataset as a histogram}
\usage{
plotHistEns(
  ens.data,
  quantiles = c(0.025, 0.25, 0.5, 0.75, 0.975),
  bins = 50,
  add.to.plot = ggplot(),
  alp = 1,
  fill = "grey50",
  font.size = 10,
  add.labels = TRUE,
  line.labels = NA,
  label.vert.position = 0.1
)
}
\arguments{
\item{ens.data}{A data.frame of values to plot as a histogram}

\item{quantiles}{a vecctor of quantiles to add the histogram}

\item{bins}{Number of bins in the histogram}

\item{add.to.plot}{A ggplot object to add these lines to. Default is ggplot()}

\item{alp}{transparency (between 0 and 1)}

\item{fill}{fill color of the histogram, following ggplot rules}

\item{font.size}{font size for the labels}

\item{add.labels}{Label the quantiles?}

\item{line.labels}{Labels for the quantiles lines}

\item{label.vert.position}{Vertical position of the quantiles (from 0 to 1)}
}
\value{
A ggplot object
}
\description{
Plots ensemble data as a histogram
}
\seealso{
Other plot: 
\code{\link{plotChron}()},
\code{\link{plotChronEns}()},
\code{\link{plotChronEnsDiff}()},
\code{\link{plotCorEns}()},
\code{\link{plotLine}()},
\code{\link{plotModelDistributions}()},
\code{\link{plotPcaEns}()},
\code{\link{plotPvalsEnsFdr}()},
\code{\link{plotRegressEns}()},
\code{\link{plotScatterEns}()},
\code{\link{plotScreeEns}()},
\code{\link{plotSpectraEns}()},
\code{\link{plotSpectrum}()},
\code{\link{plotSummary}()},
\code{\link{plotSummaryTs}()},
\code{\link{plotTimeseriesEnsLines}()},
\code{\link{plotTimeseriesEnsRibbons}()},
\code{\link{plotTimeseriesStack}()},
\code{\link{plotTrendLinesEns}()}
}
\author{
Nick McKay
}
\concept{plot}
