% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.geoChronR.R
\name{plotHistEns}
\alias{plotHistEns}
\title{Plot an ensemble dataset as a histogram}
\usage{
plotHistEns(ensData, quantiles = c(0.025, 0.25, 0.5, 0.75, 0.975),
  bins = 50, lineLabels = rownames(ensStats), add.to.plot = ggplot(),
  alp = 1, fill = "grey50")
}
\arguments{
\item{ensData}{A data.frame of values to plot as a histogram}

\item{bins}{Number of bins in the histogram}

\item{lineLabels}{Labels for the quantiles lines}

\item{add.to.plot}{A ggplot object to add these lines to. Default is ggplot()}

\item{fill}{fill color of the histogram, following ggplot rules}

\item{probs}{quantiles to calculate and plot}
}
\value{
A ggplot object
}
\description{
Plots ensemble data as a histogram
}
\seealso{
Other plot: \code{\link{axisLabel}},
  \code{\link{geoChronRPlotTheme}},
  \code{\link{getLegend}}, \code{\link{plotChronEnsDiff}},
  \code{\link{plotChronEns}}, \code{\link{plotChron}},
  \code{\link{plotCorEns}}, \code{\link{plotLine}},
  \code{\link{plotModelDistributions}},
  \code{\link{plotPcaEns}}, \code{\link{plotPvalsEnsFdr}},
  \code{\link{plotRegressEns}},
  \code{\link{plotScatterEns}},
  \code{\link{plotSpectraEns}}, \code{\link{plotSummary}},
  \code{\link{plotTimeseriesEnsLines}},
  \code{\link{plotTimeseriesEnsRibbons}},
  \code{\link{plotTimeseriesStack}},
  \code{\link{plotTrendLinesEns}}
}
\author{
Nick McKay
}
