% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.geoChronR.R
\name{plotModelDistributions}
\alias{plotModelDistributions}
\title{Plot probability distributions}
\usage{
plotModelDistributions(
  L,
  dist.var = "age",
  y.var = "depth",
  mode = "chron",
  paleo.or.chron.num = 1,
  model.num = 1,
  add.to.plot = ggplot(),
  alp = 0.5,
  color = "purple",
  scale.frac = 0.02,
  dist.plot = NA,
  dist.type = "violin",
  thick = 0.1,
  truncate.dist = NA
)
}
\arguments{
\item{L}{A LiPD object - an R serialization of a single LiPD file. It's a list, and is typically created by `readLipd()`}

\item{dist.var}{Name of the distribution variable, will be plotted along the x-axis. Use coord_flip() after running the function if you want vertical distributions. "age" by default.}

\item{y.var}{Name of the y-axis variable. "depth" by default.}

\item{mode}{chron or paleo}

\item{paleo.or.chron.num}{number of the chron or paleo Data object}

\item{model.num}{number of the model object}

\item{add.to.plot}{A ggplot object to add this plot to. Default is ggplot()}

\item{alp}{transparency, from 0 to 1}

\item{color}{distribution color (following ggplot rules)}

\item{scale.frac}{controls the vertical span of the probability distribution. Approximately the vertical fraction of the plot that the distribution will cover.}

\item{dist.plot}{vector of distribution tables to plot}

\item{dist.type}{"violin" (default), "up" for one-sided distributions pointed up, "down" for one-sided distributions pointed down}

\item{thick}{thickness of the line around the distribution}

\item{truncate.dist}{truncate probability density values below this number. NA (default) means no truncation}
}
\value{
A ggplot object
}
\description{
Plot or add probability distributions from a paleo or chron model to a plot.
}
\seealso{
Other plot: 
\code{\link{plotChronEnsDiff}()},
\code{\link{plotChronEns}()},
\code{\link{plotChron}()},
\code{\link{plotCorEns}()},
\code{\link{plotHistEns}()},
\code{\link{plotLine}()},
\code{\link{plotPcaEns}()},
\code{\link{plotPvalsEnsFdr}()},
\code{\link{plotRegressEns}()},
\code{\link{plotScatterEns}()},
\code{\link{plotScreeEns}()},
\code{\link{plotSpectraEns}()},
\code{\link{plotSpectrum}()},
\code{\link{plotSummaryTs}()},
\code{\link{plotSummary}()},
\code{\link{plotTimeseriesEnsLines}()},
\code{\link{plotTimeseriesEnsRibbons}()},
\code{\link{plotTimeseriesStack}()},
\code{\link{plotTrendLinesEns}()}

Other chron: 
\code{\link{plotChronEnsDiff}()},
\code{\link{plotChronEns}()},
\code{\link{plotChron}()}
}
\author{
Nick McKay
}
\concept{chron}
\concept{plot}
