% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.geoChronR.R
\name{plotScatterEns}
\alias{plotScatterEns}
\title{Plot an ensemble of data as a scatterplot}
\usage{
plotScatterEns(X, Y, alp = 0.2, n.ens.plot = 1000, add.to.plot = ggplot())
}
\arguments{
\item{X}{A LiPD variable list to plot, including values, units, names, and more}

\item{Y}{A LiPD variable list to plot, including values, units, names, and more}

\item{alp}{Marker transparency}

\item{n.ens.plot}{Whats the maximum number of points to plot?}

\item{add.to.plot}{A ggplot object to add this plot to. Default is ggplot() .}
}
\value{
A ggplot object
}
\description{
Plot an ensemble timeseries as a scatter plot. Useful in showing the general impact of uncertainty on a bivariate relationship.
}
\seealso{
Other plot: 
\code{\link{plotChronEnsDiff}()},
\code{\link{plotChronEns}()},
\code{\link{plotChron}()},
\code{\link{plotCorEns}()},
\code{\link{plotHistEns}()},
\code{\link{plotLine}()},
\code{\link{plotModelDistributions}()},
\code{\link{plotPcaEns}()},
\code{\link{plotPvalsEnsFdr}()},
\code{\link{plotRegressEns}()},
\code{\link{plotScreeEns}()},
\code{\link{plotSpectraEns}()},
\code{\link{plotSpectrum}()},
\code{\link{plotSummaryTs}()},
\code{\link{plotSummary}()},
\code{\link{plotTimeseriesEnsLines}()},
\code{\link{plotTimeseriesEnsRibbons}()},
\code{\link{plotTimeseriesStack}()},
\code{\link{plotTrendLinesEns}()}

Other regress: 
\code{\link{plotRegressEns}()},
\code{\link{plotTrendLinesEns}()},
\code{\link{regressEns}()},
\code{\link{regress}()}
}
\author{
Nick McKay
}
\concept{plot}
\concept{regress}
