% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.geoChronR.R
\name{plotScatterEns}
\alias{plotScatterEns}
\title{Plot an ensemble of data as a scatterplot}
\usage{
plotScatterEns(X, Y, alp = 0.2, maxPlotN = 1000, add.to.plot = ggplot())
}
\arguments{
\item{X}{A LiPD variable list to plot, including values, units, names, and more}

\item{Y}{A LiPD variable list to plot, including values, units, names, and more}

\item{alp}{Line transparency}

\item{maxPlotN}{Whats the maximum number of lines to plot?}

\item{add.to.plot}{A ggplot object to add this plot to. Default is ggplot() .}
}
\value{
A ggplot object
}
\description{
Plot an ensemble timeseries as a scatter plot. Useful in showing the general impact of uncertainty on a bivariate relationship.
}
\seealso{
Other plot: \code{\link{geoChronRPlotTheme}},
  \code{\link{plotCorrEns}}, \code{\link{plotHistEns}},
  \code{\link{plotLine}}, \code{\link{plotPvalsEnsFdr}},
  \code{\link{plotSpectraEns}}, \code{\link{plotSummary}},
  \code{\link{plotTimeseriesEnsLines}},
  \code{\link{plotTimeseriesEnsRibbons}},
  \code{\link{plotTrendLinesEns}}

Other regress: \code{\link{plotTrendLinesEns}}
}
\author{
Nick McKay
}
