% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.geoChronR.R
\name{plotSpectraEns}
\alias{plotSpectraEns}
\title{Plot ensemble spectra output}
\usage{
plotSpectraEns(
  spec.ens,
  cl.df = NULL,
  x.lims = NULL,
  x.ticks = c(10, 20, 50, 100, 200, 500, 1000),
  y.lims = NULL,
  color.low = "white",
  color.high = "grey70",
  color.line = "Black",
  color.cl = "red",
  alp = 0.5
)
}
\arguments{
\item{spec.ens}{list or dataframe containing frequency (freq) and power (pwr); typically output of computeSpectraEns}

\item{cl.df}{list or dataframe containing confidence limits (90, 95 and 99\%) as well as frequency (freq)}

\item{x.lims}{2-element vector defining the range of periods (x-axis)}

\item{x.ticks}{n-element vector of the periods labeled}

\item{y.lims}{2-element vector defining the range of spectral power (y-axis)}

\item{color.low}{Color of the outermost band; the extreme quantiles of the distribution}

\item{color.high}{Color of the innermost band; the central quantiles of the distribution}

\item{color.line}{Line color (following ggplot rules)}

\item{color.cl}{color of the lines representing the confidence limits (90, 95, 99\%)}

\item{alp}{Line transparency}
}
\value{
a ggplot object
}
\description{
Plot the output of `computeSpectraEns` as a ribbon plot of distributions, with specified confidence levels
}
\details{
`plotSpectraEns` re-uses `plotTimeseriesRibbons` and therefore the same graphical conventions. Spectra are plotted on a log-log scale, with the x-axis labeled by periods instead of frequencies, for improved intelligibility.
}
\seealso{
Other plot: 
\code{\link{AD2BP_trans}()},
\code{\link{BP2AD_trans}()},
\code{\link{axisLabel}()},
\code{\link{geoChronRPlotTheme}()},
\code{\link{getLegend}()},
\code{\link{getPlotRanges}()},
\code{\link{periodAnnotate}()},
\code{\link{plotChronEnsDiff}()},
\code{\link{plotChronEns}()},
\code{\link{plotChron}()},
\code{\link{plotCorEns}()},
\code{\link{plotHistEns}()},
\code{\link{plotLine}()},
\code{\link{plotModelDistributions}()},
\code{\link{plotPcaEns}()},
\code{\link{plotPvalsEnsFdr}()},
\code{\link{plotRegressEns}()},
\code{\link{plotScatterEns}()},
\code{\link{plotSpectrum}()},
\code{\link{plotSummary}()},
\code{\link{plotTimeseriesEnsLines}()},
\code{\link{plotTimeseriesEnsRibbons}()},
\code{\link{plotTimeseriesStack}()},
\code{\link{plotTrendLinesEns}()},
\code{\link{reverselog10_trans}()}

Other spectra: 
\code{\link{ar1Surrogates}()},
\code{\link{computeSpectraEns}()},
\code{\link{createSyntheticTimeseries}()},
\code{\link{periodAnnotate}()},
\code{\link{plotSpectrum}()},
\code{\link{reverselog10_trans}()}
}
\author{
Julien Emile-Geay
}
\concept{plot}
\concept{spectra}
