% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.geoChronR.R
\name{plotSpectrum}
\alias{plotSpectrum}
\title{Plot spectrum with confidence limits}
\usage{
plotSpectrum(
  spec.df,
  cl.df = NULL,
  x.lims = NULL,
  x.ticks = c(10, 20, 50, 100, 200, 500, 1000),
  y.lims = NULL,
  color.line = "black",
  color.cl = "red"
)
}
\arguments{
\item{spec.df}{list or dataframe containing frequency (freq) and power (pwr)}

\item{cl.df}{list or dataframe containing confidence limits (90, 95 and 99\%) as well as frequency (freq)}

\item{x.lims}{range of plotted periodicities}

\item{x.ticks}{ticks to mark on the period axis. if NULL, defaults to (10, 20, 50, 100, 200, 500, 1000)}

\item{y.lims}{2-vector for the y-axis. If NULL, computed from range(pwr)}

\item{color.cl}{color of the lines representing the confidence limits (90, 95, 99\%)}

\item{color.main}{color of the line representing the spectrum}
}
\value{
a ggplot object
}
\description{
Plot a a single spectrum, with confidence limits (no age ensemble). Useful for comparison with plotSpectraEns() in cases of no age uncertainty (e.g. GCM output)
}
\seealso{
Other plot: 
\code{\link{AD2BP_trans}()},
\code{\link{BP2AD_trans}()},
\code{\link{axisLabel}()},
\code{\link{geoChronRPlotTheme}()},
\code{\link{getLegend}()},
\code{\link{getPlotRanges}()},
\code{\link{periodAnnotate}()},
\code{\link{plotChronEnsDiff}()},
\code{\link{plotChronEns}()},
\code{\link{plotChron}()},
\code{\link{plotCorEns}()},
\code{\link{plotHistEns}()},
\code{\link{plotLine}()},
\code{\link{plotModelDistributions}()},
\code{\link{plotPcaEns}()},
\code{\link{plotPvalsEnsFdr}()},
\code{\link{plotRegressEns}()},
\code{\link{plotScatterEns}()},
\code{\link{plotSpectraEns}()},
\code{\link{plotSummary}()},
\code{\link{plotTimeseriesEnsLines}()},
\code{\link{plotTimeseriesEnsRibbons}()},
\code{\link{plotTimeseriesStack}()},
\code{\link{plotTrendLinesEns}()},
\code{\link{reverselog10_trans}()}

Other spectra: 
\code{\link{ar1Surrogates}()},
\code{\link{computeSpectraEns}()},
\code{\link{createSyntheticTimeseries}()},
\code{\link{periodAnnotate}()},
\code{\link{plotSpectraEns}()},
\code{\link{reverselog10_trans}()}
}
\author{
Julien Emile-Geay
}
\concept{plot}
\concept{spectra}
