% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.geoChronR.R
\name{plotSpectrum}
\alias{plotSpectrum}
\title{Plot single spectrum, with confidence limits}
\usage{
plotSpectrum(spec.df, cl.df = NULL, period_range = NULL,
  period_ticks = c(10, 20, 50, 100, 200, 500, 1000), ylims = NULL)
}
\arguments{
\item{spec.df}{dataframe containing frequency (freq) and power (pwr)}

\item{cl.df}{dataframe containing confidence limits (90, 95 and 99%) as well as frequency (freq)}

\item{period_range}{range of plotted periodicities}
}
\value{
ggplot object of spectrum plot
}
\description{
Plota a single spectrum, with confidence limits (no age ensemble)
}
\seealso{
Other plot: \code{\link{AD2BP_trans}},
  \code{\link{BP2AD_trans}},
  \code{\link{FrequencyAnnotate}},
  \code{\link{PeriodAnnotate}}, \code{\link{axisLabel}},
  \code{\link{geoChronRPlotTheme}},
  \code{\link{getLegend}}, \code{\link{getPlotRanges}},
  \code{\link{plotChronEnsDiff}},
  \code{\link{plotChronEns}}, \code{\link{plotChron}},
  \code{\link{plotCorEns}}, \code{\link{plotHistEns}},
  \code{\link{plotLine}},
  \code{\link{plotModelDistributions}},
  \code{\link{plotPcaEns}}, \code{\link{plotPvalsEnsFdr}},
  \code{\link{plotRegressEns}},
  \code{\link{plotScatterEns}},
  \code{\link{plotSpectraEns}}, \code{\link{plotSummary}},
  \code{\link{plotTimeseriesEnsLines}},
  \code{\link{plotTimeseriesEnsRibbons}},
  \code{\link{plotTimeseriesStack}},
  \code{\link{plotTrendLinesEns}},
  \code{\link{reverselog10_trans}}

Other spectra: \code{\link{FrequencyAnnotate}},
  \code{\link{PeriodAnnotate}},
  \code{\link{ar1Surrogates}},
  \code{\link{computeSpectraEns}},
  \code{\link{createSyntheticTimeseries}},
  \code{\link{plotSpectraEns}},
  \code{\link{reverselog10_trans}}
}
\author{
Julien Emile-Geay
}
