% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.geoChronR.R
\name{plotSummary}
\alias{plotSummary}
\title{Plot a summary figure}
\usage{
plotSummary(L)
}
\arguments{
\item{L}{A LiPD Object}
}
\value{
A gridArrange of ggplot grobs
}
\description{
shows a map, timeseries, and age model diagram, and basic simple metadata
}
\examples{
myPlot = summaryPlot(L)

}
\seealso{
Other plot: \code{\link{geoChronRPlotTheme}},
  \code{\link{plotCorrEns}}, \code{\link{plotHistEns}},
  \code{\link{plotLine}}, \code{\link{plotPvalsEnsFdr}},
  \code{\link{plotScatterEns}},
  \code{\link{plotSpectraEns}},
  \code{\link{plotTimeseriesEnsLines}},
  \code{\link{plotTimeseriesEnsRibbons}},
  \code{\link{plotTrendLinesEns}}
}
\author{
Nick McKay
}
