% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.geoChronR.R
\name{plotSummary}
\alias{plotSummary}
\title{Plot a summary figure}
\usage{
plotSummary(
  L,
  paleo.age.var = "age",
  paleo.data.var = NA,
  chron.number = NA,
  paleo.meas.num = NA,
  chron.meas.num = NA,
  chron.depth.var = "depth",
  chron.age.var = "age",
  dotSize = 5,
  summary.font.size = 10,
  text.width = 400/summary.font.size,
  legend.position = c(0.7, 0.3),
  ...
)
}
\arguments{
\item{L}{A LiPD object}

\item{paleo.age.var}{variableName to use for x axis of the paleo plot ("age" by default)}

\item{paleo.data.var}{variableName to use for the y axis of the paleo plot (NA by default, which lets you choose)}

\item{chron.number}{which chronData object to use (NA by default, will ask if needed)}

\item{paleo.meas.num}{which paleo measurement table to use (NA by default, will ask if needed)}

\item{chron.meas.num}{which chron measurement table to use (NA by default, will ask if needed)}

\item{chron.depth.var}{variableName to use for chron depth ("depth" by default)}

\item{chron.age.var}{variableName to use for chron age ("age" by default)}

\item{...}{arguments to pass on to plotChronEns()}

\item{dotsize}{what size dot for the chron plot? Only used if not plotting by plotChronEns() (default = 5)}
}
\value{
A gridArrange of ggplot grobs
}
\description{
shows a map, timeseries, and age model diagram, and basic simple metadata
}
\examples{
myPlot = summaryPlot(L)

}
\seealso{
Other plot: 
\code{\link{AD2BP_trans}()},
\code{\link{BP2AD_trans}()},
\code{\link{PeriodAnnotate}()},
\code{\link{axisLabel}()},
\code{\link{geoChronRPlotTheme}()},
\code{\link{getLegend}()},
\code{\link{getPlotRanges}()},
\code{\link{plotChronEnsDiff}()},
\code{\link{plotChronEns}()},
\code{\link{plotChron}()},
\code{\link{plotCorEns}()},
\code{\link{plotHistEns}()},
\code{\link{plotLine}()},
\code{\link{plotModelDistributions}()},
\code{\link{plotPcaEns}()},
\code{\link{plotPvalsEnsFdr}()},
\code{\link{plotRegressEns}()},
\code{\link{plotScatterEns}()},
\code{\link{plotSpectraEns}()},
\code{\link{plotSpectrum}()},
\code{\link{plotTimeseriesEnsLines}()},
\code{\link{plotTimeseriesEnsRibbons}()},
\code{\link{plotTimeseriesStack}()},
\code{\link{plotTrendLinesEns}()},
\code{\link{reverselog10_trans}()}
}
\author{
Nick McKay
}
\concept{plot}
