% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.geoChronR.R
\name{plotTimeseriesEnsLines}
\alias{plotTimeseriesEnsLines}
\title{Plot an ensemble timeseries as a set of lines}
\usage{
plotTimeseriesEnsLines(X, Y, alp = 0.2, color = "blue", maxPlotN = 1000,
  add.to.plot = ggplot())
}
\arguments{
\item{X}{A LiPD variable list to plot, including values, units, names, and more}

\item{Y}{A LiPD variable list to plot, including values, units, names, and more}

\item{alp}{Line transparency}

\item{color}{Line color (following ggplot rules)}

\item{maxPlotN}{Whats the maximum number of lines to plot?}

\item{add.to.plot}{A ggplot object to add these lines to. Default is ggplot() .}
}
\value{
A ggplot object
}
\description{
Plot an ensemble timeseries as a set of lines. Useful for displaying a handful of ensemble members to characterize individual paths.
}
\seealso{
Other plot: \code{\link{axisLabel}},
  \code{\link{geoChronRPlotTheme}},
  \code{\link{plotChron}}, \code{\link{plotCorrEns}},
  \code{\link{plotHistEns}}, \code{\link{plotLine}},
  \code{\link{plotModelDistributions}},
  \code{\link{plotPcaEns}}, \code{\link{plotPvalsEnsFdr}},
  \code{\link{plotRegressEns}},
  \code{\link{plotScatterEns}},
  \code{\link{plotSpectraEns}}, \code{\link{plotSummary}},
  \code{\link{plotTimeseriesEnsRibbons}},
  \code{\link{plotTrendLinesEns}}
}
\author{
Nick McKay
}
