% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.geoChronR.R
\name{plotTimeseriesEnsRibbons}
\alias{plotTimeseriesEnsRibbons}
\title{Plot an ensemble timeseries as ribbons of probabilities}
\usage{
plotTimeseriesEnsRibbons(X, Y, alp = 1, probs = c(0.025, 0.25, 0.5, 0.75,
  0.975), x.bin = NA, y.bin = NA, nbins = 200, colorLow = "white",
  colorHigh = "grey70", lineColor = "Black", lineWidth = 1,
  add.to.plot = ggplot())
}
\arguments{
\item{X}{A LiPD variable list to plot, including values, units, names, and more}

\item{Y}{A LiPD variable list to plot, including values, units, names, and more}

\item{alp}{Line transparency}

\item{probs}{a vector of probabilities to plot as ribbons. It will create bands as ribbons of quantiles moving inward. If there's an odd number, it plots the middle quantile as a line.}

\item{x.bin}{vector of bin edges over which to bin.}

\item{y.bin}{vector of bin edges over which to bin.}

\item{nbins}{number bins over which to calculate intervals. Used to calculate x.bin if not provided.}

\item{colorLow}{Band color of the outer most band.}

\item{colorHigh}{Band color of the inner most band.}

\item{lineColor}{Line color (following ggplot rules)}

\item{lineWidth}{Width of the line}

\item{add.to.plot}{A ggplot object to add this plot to. Default is ggplot() .}
}
\value{
A ggplot object
}
\description{
Plot an ensemble timeseries as a set of bands of probability. Useful for displaying the full range of probability across ensemble members.
}
\seealso{
Other plot: \code{\link{axisLabel}},
  \code{\link{geoChronRPlotTheme}},
  \code{\link{plotChron}}, \code{\link{plotCorrEns}},
  \code{\link{plotHistEns}}, \code{\link{plotLine}},
  \code{\link{plotModelDistributions}},
  \code{\link{plotPcaEns}}, \code{\link{plotPvalsEnsFdr}},
  \code{\link{plotRegressEns}},
  \code{\link{plotScatterEns}},
  \code{\link{plotSpectraEns}}, \code{\link{plotSummary}},
  \code{\link{plotTimeseriesEnsLines}},
  \code{\link{plotTrendLinesEns}}
}
\author{
Nick McKay
}
