% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.geoChronR.R
\name{plotTrendLinesEns}
\alias{plotTrendLinesEns}
\title{Plot an ensemble of trendlines}
\usage{
plotTrendLinesEns(mb.df, xrange, pXY = 1:nrow(mb.df), alp = 0.2,
  color = "red", add.to.plot = ggplot())
}
\arguments{
\item{mb.df}{A data.frame of slopes (column 1) and intercepts (column 2)}

\item{xrange}{range of x values (min and max)}

\item{pXY}{index of which observations to use}

\item{alp}{Line transparency}

\item{add.to.plot}{A ggplot object to add these lines to. Default is ggplot() .}
}
\value{
A ggplot object
}
\description{
Plot an ensemble of trendlines based on slope and intercept.
}
\seealso{
Other plot: \code{\link{AD2BP_trans}},
  \code{\link{BP2AD_trans}}, \code{\link{axisLabel}},
  \code{\link{geoChronRPlotTheme}},
  \code{\link{getLegend}}, \code{\link{getPlotRanges}},
  \code{\link{plotChronEnsDiff}},
  \code{\link{plotChronEns}}, \code{\link{plotChron}},
  \code{\link{plotCorEns}}, \code{\link{plotHistEns}},
  \code{\link{plotLine}},
  \code{\link{plotModelDistributions}},
  \code{\link{plotPcaEns}}, \code{\link{plotPvalsEnsFdr}},
  \code{\link{plotRegressEns}},
  \code{\link{plotScatterEns}},
  \code{\link{plotSpectraEns}}, \code{\link{plotSummary}},
  \code{\link{plotTimeseriesEnsLines}},
  \code{\link{plotTimeseriesEnsRibbons}},
  \code{\link{plotTimeseriesStack}}

Other regress: \code{\link{plotRegressEns}},
  \code{\link{plotScatterEns}}, \code{\link{regressEns}}
}
\author{
Nick McKay
}
