% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembleComposite.R
\name{pullInstance}
\alias{pullInstance}
\title{Pull one instance from an ensemble}
\usage{
pullInstance(X)
}
\arguments{
\item{X}{A LiPD column object}
}
\value{
a vector randomly drawn from the matrix
}
\description{
Pull one instance from an ensemble
}
\seealso{
Other utility: 
\code{\link{askUser}()},
\code{\link{concatEnsembleTimeseries}()},
\code{\link{convertAD2BP}()},
\code{\link{convertBP2AD}()},
\code{\link{createChronMeasInputDf}()},
\code{\link{gaussianize}()},
\code{\link{generateEnsembleFromUncertainty}()},
\code{\link{getLastVarString}()},
\code{\link{getOs}()},
\code{\link{heuristicUnits}()},
\code{\link{loadRemote}()},
\code{\link{simulateAutoCorrelatedUncertainty}()},
\code{\link{stringifyVariables}()},
\code{\link{surrogateDataFun}()}
}
\concept{utility}
