% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipd.manipulation.R
\name{pullTsVariable}
\alias{pullTsVariable}
\title{pull variable out of TS object}
\usage{
pullTsVariable(TS, variable)
}
\arguments{
\item{TS}{LiPD timeseries object See \url{http://nickmckay.github.io/LiPD-utilities/r/index.html#what-is-a-time-series}}

\item{variable}{the name of variable in a TS object}
}
\value{
a vector of the values, with NA representing instances without this variable.
}
\description{
pulls all instances of a single variable out of a TS
}
\seealso{
Other LiPD manipulation: 
\code{\link{createConcatenatedEnsembleMeasurementTable}()},
\code{\link{createModel}()},
\code{\link{createMultiModelEnsemble}()},
\code{\link{createSummaryTableFromEnsembleTable}()},
\code{\link{createTSid}()},
\code{\link{estimateUncertaintyFromRange}()},
\code{\link{getVariableIndex}()},
\code{\link{mapAgeEnsembleToPaleoData}()},
\code{\link{selectData}()}
}
\concept{LiPD manipulation}
