% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.regression.functions.R
\name{pvalMonteCarlo}
\alias{pvalMonteCarlo}
\title{Correlations and their significance according to AR(1) benchmarks}
\usage{
pvalMonteCarlo(X, Y, n.sim = 100, method = "isospectral")
}
\arguments{
\item{X}{a 1-column vector}

\item{Y}{a 1-column vector of the same}

\item{n.sim}{number of simulations}

\item{method}{the method applies. Possible choices are "isospectral" (default) and "isopersistent"}
}
\value{
output
}
\description{
Estimates correlation p-values for two timeseries using a Monte Carlo based method
}
\author{
Julien Emile-Geay

Nick McKay
}
