% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.regression.functions.R
\name{pvalMonteCarlo}
\alias{pvalMonteCarlo}
\title{Correlations and their significance according to AR(1) benchmarks}
\usage{
pvalMonteCarlo(
  X,
  Y,
  n.sim = 100,
  method = "isospectral",
  cor.method = "pearson"
)
}
\arguments{
\item{X}{a 1-column vector}

\item{Y}{a 1-column vector of the same}

\item{n.sim}{number of simulations}

\item{method}{the method applies. Possible choices are "isospectral" (default) and "isopersistent"}

\item{cor.method}{correlation method to pass to cor() "pearson" (default), "kendall", or "spearman"}
}
\value{
output
}
\description{
Generate parametric or non-parametric surrogates of two series X & Y
}
\seealso{
Other correlation: 
\code{\link{ar1Surrogates}()},
\code{\link{ar1}()},
\code{\link{corMatrix}},
\code{\link{effectiveN}()},
\code{\link{plotCorEns}()},
\code{\link{pvalPearsonSerialCorrected}()}
}
\author{
Julien Emile-Geay

Nick McKay
}
\concept{correlation}
