% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.regression.functions.R
\name{pvalPearsonSerialCorrected}
\alias{pvalPearsonSerialCorrected}
\title{Calculate correlation p-value given sample size.}
\usage{
pvalPearsonSerialCorrected(r, n)
}
\arguments{
\item{r}{correlation coefficient}

\item{n}{sample size}
}
\value{
p-value based on two-tailed t-test
}
\description{
Calculate Pearson p-values accounting for effective sample size
}
\seealso{
Other correlation: 
\code{\link{ar1Surrogates}()},
\code{\link{ar1}()},
\code{\link{corMatrix}},
\code{\link{effectiveN}()},
\code{\link{plotCorEns}()},
\code{\link{pvalMonteCarlo}()}
}
\author{
Nick McKay
}
\concept{correlation}
