% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.geoChronR.R
\name{quantile2d}
\alias{quantile2d}
\title{Find quantiles across an ensemble}
\usage{
quantile2d(
  x,
  y,
  n.bins = 500,
  x.bin = NA,
  probs = c(0.025, 0.25, 0.5, 0.75, 0.975),
  n.ens = max(c(ncol(x), ncol(y))),
  seed = 111,
  interp.method = "linear",
  limit.outliers.x = 0.0025,
  na.thresh = 0.5,
  ...
)
}
\arguments{
\item{x}{n by m matrix where n is the number of observations and m is >= 1}

\item{y}{n by j matrix where n is the number of observations and j is >= 1}

\item{n.bins}{number bins over which to calculate intervals. Used to calculate x.bin if not provided.}

\item{x.bin}{vector of bin edges over which to bin.}

\item{probs}{quantiles to calculate}

\item{n.ens}{how many ensemble members?}

\item{seed}{set a seed for reproducibility}

\item{limit.outliers.x}{limit the plotting of outliers on the x axis to exclude values below this probability limit (default = 0.025)}
}
\value{
list of quantiles and x.bin
}
\description{
Determine quantiles across ensembles of x and/or y, as a function of x, using interpolation
}
\examples{
\dontrun{
quantiles <- quantile2d(ageEnsemble,paleoEnsemble)
}

}
\seealso{
Other plot help: 
\code{\link{AD2BP_trans}()},
\code{\link{BP2AD_trans}()},
\code{\link{axisLabel}()},
\code{\link{bin2d}()},
\code{\link{geoChronRPlotTheme}()},
\code{\link{getLegend}()},
\code{\link{getPlotRanges}()},
\code{\link{kde2d}()},
\code{\link{meltDistributionTable}()},
\code{\link{modeSelektor}()},
\code{\link{periodAnnotate}()},
\code{\link{reverselog10_trans}()}

Other gridding: 
\code{\link{bin2d}()},
\code{\link{kde2d}()}
}
\author{
Nick McKay
}
\concept{gridding}
\concept{plot help}
