% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.geoChronR.R
\name{reverselog10_trans}
\alias{reverselog10_trans}
\title{Reverse axis in log10 scale}
\usage{
reverselog10_trans()
}
\description{
Reverse axis in log10 scale
}
\seealso{
Other plot help: 
\code{\link{AD2BP_trans}()},
\code{\link{BP2AD_trans}()},
\code{\link{axisLabel}()},
\code{\link{bin2d}()},
\code{\link{geoChronRPlotTheme}()},
\code{\link{getLegend}()},
\code{\link{getPlotRanges}()},
\code{\link{kde2d}()},
\code{\link{meltDistributionTable}()},
\code{\link{modeSelektor}()},
\code{\link{periodAnnotate}()},
\code{\link{quantile2d}()}

Other spectra: 
\code{\link{ar1Surrogates}()},
\code{\link{computeSpectraEns}()},
\code{\link{createSyntheticTimeseries}()},
\code{\link{periodAnnotate}()},
\code{\link{plotSpectrum}()}
}
\author{
Nick McKay, Julien Emile-Geay
}
\concept{plot help}
\concept{spectra}
