% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BAM.lipd.R
\name{runBam}
\alias{runBam}
\title{Generate a Banded Age Model (BAM) and add it into a LiPD object}
\usage{
runBam(
  L,
  paleo.num = NA,
  paleo.meas.table.num = NA,
  chron.num = 1,
  model.num = NA,
  ens.table.number = NA,
  make.new = FALSE,
  n.ens = 1000,
  model = NA
)
}
\arguments{
\item{L}{A LiPD object - an R serialization of a single LiPD file. It's a list, and is typically created by `readLipd()`}

\item{paleo.num}{the number of the paleoData object that you'll be working in}

\item{paleo.meas.table.num}{the number of the measurementTable you'll be working in}

\item{model.num}{the number of the chronData model where you want to store the model information}

\item{ens.table.number}{the number of the ensembleTable you want to put the output into}

\item{make.new}{Forces the creation of a new model (TRUE or FALSE{default})}

\item{n.ens}{The number of members in the ensemble}

\item{model}{a list that describes the model to use in BAM
\itemize{
\item model$ns: number of samples 
\item model$name: 'poisson' or 'bernoulli' 
\item model$param: probability of growth band being perturbed (default: prob of missing band = prob of doubly-counted band = 0.05)
\itemize{
     \item if model$param is a single argument, then the perturbations are symmetric (prob of missing band = prob of doubly-counted band)
     \item if model$param = [a1 a2] and a1 neq a2 the model is asymmetric
     \itemize{
                 \item a1 = prob(missing layer) - undercounted
                 \item a2 = prob(layer counted multiple times) - overcounted
                 }
     \item if model$param: 2xp matrix, then different miscounting prob. are defined for each time series.
     }
\item model$resize: do not resize: 0 (default), resize to shortest sample: -1, resize to longest sample: 1
\item model$tm: if a time model is provided, the code returns the corresponding perturbed data
}}
}
\value{
L The single LiPD object that was entered, with ageEnsemble and chronData model added.
}
\description{
This is a high-level function that uses BAM to simulate age uncertainty in layer counted records, and stores this as an age-ensemble in a paleoData measurementTable, and in a model in chronData. If needed input variables are not entered, and cannot be deduced, it will run in interactive mode. BAM produces reasonable results for non-layer counted data, and can generate ensembles for unevenly spaced data, and thus is useful for generating ensembles for tie-point chronologies that are missing the necessary data to calculate ensembles properly. See Comboul et al. (2015) doi:10.5194/cp-10-825-2014 for details.
}
\section{Long-form example}{

\href{http://nickmckay.github.io/GeoChronR/articles/Introduction.html}{View a full-fledged example of how to use this function.}
}

\examples{
\dontrun{
Run in interactive mode:
L = runBam(L)

Run in noninteractive mode, describing everything:
L = runBam(L,paleo.num = 1, paleo.meas.table.num = 1, model.num = 3, make.new = TRUE,
nEns = 100, model = list(name = "poisson",param = 0.05, resize = 0, ns = n.ens))
}
}
\seealso{
Other BAM: 
\code{\link{bamCorrect}()},
\code{\link{simulateBam}()}
}
\author{
Nick McKay

Maud Comboul (BAM)
}
\concept{BAM}
