% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.bchron.LiPD.R
\name{runBchron}
\alias{runBchron}
\title{Generate a Bayesian Reconstruction Age Model  (Bacon) and add it into a LiPD object}
\usage{
runBchron(
  L,
  chron.num = NA,
  which.table = NA,
  site.name = L$dataSetName,
  model.num = NA,
  cal.curves = NA,
  iter = 10000,
  outlier.probs = 0.05,
  ask = TRUE,
  lab.id.var = "labID",
  age.14c.var = "age14C",
  age.14c.uncertainty.var = "age14CUnc",
  age.var = "age",
  age.uncertainty.var = "ageUnc",
  depth.var = "depth",
  reservoir.age.14c.var = "reservoirAge",
  reservoir.age.14c.uncertainty.var = "reservoirAge14C",
  rejected.ages.var = "rejected",
  depth.units = "cm",
  ...
)
}
\arguments{
\item{L}{A LiPD object - an R serialization of a single LiPD file. It's a list, and is typically created by `readLipd()`}

\item{chron.num}{the number of the chronData object that you'll be working in}

\item{site.name}{the name of the site}

\item{model.num}{chron.numModel do you want to use?}

\item{cal.curves}{The calibration curves to be used. Enter either "marine13", intcal13", "shcal13" or "normal". Will prompt if not provided.}
}
\value{
L. The single LiPD object that was entered, with methods, ensembleTable, summaryTable and distributionTable added to the chronData model.
}
\description{
This is a high-level function that uses Bchron to simulate an age model, and stores this as an age-ensemble in a model in chronData. If needed input variables are not entered, and cannot be deduced, it will run in interactive mode. See Haslett and Parnell (2008) doi:10.1111/j.1467-9876.2008.00623.x for details.
}
\examples{
\dontrun{
#Run in interactive mode:
L = runBchron(L)

#Run in noninteractive mode:
L = runBchron(L,chron.num = 1, site.name = "MyWonderfulSite", model.num = 3, cal.curves = "marine13") 
}
}
\author{
Deborah Khider

Andrew Parnell

Nick McKay
}
\concept{Bchron}
