% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.bchron.LiPD.R
\name{runBchron}
\alias{runBchron}
\title{Generate a Bayesian Reconstruction Age Model  (Bacon) and add it into a LiPD object}
\usage{
runBchron(L, which.chron = NA, site.name = L$dataSetName, modelNum = NA,
  calCurves = NA)
}
\arguments{
\item{L}{a single LiPD object}

\item{which.chron}{the number of the chronData object that you'll be working in}

\item{site.name}{the name of the site}

\item{modelNum}{which chronModel do you want to use?}

\item{calCurves}{The calibration curves to be used. Enter either "marine13", intcal13", "shcal13" or "normal". Will prompt if not provided.}
}
\value{
L. The single LiPD object that was entered, with methods, ensembleTable, summaryTable and distributionTable added to the chronData model.
}
\description{
This is a high-level function that uses Bchron to simulate an age model, and stores this as an age-ensemble in a model in chronData. If needed input variables are not entered, and cannot be deduced, it will run in interactive mode. See Haslett and Parnell (2008) doi:10.1111/j.1467-9876.2008.00623.x for details.
}
\examples{
Run in interactive mode:
L = runBchron(L)

Run in noninteractive mode:
L = runBchron(L,which.chron = 1, site.name = "MyWonderfulSite", modelNum = 3, calCurves = "marine13") 
}
\author{
Deborah Khider

Andrew Parnell
}
