% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxcal.R
\name{runOxcal}
\alias{runOxcal}
\title{Generate an oxcal Age Model and add it into a LiPD object}
\usage{
runOxcal(
  L,
  chron.num = NA,
  meas.table.num = NA,
  oxcal.path = "~/OxCal/",
  model.num = NA,
  remove.rejected = TRUE,
  overwrite = TRUE,
  max.ens = 1000,
  static.reservoir.age = NA,
  static.reservoir.age.unc = NA,
  surface.age = NA,
  surface.age.unc = NA,
  surface.age.depth = 0,
  depth.interval = NA,
  events.per.unit.length = NA,
  events.per.unit.length.uncertainty = 0,
  outlier.prob = 0.05,
  cal.curve = "IntCal13",
  ...
)
}
\arguments{
\item{L}{A LiPD object - an R serialization of a single LiPD file. It's a list, and is typically created by `readLipd()`}

\item{chron.num}{the number of the chronData object that you'll be working in}

\item{meas.table.num}{the number of the measurementTable you'll be working in}

\item{model.num}{chron.numModel do you want to use?}

\item{remove.rejected}{don't write out dates that are marked as rejected}

\item{overwrite}{overwrite files and directories}

\item{max.ens}{the maximum number of ensembles to load in (default = 1000)}

\item{static.reservoir.age}{optionally assign a deltaR to all radiocarbon dates}

\item{static.reservoir.age.unc}{optionally assign a deltaR uncertainty to all radiocarbon dates}

\item{...}{
  Arguments passed on to \code{\link[=createChronMeasInputDf]{createChronMeasInputDf}}
  \describe{
    \item{\code{lab.id.var}}{}
    \item{\code{age.14c.var}}{}
    \item{\code{age.14c.uncertainty.var}}{}
    \item{\code{age.var}}{}
    \item{\code{age.uncertainty.var}}{}
    \item{\code{depth.var}}{}
    \item{\code{reservoir.age.14c.var}}{}
    \item{\code{reservoir.age.14c.uncertainty.var}}{}
    \item{\code{rejected.ages.var}}{}
    \item{\code{split.ages}}{if there's an age_type column, and only one age column, intelligently split between age and age14C}
  }}
}
\value{
L The single LiPD object that was entered, with methods, ensembleTable, summaryTable and distributionTable added to the chronData model.
}
\description{
This is a high-level function that uses oxcal to simulate an age model, and stores this as an age-ensemble in a model in chronData. If needed input variables are not entered, and cannot be deduced, it will run in interactive mode. See Bronk Ramsey et al. 2008 doi:10.1016/j.quascirev.2007.01.019
}
\examples{
Run in interactive mode:
L = runOxcal(L)

Run in noninteractive mode, describing everything:
L = runOxcal(L,chron.num = 1, meas.table.num = 1, model.num = 3, bacon.dir = "~/Bacon/",site.name = "MSB2K", cc = 1)
}
\author{
Nick McKay
}
\concept{oxcal}
