% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipd.manipulation.R
\name{selectData}
\alias{selectData}
\title{Select a LiPD "variable list"}
\usage{
selectData(L, varName = NA, where = "paleoData", which.data = NA,
  tableType = "measurement", which.mt = NA, always.choose = FALSE,
  altNames = NA, model.num = 1, which.ens = 1, which.sum = 1,
  strictSearch = FALSE)
}
\arguments{
\item{L}{a lipd object}

\item{varName}{string name of the variable to extract}

\item{where}{"paleoData" or "chronData"}

\item{which.data}{an integer that corresponds to which paleo or chron Data object (L$<where>Data[[?]]) has the variable you want?}

\item{tableType}{What type of table do you want to select data from? ("measurement", "summary" or "ensemble")}

\item{which.mt}{an integer that corresponds to which paleo measurementTable has the variable you want?}

\item{always.choose}{Force selection of the variable from a list}

\item{altNames}{A vector of strings for alternative names to search for}

\item{model.num}{an integer that corresponds to which model that has the variable you want}

\item{which.ens}{an integer that corresponds to which ensembleTable you want to get the variable from?}

\item{which.sum}{an integer that corresponds to which summaryTable you want to get the variable from?}

\item{strictSearch}{Use a strictSearch to look for the ageEnsemble and depth variables. TRUE(default) or FALSE.}
}
\value{
A LiPD "variable list" object
}
\description{
Selects and extracts a LiPD "variable list"
}
\seealso{
Other LiPD manipulation: \code{\link{estimateUncertaintyFromRange}},
  \code{\link{flipCoords}}, \code{\link{getVariableIndex}},
  \code{\link{mapAgeEnsembleToPaleoData}},
  \code{\link{pullTsVariable}},
  \code{\link{pushTsVariable}}
}
