% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacon.lipd.R
\name{setBaconDir}
\alias{setBaconDir}
\title{Set the name of bacon core directory}
\usage{
setBaconDir(baconDir)
}
\arguments{
\item{baconDir}{if you already have it}
}
\value{
baconDir the bacon core directory
}
\description{
Use this to programmatically set the baconDir
}
\author{
Nick McKay
}
