% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronBuilding.R
\name{stringifyVariables}
\alias{stringifyVariables}
\title{Convert variable names to strings}
\usage{
stringifyVariables(varName, usedValue)
}
\arguments{
\item{varName}{}
}
\value{

}
\description{
Convert variable names to strings
}
