% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronBuilding.R
\name{stringifyVariables}
\alias{stringifyVariables}
\title{Convert var.names to strings}
\usage{
stringifyVariables(var.name, used.value)
}
\arguments{
\item{var.name}{string name of the variable to extract}

\item{used.value}{}
}
\value{
the var.name as a string
}
\description{
Convert var.names to strings
}
