% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronBuilding.R
\name{stringifyVariables}
\alias{stringifyVariables}
\title{Convert var.names to strings}
\usage{
stringifyVariables(var.name, used.value)
}
\arguments{
\item{var.name}{string name of the variable to extract}

\item{used.value}{the value that was used}
}
\value{
the var.name as a string
}
\description{
Convert var.names to strings
}
\seealso{
Other utility: 
\code{\link{askUser}()},
\code{\link{concatEnsembleTimeseries}()},
\code{\link{convertAD2BP}()},
\code{\link{convertBP2AD}()},
\code{\link{createChronMeasInputDf}()},
\code{\link{gaussianize}()},
\code{\link{generateEnsembleFromUncertainty}()},
\code{\link{getLastVarString}()},
\code{\link{getOs}()},
\code{\link{heuristicUnits}()},
\code{\link{loadRemote}()},
\code{\link{pullInstance}()},
\code{\link{simulateAutoCorrelatedUncertainty}()},
\code{\link{surrogateDataFun}()}
}
\concept{utility}
