% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.functions.R
\name{summaryPlotTs}
\alias{summaryPlotTs}
\title{TS overview plot}
\usage{
summaryPlotTs(
  TS,
  sort.var = "archiveType",
  age.range = NA,
  age.var = "age",
  step = NA,
  ...
)
}
\arguments{
\item{TS}{}

\item{sort.var}{What variable in the TS should the map and histogram be organized by? (default = "archiveType)}

\item{age.range}{a two element vector that includes the range over which you want to count (default = NA, which calculates the full range in the TS)}

\item{age.var}{What age variable should be used (default = "age)}

\item{step}{What step size should be used along the age axis? (default = NA, which will calculate 0.5\% of the age.range)}
}
\value{
a gridExtra ggplot output
}
\description{
TS overview plot
}
