% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacon.lipd.R
\name{writeBacon}
\alias{writeBacon}
\title{Create the input file for a Bacon model from a LiPD object}
\usage{
writeBacon(
  L,
  bacon.dir = NA,
  chron.num = 1,
  remove.rejected = TRUE,
  overwrite = TRUE,
  cc = NA,
  site.name = L$dataSetName,
  model.num = NA,
  use.marine = NULL,
  ask.reservoir = TRUE,
  ...
)
}
\arguments{
\item{L}{a LiPD object}

\item{bacon.dir}{the directory where Bacon is installed on this computer.}

\item{chron.num}{the number of the chronData object that you'll be working in}

\item{remove.rejected}{don't write out dates that are marked as rejected}

\item{overwrite}{overwrite files and directories}

\item{cc}{An integer, or vector of integers corresponding to age that describes the calibration curve. You can specify here (see below) or if it's NA the code will guess based on archiveType
\itemize{
\item cc=1 IntCal13
\item cc=2 MarineCal 
\item cc=3 SHCal13
}}

\item{site.name}{the name used for the bacon model (and directories)}

\item{model.num}{chron.numModel do you want to use?}

\item{use.marine}{use the marine 13C curve? (yes or no, or NULL to choose)}

\item{ask.reservoir}{ask about reservoir corrections}

\item{...}{
  Arguments passed on to \code{\link[=selectData]{selectData}}
  \describe{
    \item{\code{var.name}}{string name of the variable to extract}
    \item{\code{paleo.or.chron}}{"paleoData" or "chronData"}
    \item{\code{paleo.or.chron.num}}{an integer that corresponds to paleo.num or chron Data object (L$<paleo.or.chron>[[?]]) has the variable you want?}
    \item{\code{meas.table.num}}{an integer that corresponds to paleo.num measurementTable has the variable you want?}
    \item{\code{table.type}}{What type of table do you want to select data from? ("measurement", "summary" or "ensemble")}
    \item{\code{always.choose}}{Force selection of the variable from a list}
    \item{\code{alt.names}}{A vector of strings for alternative names to search for}
    \item{\code{ens.table.num}}{an integer that corresponds to ensembleTable you want to get the variable from?}
    \item{\code{sum.table.num}}{an integer that corresponds to which summaryTable you want to get the variable from?}
    \item{\code{strict.search}}{Use a strict.search to look for the ageEnsemble and depth variables. TRUE(default) or FALSE.}
  }}
}
\value{
L the input LiPD file with methods added to the chronModel.
}
\description{
This generates the csv file that is used for input to Bacon. Will be run in interactive mode if necessary parameters aren't specified. Most users will want to use runBacon for their bacon needs.
}
\examples{
\dontrun{
writeBacon(L)
#Run in interactive mode

writeBacon(L,chron.num=1,meas.table.num = 1,bacon.dir="~/Bacon/",remove.rejected=TRUE,overwrite=TRUE,cc=NA,site.name=L$dataSetName,model.num=NA)
}
}
\seealso{
Other Bacon: 
\code{\link{getBaconDir}()},
\code{\link{loadBaconOutput}()},
\code{\link{runBacon}()},
\code{\link{sampleBaconAges}()},
\code{\link{setBaconDir}()}
}
\author{
Nick McKay
}
\concept{Bacon}
