% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacon.lipd.R
\name{writeBacon}
\alias{writeBacon}
\title{Create the input file for a Bacon model from a LiPD object}
\usage{
writeBacon(
  L,
  baconDir = NA,
  which.chron = 1,
  remove.rejected = TRUE,
  overwrite = TRUE,
  cc = NA,
  site.name = L$dataSetName,
  modelNum = NA,
  useMarine = NULL,
  askReservoir = TRUE,
  ...
)
}
\arguments{
\item{L}{a single LiPD object}

\item{baconDir}{the directory where Bacon is installed on this computer.}

\item{which.chron}{the number of the chronData object that you'll be working in}

\item{remove.rejected}{don't write out dates that are marked as rejected}

\item{overwrite}{overwrite files and directories}

\item{cc}{An integer, or vector of integers corresponding to age that describes the calibration curve. You can specify here (see below) or if it's NA the code will guess based on archiveType
\itemize{
\item cc=1 IntCal13
\item cc=2 MarineCal 
\item cc=3 SHCal13
}}

\item{site.name}{the name used for the bacon model (and directories)}

\item{modelNum}{which chronModel do you want to use?}

\item{which.mt}{the number of the measurementTable you'll be working in}
}
\value{
L the input LiPD file with methods added to the chronModel.
}
\description{
This generates the csv file that is used for input to Bacon. Will be run in interactive mode if necessary parameters aren't specified. Most users will want to use runBacon for their bacon needs.
}
\examples{
writeBacon(L)
#Run in interactive mode

writeBacon(L,which.chron=1,which.mt = 1,baconDir="~/Bacon/",remove.rejected=TRUE,overwrite=TRUE,cc=NA,site.name=L$dataSetName,modelNum=NA)
}
\seealso{
Other Bacon: 
\code{\link{loadBaconOutput}()},
\code{\link{runBacon}()},
\code{\link{sampleBaconAges}()}
}
\author{
Nick McKay
}
\concept{Bacon}
